desktop-pwas: Always create main shortcut for policy installed apps
Previously the create_desktop_shortcut policy option also controlled
creating the main shortcut (in the Start Menu on Windows and in the
Applications folder on macOS). This CL changes to always create
the main shortcut and have the policy only control if a desktop shortcut
should be created.

Bug: 844278
Change-Id: I6f877ed7fbf53e18ae17e111cbf277c86a95e2c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1545332
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Commit-Queue: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651620}