Image annotation service: support non-English descriptions.
This change adds a parameter to image annotation requests that
controls the language in which descriptions should be generated.

Currently only English descriptions are generated, with non-en
requests triggering OCR only.

We use the first accept language
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Accept-Language
as a proxy for the (only) language a user understands. We do
this (instead of e.g. passing the entire list) since almost all
accept language lists include en somewhere (and are therefore
not good representations of user fluency).

We use accept languages instead of more sophisticated language
modeling (e.g. the fluent language model
https://cs.chromium.org/chromium/src/components/language/core/browser/fluent_language_model.h
) since accept languages are the only readily-accessible source
of user language data in the renderer process. In future, better
sources of user language info should be used when they become
available.

We register as a renderer preference watcher to be informed
about changes to accept language as they occur. To do this, we
rename the RegisterRendererPreferenceWatcherForWorker method to
be more general.

Change-Id: I73d1c1f62bf0e5f3bcd3eec1d28e28240e70d9ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1554466
Commit-Queue: Michael Martis <martis@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651633}