chromeos: Add test for SettingsWindowManager with SplitSettings feature
With SplitSettings, browser settings opens in a tab, but OS settings
open in a window.

Delete an old test that was never running. The test was left over from
a time when we had experimental support for settings-in-a-window on
non-Chrome OS platforms. That support was removed in 2017.

Bug: 950007
Test: added to browser_tests
Change-Id: I8c2738263cbb43530fdcd6119aab2a2e2d088446
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1570577
Reviewed-by: calamity <calamity@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651715}