Replace child_count()/child_at() with STL algorithms.
These functions are deprecated.  Where we can't map the existing functionality
to algorithms, they're being replaced by range-based for, iterators, etc.  In
the cases in this CL, the underlying functionality maps to STL algorithms, so
make use of them; this provides the most high-level description of the intent of
the code, and in many cases can be simplified further in C++20 with ranges.  See
also
https://sean-parent.stlab.cc/presentations/2013-09-11-cpp-seasoning/cpp-seasoning.pdf
for arguments against loops (range-based or otherwise).

Bug: 940135
Change-Id: Ic4ef8a9fd6840645180472828e91306414bb795b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1565507
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651754}