[iOS] Ignore scrolls past bottom when FS provider is uninitialized.
When the FullscreenProvider is uninitialized, fullscreen is implemented
by resizing the entire web view.  When this occurs, the contentOffset
updates caused by the UIScrollView "snap back" CADisplayLink animation
interfere with the frame adjustments.  These scrolls should be ignored,
as they produce a jittery animation.

This issue does not occur for smooth scrolling, as the frame is not
adjusted in that scenario.

Bug: 920966
Change-Id: Ie918b825c83ec48436093dbad29fbcc063a18efb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1549287
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Auto-Submit: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651935}