ServiceWorker: Extends ServiceWorkerNewScriptloader for resuming paused download
If ServiceWorkerImportedScriptUpdateCheck is enabled, there are two
phases in updating: byte-to-byte checking and launching a new service
worker for install/activate events. If script change is found in phase
one, the download of changed script is paused. In the second phase,
we need to resume the paused download to avoid making new network
request again.

This patch is for second phase: Extends ServiceWorkerNewScriptLoader
to support resuming download of remaining script data from network.

Bug: 648295
Change-Id: I3bde3896d3093dfbfc0b9310a0410473adc3387d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1547410
Commit-Queue: Ting Shao <ting.shao@intel.com>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652042}