[TTS] Update Bar layout for Definitions.
This CL adds new layout treatment for dictionary definitions when they
are returned from the server:
1) Uses the existing "Book Round" vector icon when a dictionary is present.
2) Styles the Title to have two parts, with the first part being the word
the second part being the pronunciation shown gray.

The icon is controlled by a new ContextualSearchCardIconControl to allow
displaying an icon for cards from the server that are not quick actions.

Also update the ContextualSearchImageControl to use the term "Card Icon" in
place of "Quick Action Icon" since the icon can come from either a Quick
Action or some other kind of card.

BUG=936695, 954225

Change-Id: I26230ae060a00deda621ab776e673da9103e5730
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1572244
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Donn Denman <donnd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652315}