Parsing cookie SameSite: add None & Extended options, change default
This CL changes the default net::CookieSameSite value to UNSPECIFIED
(it was previously NO_RESTRICTION). The default is used when the cookie
does not specify a SameSite attribute, or if the given SameSite
attribute does not specify a valid value.

Support is added for the new `SameSite=None` option, which parses as
CookieSameSite::NO_RESTRICTION. Support is also added for the new
`SameSite=Extended` option, which for now is identical to Lax mode.

Bug: 953306, 953995
Change-Id: I96ee07f5e626a2f99e65e9c27260a8f6e5b6b9f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1573081
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652529}