Add GetInnerRangeBoundsRect To AXPlatformNodeDelegate
This change adds the GetInnerRangeBoundsRect interface to
AXPlatformNodeDelegate. Similar to GetRangeBoundsRect,
GetInnerRangeBoundsRect also calculates bounds for a range. Unlike
GetRangeBoundsRect, GetInnerRangeBoundsRect accepts text offsets
relative to GetInnerText in which text offsets on descendant nodes are
included. This interface is intended for clients which calculate text
offsets relative to GetInnerText, such as the UIA ITextRangeProvider
implementation which relies on AXNodePosition.

This change also adds values to the result of
BrowserAccessibility::GetInnerText so that it has parity with
AXNodePosition::GetInnerText.

GetHypertext and GetInnerText interfaces are added to
AXPlatformNodeDelegate although they are currently only implemented on
BrowserAccessibility and its subclasses.

Bug: 928948
Change-Id: I5607ee796a9f4bfc4352e43b2f7209200685b155
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1558708
Commit-Queue: Frank Poole <frpoole@microsoft.com>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Martin Robinson <mrobinson@igalia.com>
Cr-Commit-Position: refs/heads/master@{#652605}