WebUI: Fix Welcome Windows 10 Handler's navigation to the NTP
Previously, the Windows 10 Handler navigated to the NTP by making the
WebContents load the URL of the NTP.

That's is mostly correct, except that it doesn't tell the Browser to
update the omnibox.

This CL makes the Windows 10 Welcome Handler act the same way as the
normal Welcome Handler, which performs a navigation at the chrome/
with a Browser object instead of at the content/ level with a
WebContents object.

It used to work before accidentally, essentially.

Bug: 954120
Change-Id: I5514044f32ae542de52c8aba8c08373de1e84b53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1576138
Auto-Submit: Tommy Li <tommycli@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652905}