[DeviceSync v2] Add "DeviceSync:BetterTogether" key
Enrolling a key with name "DeviceSync:BetterTogether" adds the device
to the user's DeviceSync v2 "DeviceSync:BetterTogether" group. Only
devices that are in this group can send and receive device metadata
using SyncMetadata calls.

Note: This key is not to be confused with the unenrolled
"group key-pair" which is used to end-to-end encrypt device metadata.

Bug: 951969
Change-Id: I98d7a7aeda2656a86797d16410c2772d27d4a5a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1575681
Auto-Submit: Josh Nohle <nohle@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652975}