[Files app] ES6 class for util.js
Caveat: DOMError is deprecated according to MDN [1], but it's still
used in File System API so we have to stick with it. However Closure
externs for DOMError doesn't include the constructor with two arguments
which is available (at least in Chrome), so I kept the existing helper
createDOMError to take care of closure typing for UserDOMError and
DOMError.

I tried to use DOMError directly in CL:1571270 but closure didn't like
it. :-/

[1] - https://developer.mozilla.org/en-US/docs/Web/API/DOMError

Bug: 778674
Change-Id: Id316ecbcec23328ed62a0135450ad9edd7403553
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1571327
Auto-Submit: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653100}