Reland: Fix MOCK_TIME bug with racy cross thread immediate task posting
The fix is to disable CrossThreadTaskPostingDoesntAffectMockTime test on
iOS and to reduce the rate at which the spammy thread posts tasks as a
precaution.

Original patch: https://crrev.com/c/1565983

In https://crrev.com/c/1564008 we have a test that only cares about main
thread delayed tasks.  However it invokes an API which as a side effect
triggers cross thread PostTask to the main thread.  These tasks are
uninteresting except occasionally they happen in the window between
MockTimeDomain::DelayTillNextTask and
MockTimeDomain::MaybeFastForwardToNextTask.  Due to a bug in
MaybeFastForwardToNextTask it was treating this racy cross thread post
the same as running out of tasks and exiting early.

This is clearly undesirable and this patch fixes that behaviour.

TBR=gab@chromium.org, fdoray@chromium.org

Change-Id: I099c905761d68eb5723d53ed8796733a9cf60380
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1572121
Reviewed-by: Alex Clarke <alexclarke@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Alex Clarke <alexclarke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653277}