[omnibox] Create omnibox_resources containing Pedal concept data
This CL is like the previously landed (and reverted) CL...
1e6642b0358084c7d9069dca4fd03cad16641700
[omnibox] Add and load Pedal concept data for de, en, fr, ja, zh-CN
Previously, Pedal concept data was a singular resource in English but
this CL restructures the resources to load localized data from a
pak depending on selected language. Data for the first four non-English
languages localized by the translation team is included.

...But has the following differences:
* Pedal concept data itself is improved (esp. for fr, ja, and zh-CN).
* The per-locale resources are contained in a separate .grd file
  instead of being included in components_strings.grd.

See https://crbug.com/954348 for details about why the revert and
reworking of the previous CL was necessary.

Bug: 893183, 954348
Change-Id: I337ce9a72f6e27165e895ac9b5055641b4019ac5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1577782
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Orin Jaworski <orinj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653314}