[omnibox]: Dedupe keyword provider classification.
This is the 12th refactoring CL aimed at reducing duplication and
inconsistency for classifying omnibox results.

Keyword classification is trivial; i.e. the entirety of the contents and
description are styled uniformly. Descriptions are styled DIM. Contents
are styled 1) DIM (when the user input includes only the keyword), 2)
NONE (when the user input contains text following the keyword), or 3)
MATCH (when the user input contains only the keyword and the keyword has
no replacement text).

This CL changes contents classification in case (3) to use
`emplace_back` instead of `ClassifyLocationInString`. This is both
consistent with how the other cases are handled and inline with our goal
to deprecate `ClassifyLocationInString` and other duplicated
classification helpers.

Bug: 366623
Change-Id: I205a7bea19391d6343bc071574797f9a0f61c324
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1579763
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653410}