Clipboard: Remove unnecessary ClipboardBuffer argument
of mojom::ClipboardBuffer::kStandard. While it could be useful
to allow this argument to have other values, perhaps to use
for ClipboardBuffer::kSelection in X11, this is currently not
used and always has the same value, so let's remove the unnecessary
checks and passing around of this buffer value.

SystemClipboard: :Read{Image,PlainText}() have an argument,
mojom: :ClipboardBuffer buffer, which always has the same value
Change-Id: Ide1d9ecb5b8f2869b14232dc2acbfe4c681419e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1576184
Commit-Queue: Victor Costan <pwnall@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653521}