LoginShelfView: Hide Guest button after session starts
On LoginShelfView we maintain a list of screens where the guest button
should be hidden.

This CL adds a check that hides the guest button if any user session
has started. This way we don't have to keep updating the visibility
rules whenever we add or remove a post login screen.

This CL also introduces a couple of Guest button visibility tests
to prevent regressions.

Bug: 949658
Change-Id: I60c1559c4123f3b9bdec49002f1c440abb5c0e6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1579525
Commit-Queue: Tony De Luna <tonydeluna@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653862}