Worker: Introduce WorkerGlobalScope::Initialize()
This is a preparation for consolidating global scope initialization code
scattered among global scopes and module script fetchers into
WorkerGlobalScope::Initialize() overrides.

This CL adds pure-virtual WorkerGlobalScope::Initialize(), and makes each
derivative of WorkerGlobalScope implement it based on its spec texts. As the
first step, this moves initialization of WorkerGlobalScope's URL to
Initialize(). This does not cause behavior change because the initialization
order is not changed. Other initialization routines for referrer policy, CSP
etc will be moved by subsequent CLs.

Design doc:
https://docs.google.com/document/d/1JCv8TD2nPLNC2iRCp_D1OM4I3uTS0HoEobuTymaMqgw/edit?usp=sharing

For the record, this change is split from the following CL to reduce the
complexity of that:
https://chromium-review.googlesource.com/c/chromium/src/+/1525785

Bug: 835717, 924041, 924043
Change-Id: I7a0ac11d336216c6c2dca12b61bb8b98358c232b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1580985
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Hiroshige Hayashizaki <hiroshige@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653909}