[mojo-bindings] Move some unittest files around
These will be heavily modified in a follow-up CL and should also be
renamed to reflect the changes in what they cover (new types instead of
old types). This does the rename seprately so that the diffs in the
follow-up CL are clearer.

Bug: 955171
Change-Id: Ica24749cce65f163ba9a9fee3994b9ad37ab14ba
TBR: oksamyt@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1581161
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#654108}