Fix UAF in NavigationPredictor test.
The test was initializing two nested ScopedFeatureLists in a way
that caused UAF on FieldTrial objects. The nested lists are not
actually required and this CL changes the test not to do this.

In separate CL, I'm going to look at a way to prevent this issue
to begin with, for example via a DCHECK, but this requires seeing
if other tests may be affected.

Bug: 955643
Change-Id: Ic63543786b45b0fccde51d1287f42de23f183765
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1584889
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654169}