compiler: reenable ThinLTO caching for Chrome OS
We had to disable this due to a bug in lld. The bug has been fixed, and
the fix has been integrated into our toolchain.

I built chromeos-chrome with this applied, and ran autotest's smoke
tests and security.ChromeSandboxed on it. Neither failed, so I assume we
should be good to try this again.

Bug: 889967
Change-Id: I22ace0450b18fba942d59802d6e2b7636ccf8c35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1558713
Reviewed-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654305}