Add a new test, testServerReturnsBypassWithHostBlacklisted
I also added a redirect_chain field to HTTPResponse. This is the ordered
list of URLs in the redirect chain. If a request did not have any
redirects, this list is empty. This is used to verify that a server
bypass happend.

For example, testServerReturnsBypass checks that we show the original
page, and also that we got there via a redirect (bypass) from the
server. OTOH, testServerReturnsBypassWithHostBlacklisted loads the page
twice, where the second load should _not_ go through the server.

R=robertogden@chromium.org

Bug: 908491
Change-Id: I0e4c66917098013f2137d0cf2306567ddf8f4a60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1584436
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Commit-Queue: Tom Bergan <tombergan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654594}