Schedule background task when changing some offline prefetching settings
Causes the background task to be scheduled when limitless prefetching is enabled
and when the prefetch testing header value is changed. There is a five-second
delay built in when the background task is scheduled this way; this gives the
user a chance to change both limitless and the testing header before the task
starts.

Bug: 954252
Change-Id: Ib21e8e0fd4dca29226c2b3e2080da1afe7650ed8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1574139
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Reviewed-by: Dan H <harringtond@chromium.org>
Commit-Queue: Ian Wells <iwells@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654608}