[ozone/wayland] Add possibility to create wl buffer immediately.
zwp_linux_dmabuf protocol is able to create the wl_buffer both
synchronously and asynchronously. This feature depends on the
protocol version.

Thus, if the protocol corresponds to the version, when the buffer
can be created immediately, utilize it and do not wait for the
answer from the Wayland compositor.

Bug: 939794
Change-Id: I14fcad5eae8bcce3b24b662867c13948eb7e4ae6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1572344
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Michael Spang <spang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654800}