[Image Capture] Enable image capture tests with real webcams using video capture service.
This is a new attempt to enable tests that were disabled due to
flakiness on WebRTC bots (this is a roll-forward of http://crrev.com/c/896062).
This is a further attempt to get these tests to run on the bots.

NB: running the tests locally, I was able to verify that due to issues
in the UVC kernel module on Linux, the test
WebRtcImageCaptureSucceedsBrowserTest.ManipulateFocusDistance was
consistently failing with and without the video capture service. Since
the issue is not in Chromium, I am disabling the test and tracking the
work to re-enable it in a separate bug (http://crbug.com/c/957020).

I was able to run the tests locally and they passed with 3 different
cameras, namely Logitech C920, Logitech C930e, and Sonix Technoligy
Facevision.

TBR=chfremer@chromium.org
TEST=content_browsertests --gtest_filter=UsingRealWebcam* --run-manual --test-launcher-jobs=1
BUG=721812

Change-Id: I914dde51664a26ee2134d5c7a4c663bc801db9fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1583709
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Armando Miraglia <armax@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654819}