Experimentally raise font scanning timeout to 2,5 minutes
Font scanning and lookup table construction is only required for font
local matching pre Windows 10. If the timeout is hit, no local font
matching will be performed on this particular pre Win 10
system. Previously we had this set to 15 seconds, which did mean
timeouts on a number of systems, but also previously the render was
waiting synchronously for a result, which is no longer the case.
Experimentally raise this to 150 to see what scanning times are occuring
in the field.

Since the renderer is no longer synchronously waiting, in order to avoid
startup jank due to contention, reduce priority to BEST_EFFORT.

Currently only active behind an experiment configuration.

Bug: 889864, 931366
Change-Id: I19da2ef1e826b8c72a7ea6f0d22814e3e1c35366
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1581627
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654844}