Simplify biquad usage in getFrequencyResponse
The existing implementation was confusing because it looked like it
was modifying the filter coefficients of the running biquad filter.
Change this so that the biquad kernel is passed that will be used to
compute the response.  This makes it clear that the filter
coefficients aren't being modified. It's up to the caller to use the
correct biquad kernel.

Bug: 390266
Change-Id: Ib67afa9112dfa666875b40be3e7c582eef767e65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1582869
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654881}