Refactor Windows Mixed Reality COM Events to Callbacks
In order to test Windows Mixed Reality, there needs to be a place to add
calls to the VR Test Hooks.  Currently, the direct calls to COM are not
mockable or easily replaced.  By creating a simple wrapper for the COM
calls, this change creates places for the test hooks to be added, while
also cleaning up the code to make it more readable.

This change finishes the transition from COM Calls to wrapper classes by
wrapping the COM events into chromium callback types, and creating
classes to wrap the relevant args, as well as removing the "GetComPtr"
methods.  The "GetRawPtr" methods are necessary in order to pass the
underlying pointers to various COM calls, but it is not expected that
this method is needed outside of the wrapper classes.

Bug: 954413
Change-Id: I9890de1b33ca12e50fe445aafe8f824c7af7657f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1585096
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Bill Orr <billorr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654931}