Splitting relative motion from pointer lock.
Currently, exo's implementation locks the pointer immediately when a
relative motion listener is requested. As a corollary, that errors when
exo can not figure out which window to lock to. In XWayland the
relative pointer is requested orthoganally to pointer locking, meaning
no relative motion events will be sent, since exo refused to register
the motion listener.

This cl splits the relative motion registration from window locking. We
are emulating the old behaviour as much as possible (i.e. we still try
to lock when relative motion is requested) though in a future cl we
plan to implement pointer-constraints-unstable-v1.xml, and remove the
request to lock from the relative-motion code.

Bug: 927521
Change-Id: I8db502b337acf00104ec101b51e4196d17ca6fc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1568347
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: Mitsuru Oshima (Slow 4/22-26) <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655083}