[remoting][FTL] Make FtlSignalStrategy take unique_ptr to OAuthTokenGetter
We have been seeing some lifetime management issue with
FtlSignalStrategy. It currently take raw pointer to OAuthTokenGetter but
we delete consumers of it ~5 seconds after the signal strategy is
deleted so that we still have a chance to send out all pending requests.
There is a chance that the signal strategy tries to access the token
getter after it is deleted by the caller.

This CL makes OAuthTokenGetterProxy usable by the host and upgrades it
to handle sequences. The host will now create an owned token getter
proxy and pass it to FtlSignalStrategy, which is not different than
passing a WeakPtr. Mobile clients OTH already have
ChromotingClientRuntime::CreateOAuthTokenGetter() that creates proxied
token getter from a token getter singleton, so this change will allow
them to send out pending messages after the session is terminated and
prevent use-after-free from happening.

Bug: 954594
Change-Id: Iaf2052c936ef9555f2d017773cfba01e9a5b819d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1588959
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655098}