Allow download clients to be lazily inflated
Currently download Clients need to be eagerly created and passed to the
DownloadService. This CL implements DeferredClientWrapper that allows
a client that requires the full browser to not be inflated unless it is
needed. The cl also creates the NativeStartupBridge in java that would
allow native code running in reduced mode to request a full browser
start

Bug: 943724
Change-Id: I77330db859a894b9fc45183c251ec22a1c3eba7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1567904
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>
Auto-Submit: Mohamed Heikal <mheikal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655119}