Blob: Eagerly complete reading when there are no more remaining bytes.
Similar to crrev.com/c/1586258 the MojoBlobReader can possibly delay
completion under the following conditions:

1. The mojo data pipe is sized to exactly the blob length since the
   blob is less than 512kb.
2. The first read fills the entire pipe exactly and triggers
   ArmOrNotify to wait for the next writable event.
3. Potentially long wait for the consuming renderer to make room in
   the pipe.
4. When the pipe is finally writable the reader triggers another read.
5. The read returns 0 and realizes its done.

This CL allows the reader to complete at step (2) when there are zero
remaining bytes expected.

The completion delay could be significant depending on the what is
consuming the blob.  For example, if a blob is being consumed as a
service worker provided response, then the completion delay may cause
subresource processing to miss an idle window on the main thread and
be further delayed.

This fix is behind a feature flag to allow us to determine how much
of an impact the change has on service worker loading performance.

Bug: 957175
Change-Id: Ib309d1663b9fe5255bcde06d1374faec5ec1f39d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1588201
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655125}