Add CPU usage to PageLoad/AdPageLoad UKM
Record CPU usage for ad page loads. Also records total page cpu usage
in the PageLoad UKM to use as a comparison. This number can be combined
with PageLoad.PageTiming.Foreground duration to get CPU utilization.
Only CPU usage in the foreground will be recorded as there are many
background interventions that skew this data when a tab is
backgrounded.

Change-Id: I4d036e3d7c2c4853577c1459b7230b76d43673aa
Bug: 949800
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1554142
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Commit-Queue: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655247}