Connect the new Memory Pressure Monitor to ChromeBrowserMain
This monitor doesn't do anything except recording metrics for now, the
goal of this CL is to have all the plumbing in place so we can start
conducting Finch experiments and tune this signal. Once calibrated a
new parameter will get added to the Feature to allow replacing the
existing memory pressure monitor (from base) with this one.

Bug: 771478
Change-Id: I1c063fb82468b76577107473c6129b8b3300fe45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1588921
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Commit-Queue: Sébastien Marchand <sebmarchand@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655264}