[omnibox] Clean up answer styling.
For the bolding, the stock price should not be bold because it is not
related to the user's query.

I also think this makes it a bit clearer why the styling is reversed
and makes it a bit easier to add other non-reversed styling types.

Bug: 956829
Change-Id: I1da57517842b1743be9a6459cf4dd581a7f47653
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1571704
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#655321}