Fixing timestamp in workers for the callback in request animation frame
The time that the worker is presention in the requestAnimationFrame
callback, has to take into account if the worker is in the WorkerGlobalScope
as in https://developer.mozilla.org/en-US/docs/Web/API/WorkerGlobalScope/performance.

Replace how the time was capture for the callback to take into account the time
since the worker started by calling performance.now with the right performance
obtained from WorkerGlobalScope.

Bug: 935096
Change-Id: If35ae0ab7e734357005e07122af75ef5da12e009
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1582527
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655323}