Policy management : Add atomic group policies concept
Some policies like the proxy settings are part of a logical group
and it makes sense that they are treated as one. Those policies will
now be part of an atomic policy group. It will help to have only one source of
truth for the values of the from a same group when they can be set from
multiple sources.

Memory footprint:
Around 511 bytes to have the atomic groups.
Around 592 bytes to have the map of policy to atomic group.
Max of 2280 bytes to have all policies in a group.

Bug: 949268
Change-Id: I69ba0bfecfe42ebd36310fc63cb8f81afe8f744b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1551457
Commit-Queue: Yann Dago <ydago@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655327}