Retry YUV overlay format if swap chain size changes
The default overlay format is YUY2 but switched to BGRA if the overlay swap
chain fails to promote to hardware overlay. To save power consumption, the
current code will force it to stay in BGRA mode for the rest of life if one
of the swap chains fails.

Because the current overlay mode is about 70%, we want to see if there is a
chance to increase it. This CL will give it a chance to retry YUY2 format.
We can monitor the UMA histogram and see if this helps.

Bug:957678

Change-Id: I3b2c73d7553eb7a166a74015f9265f141217ae5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1588922
Commit-Queue: Maggie Chen <magchen@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655364}