Add 2 more browsertest and additional checks for existing tests.
(1) Add browsertest to verify that DRP (Data Reduction Proxy)
is used only when render frame ID is set.

(2) Add browsertest to verify that DRP is used on render
initiated navigations

(3) Add tests to ensure that chrome-proxy headers are
set when the request is sent to the proxy server.

Change-Id: Iba3c6cd0b9e74f2d60e5c1686687c1c20e690ca2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1589114
Commit-Queue: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Reviewed-by: rajendrant <rajendrant@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655400}