Enabled crostini's uninstallation UI by default
This allows users to uninstall applications from the launcher menu. It
has worked for some time but we held off on enabling the flag pending
resolution on crbug.com/909071.

The resolution is that we will enable this feature while working on
that last bug.

Bug: 822514
Change-Id: If1a8937d0edeb20aad076256fef82bd64982de9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1575310
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655507}