views: fix bug in activation handling
Interesting enough it is possible to end up in
DesktopNativeWidget::OnWindowActivated() and *not* have the Widget be active.
This ensures we don't attempt to restore a focused view in OnWindowActivated()
if the Widget isn't really active. To do otherwise might result in activating
the widget.

BUG=956963
TEST=see bug

Change-Id: I7e6e5feee12edc062ea314405995c122750343ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1586467
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655523}