Avoid DCHECK when marking partially disconnected ancestor chains.
Node::MarkAncestorsWithChildNeedsStyleRecalc currently relies on Node's
(entire) ancestor chain to either be connected, or not be connected.
Unfortunately, it's possible to enter a state which violates this
during removal of nodes.

When removing nodes from a <form> subtree, HTMLFormElement::
InvalidateDefaultButtonStyle can call PseudoStateChanged on another
node in the <form> subtree, which in turn eventually triggers a
MarkAncestorsWithChildNeedsStyleRecalc on the partially disconnected
ancestor chain.

BUG=956537
R=futhark@chromium.org

Change-Id: I22a2a5a51dbd7dc069abdb98e03345754735caf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1588393
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655552}