Fix flaky AdsPLMOBrowserTest.FrameWithSmallAreaNotConsideredVisible
Add a FrameSizeExpectation() method to the PLMTestWaiter that waits for
a frame size update of a given size to be received. This is added to
prevent a race condition in browsertests where we navigate the page
before the update is received.

This test is flaking on a Skia FYI bot. This is an existing race
condition that is exposed with certain features enabled. See bug for more
info.

Bug: 958094
Change-Id: I171e3de6d37b55119f6aa84554f09252990e9ebc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1592068
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Commit-Queue: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655756}