Use chrome_test_util::ClearBrowsingHistory() in ChromeEarlGrey only.
Replace GREYAsserts in EG2 tests to CHROME_EG_ASSERT_NO_ERROR tests and
use [ChromeEarlGrey clearBrowsingHistory]
instead of chrome_test_util::ClearBrowsingHistory().

This is part of EG1 to EG2 migration, which involves moving EarlGrey code from
app-side helpers into test code. For GREYAsserts, the policy is to take assert-
containing-helpers, modify them to return NSError* with localizedDescription
containing the assert failure (in the event of a failure), and then assert
on the result of the helper on the test side.

Bug: 922813
Change-Id: I939aaa81f50521cafa69f345119c868aacd83dfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1590546
Commit-Queue: Maksym Onufriienko <monufriienko@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655785}