Fix 'hidden' argument to LocalFrameView::UpdateRenderThrottlingStatus
For the non-OOPIF case, the 'hidden' argument indicates only whether
the LFV is scrolled outside of the parent document's viewport. It is
used to set LocalFrameView::hidden_for_throttling_, which also depends
on whether the iframe has zero size or is display:none.

Several call sites to UpdateRenderThrottlingStatus would pass in
hidden_for_throttling_ as the 'hidden' argument, in an attempt to
preserve the existing value. However, that's wrong, because the
'hidden' argument should only represent the scrolling state, not the
other determinants of hidden_for_throttling_.

This patch changes the meaning of hidden_for_throttling_ for non-OOPIF
iframes: it now represents only whether the iframe is scrolled outside
the viewport. Passing it as the 'hidden' argument to
UpdateRenderThrottlingStatus is now correct.

This also fixes the syntax of a couple of web_test that recently
turned flaky. By rights, the tests should have been consistently
failing; they only appeared flaky because the test harness was
crashing and not terminating.

BUG=956547

Change-Id: Ie635a82125e08e4252534e0b3a409aad782cf7ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1589657
Commit-Queue: Stefan Zager <szager@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655853}