[omnibox bookmarks] Make short symbols match beginning of words
Currently, a symbol (that a user might enter in the Omnibox) that is
very short (e.g. 1 char) will not be allowed to match the beginning of
a word. This CL flips that match behavior to allow a partial match,
but only when a flag is set. We intend to measure and make sure that
suggestion quality doesn't degrade.

TBR: mpearson@chromium.org
Bug: 925815
Change-Id: If279d0add5015a11d7528710a8cde020fba5cdef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1488615
Reviewed-by: Kevin Bailey <krb@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Kevin Bailey <krb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656034}