wm: Fix swipe up shelf to exit overview.
1) Fixes a bug where non resizable windows shift when getting swiped up.
2) Fixes a bug where windows do not have their transforms reset when swiped
away while in overview.
3) Fixes a bug where the "cannot snap this window" label does not move
with the overview item while swiping.

Introduced a cache of initial y positions as the calcuations to fix 1) got
really confusing (and already was). This introduced a weird bug where
the transform could not be read from the first window in the window grid,
though it was visually transformed. Tackled this with a workaround, but I
think its worth it for the codes simplification.

Test: manual
Bug: 891463
Change-Id: I2726ea5c10aad9ab308ecbfffacd0421c4040aa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1592809
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656168}