vk: Don't rebuild VK in ShellTestApi.
There was a bug in the select_file_dialog_extension browser test where
the virtual keyboard had to be reloaded for it to work. This was
because we were loading the VK using command line flags, but the
test was delaying the component extension loading until later, so
the virtual keyboard extension could not load.

This patch removes the initial command line load and changes the
EnableVirtualKeyboard helper to enable instead of reload.

Bug: 849995
Change-Id: I51a7ac2ad3cb9d804d764a357c2da2e1fb3d4aef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1592952
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Darren Shen <shend@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656216}