Add fallback fonts unittests
This CL is adding unittests for the GetFallbackFont on Windows.
It's fixing GetUniscribeFallbackFont which was failing on multiple
added tests.

The ShapeRuns (ui/gfx/render_text_harfbuzz.h) function applies fonts on runs
in a specific order. The FallbackFont is a fastpath to determine a font based
on a sequence of characters. If this is failing, the FallbackFonts returns
the list of fonts in that family.

FallbackFont on windows 7 is using the Uniscribe API and on windows 10 it's
using the DWrite API. We are exposing the Uniscribe API to allow better
unittests.

GetUniscribeFallbackFont is determining the fallback fonts by parsing the
uniscribe commands generated in a metafile. The previous code was assuming
the fallback font is the first font to be set. By looking to failing cases
we realized that multiple fonts may be selected and some spacing may be
produced into the metafile. A state machine is now assuming the fallback
font is the font used when the requested text is produced.

We are adding tests for script that are currently working. A CL will follow
with fixes for broken scripts. They are working in Chrome because the
FallbackFonts is able to return the full list of fonts.

BUG: 944227

Change-Id: I272a877d82fa031eff9e129e9cd4af915c756e89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1549882
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656413}