[omnibox] Remove UI logic that depends on the entity suggest Feature.
That logic is not actually doing anything useful. It used to be
necessary as part of the birthday launch where a variety of different
features might be enabled in different combinations. Now we should
simply render as a rich suggestion anything that has an associated
image_url (in other words, anything that's an entity suggestion).

Additionally, enable the Feature by default on desktop (this now has
no effect but is consistent with the normal process where we enable
features by default before launch).

Bug: 806995
Change-Id: I03b1e43cd9051d02def09915041cc34f937e9ce6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1593844
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Justin Donnelly <jdonnelly@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656498}