Add support for Xbox Adaptive gamepad on Linux/Mac/ChromeOS
This CL adds support for the Xbox Adaptive Controller over USB on
Mac, Linux, and Chrome OS, and Bluetooth on Mac. It also fixes some
issues that prevented Xbox One devices from working correctly on Mac.

On Linux and Chrome OS, the device does not stay connected over
Bluetooth but is usable as a USB gamepad.

This CL corrects the XInputType field for Xbox devices that operate
as HID gamepads when connected over Bluetooth.

On Mac, XboxControllerMac objects were not correctly shut down before
they were destroyed. This CL ensures these devices are shut down when
they are disconnected or the tab is closed.

BUG=953999

Change-Id: I3e98157647c8c408060a389de270445831782d8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1573174
Reviewed-by: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Commit-Queue: Matt Reynolds <mattreynolds@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656550}