[WebUSB] Trigger Android permission dialog when Chrome dialog dismissed
This change triggers Android's USB permission dialog when the user
selects a USB device in Chrome's USB permission dialog. Assuming
permission is granted device information, including serial number, is
refreshed and the new device information is what is passed to Blink.

This fixes the problem that the USBDevice.serialNumber field is not
populated when the object is passed to script. It doesn't solve the
problem that USB device permissions are not remembered across browser
restarts because while Chrome can now remember the device serial number
Android does not remember the user's choice and so Chrome cannot
compare this data to new device connections.

Bug: 932672
Change-Id: If90ee0a90de58561b091c19f7bafe24a83cfd723
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1592762
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656595}