android: crazy-linker: Add LoadParams type and use it in ElfLoader.
This CL introduces a new LoadParams struct used to keep various
parameters for loading a library with the crazy linker, and
refactors the internal API to use it.

This struct will be expanded later to allow loading from a
reserved memory mapping, and a file descriptor, to implement
a proper wrapper for android_dlopen_ext().

+ Ensure LoadLibraryWithSystemLinker aborts if library is
  already loaded.

BUG=936001
R=pasko@chromium.org, agrieve@chromium, cjgrant@chromium.org

Change-Id: I418cdbe0d7e3cb35615f0bfe6aa42c51abef7ac8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1564156
Reviewed-by: Egor Pasko <pasko@chromium.org>
Commit-Queue: David Turner <digit@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656695}