Switch chromeos test to using UnescapeBinaryURLComponent().
This CL is part of a project to cleaning up uses of
UnescapeURLComponent(), as it's been used for wildly disparate
purposes, and much of its logic is only suitable for cases where
we want to format a URL for safe display to the user. We want
to make it leave more weird characters escaped, but need to
eliminate usage of it in other contexts before we can do that.

UnescapeBinaryURLComponent() unconditionally unescapes all escaped
characters, and should be used instead in most cases, particulars where
URL_SPECIAL_CHARS_EXCEPT_PATH_SEPARATORS is in use. The primary cases
where it shouldn't be used (without additional checks), are cases where
unescaping path separators or funky characters (like %01) is a problem.
In those places, we should generally fail, instead of partially
unescaping a URL.

Bug: 849998
Change-Id: I206bbf6a29cc615b37d5f1e4c9c306bbd7efdc92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1592710
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656708}