Give users the option not to auto connect to WiFi networks.
In order to auto connect to hidden Wifi networks, CrOS must broadcast
the SSID of the network to see if it's available, which leaks
information about what networks the device is configured to
auto-connect to.  This page gives users the option of whether or not to
auto connect to a WiFi network when first configuring the network, and
provides a warning that auto connecting to a hidden network is bad from
a privacy point of view.  A pop up dialog with a longer explanation
about why auto connecting to a hidden network is bad from a privacy
point of view will be included in a later cl but is currently blocked
on a CSS bug (crbug/954202).

Change-Id: I0abd56ff2b3a57ec77f31feb6b66d800cf2ebeb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1476090
Commit-Queue: Melissa Galonsky <mgalonsky@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656767}