PaymentRequest: Fix an infinite spinning loop when selecting user infos
There is a bug that the payment sheet UI falls into an infinite spinning
loop when users re-select another item (e.g. shipping address, payer
detail info) since retry() is called.

This problem has been occured since http://crrev.com/c/1512594. In the
patch, the logic was changed so that updateWith() calls were forwarded
to the invoked payment app. The change is very reasonable and important,
but in the case of the autofilll payment app provided by the UA itself,
the change might not be appropriate. Therefore, this CL applies an
exception to the autofill payment app for the logic.

Test:
  chrome/android/javatests/src/org/chromium/chrome/browser/payments/PaymentRequestRetryTest.java

Manual Test:
  https://w3c-test.org/payment-request/payment-response/onpayerdetailchange-attribute-manual.https.html

Bug: 861704, 941837
Change-Id: Idf5ab1a45795490c7ee0e0e8e313e8f6213b0adf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1589680
Commit-Queue: Jinho Bang <jinho.bang@samsung.com>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656934}