[ozone/wayland] Ease the buffer swap and maintenance
This CL adds a Surface class, which has 1:1 relationship
with a WaylandWindow (whenever the former is created, the
manager is notified, and it creates an internal representation
of the window).

The CL does not bring any major changes, but rather improves
and eases the maintenance of buffers and distinguishes
buffers by the widget they have been created for.

This is a prerequisite to switch from post frame callback wait
to the pre frame callback wait.

Bug: 943096
Change-Id: Ia70f061c4c47a9675c2a018a346b6bc38d0e8cf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1570014
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Michael Spang <spang@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657172}