Change policy template length check to be an error instead of warning
The policy description check provided a warning when the length exceeded 3100
characters, and an error when exceeding 4096 characters.

This change removes the warning, and lowers the limit to 3500. 3500 is judged
to be a safe value by checking the current maximum length of a policy's
description translations.

Bug: 958784
Change-Id: Iacb1e52cb98118b7eecea5c2947bf4f9ba660fb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1598801
Commit-Queue: Jit Yao Yap <jityao@google.com>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657214}