[iOS] Should not show "Initial sync setup" error without Unity
With crrev.com/1560152, "Initial sync setup" error is now shown even
when Unity is not on. This patch only changed the UI. When the settings
are closed, the sync with started (like it used to be before the patch).
To keep the same behaviour before the patch, kSyncSettingsNotConfirmed
needs to be ignored in SyncSettingsTableViewController.

Bug: 956506
Change-Id: I9f17778273c62a2dc7dff816abee827d6d3b8333
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1599169
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657307}