Download service: Support asynchronous BlobStorageContextGetter.
In reduced memory mode of Chrome, the browser context will be loaded
asynchronously, so we need to wait until the profile being created and
then retrieve the BlobStorageContextGetter from a Profile.

This CL creates an interface to retrieve the BlobStorageContextGetter
in a callback to download service.

TBR=peter@chromium.org

Bug:959208

Change-Id: I444f9836324428ed285ac67a135e14f6380babc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1591948
Reviewed-by: Xi Han <hanxi@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657318}