Switch index_db_browsertest to using UnescapeBinaryURLComponent().
It was using UnescapeURLComponent(), which, unlike
UnescapeBinaryURLComponent(), leaves certain characters escaped. The
behavior difference doesn't really matter for the purposes of these
tests, we're just trying to reduce the consumers of
UnescapeURLComponent() to only those that need and expect it not to
unescape UTF-8 characters that are unsafe to display, so we can expand
that logic.

This CL is part of a project to clean up uses of
UnescapeURLComponent(), as it's been used for wildly disparate
purposes, and much of its logic is only suitable for cases where
we want to format a URL for safe display to the user. We want
to make it leave more weird characters escaped, but need to
eliminate usage of it in other contexts before we can do that.

UnescapeBinaryURLComponent() should be used instead in most cases,
particularly where URL_SPECIAL_CHARS_EXCEPT_PATH_SEPARATORS is in use,
as all consumers of that option seem to either be slightly buggy, or
don't actually need to leave any characters unescaped. The primary
cases where switching to UnescapeBinaryURLComponent() may cause
problems (without additional checks), are cases where unescaping
path separators or funky characters (like %01) is a problem. In those
places, we should generally fail instead of partially unescaping a URL.

Bug: 849998
Change-Id: I7d1f64b5fab51b8120a0c3504e4e8cd1610b79d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1600092
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Chase Phillips <cmp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657455}