WebUI: cr-checkbox, handle space on keyup similar to other UI elements
The issue referenced in the bug is due to the checkbox sending the
changed event when the space key is pressed. The action menu is closed
and when the space key is released, the action menu button is activated.

Making the cr-checkbox similar to <button> and cr-icon-button by
only handling the space key being released fixes this issue.

Bug: 956497
Change-Id: I9eff478f4b7824d76d35e9c8507ea3d036844ed8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1592219
Commit-Queue: Esmael El-Moslimany <aee@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657491}