Custom scrollbar hit-test fallback.
When hit-testing a scrollbar, ScrollbarController::HandleMouseDown
applies a static_cast on the returned layer. For standard scrollbars,
the layer returned is PaintedScrollbarLayerImpl where as for custom
scrollbars, PictureLayerImpl is returned and everything was getting
casted to ScrollbarLayerImplBase. This would then return true for
the is_scrollbar check and fail later on. The fix here would be to
avoid using static_cast and instead, use LayerImpl::ToScrollbarLayer().

Bug: 960521
Change-Id: I2fd6e7e6d239c7293ad7af3d73875fb758e803fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1598843
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Rahul Arakeri <arakeri@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#657495}