Ensure DownloadManager is created for DownloadService on profile creation
For browser tests that use download service, content::DownloadManager needs to be
created so that DownloadDriverImpl has the SimpleDownloadManager object to carry
out the download task. However, some browser tests didn't explicitly create the
DownloadManager after profile creation, as this is done automatically in
DownloadServiceFactory.
This CL makes the download service creation to proceed without the
DownloadManager. But will set it whenever profile is created. It also works with
reduced mode as inprogressDownloadManager will be created to handle the download

BUG=942770

Change-Id: Idc2a1d996dd65c742dea0263e80a3e422c8d9b1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1597948
Commit-Queue: Min Qin <qinmin@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657518}