viz: Avoid damaging the main overlay when updating video overlays.
Since the media code marks candidates opaque based on the video format,
it is common to have video quads be marked not opaque. If we use any
strategy other than an underlay in this case, the damage rect for the
overlay is not subtracted from the damage rect for the OutputSurface
plane which causes us to unnecessarily redraw the main plane.

R=liberato@chromium.org, piman@chromium.org

Bug: 889328
Change-Id: I3e5436b051b1becc39b5b5396ec3f676afc203b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1600492
Reviewed-by: Antoine Labour <piman@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657542}