Makes it possible to open PWAs to handle files on ChromeOS.
This CL allow PWAs that specify they can handle files in their manifest
to integrate with the ChromeOS file manager.

Allowing PWAs to access file handles they were opened with will be part
of a future CL, this CL purely causes the app to be opened.

This feature is gated behind two feature flags: NativeFileSystemAPI and
FileHandlingAPI.

This is part of the file-handling proposal:
https://github.com/WICG/file-handling/blob/master/explainer.md

Patch: 1/4

Bug: 829689
Change-Id: Ibbe5181f8d4dccd4a44bda756cea2268612e0241
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1597515
Commit-Queue: Jay Harris <harrisjay@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657949}