vaapi: introduce |num_extra_pics_| for H264 and VP8 allocation
Vaapi has a kNone allocation mode, in which we decode onto the client's
PictureBuffers' VASurfaces. Sadly, this doesn't work for VP8 nor H264,
because the power consumption regresses: crbug.com/910986#c32 and
crbug.com/910986#c35 explain why: when |decoder_| calls back to
CreateSurface(), it doesn't find |available_picture_buffers_| [1] and
it tries and tries in a busy loop. This is because most of the
VASurfaces are in use in the media pipeline -- we take this to be
kMaxVideoFrames (4) [2], but in practice is more like 7 to 9 (see
[3]). This situation doesn't happen in VP9 because this has a larger
amount of codec reference frames that are hardly ever exhausted.

This CL fixes that by introducing a member variable |num_extra_pics_|
that increases the amount of PictureBuffers allocated, accommodating
the deeper media pipeline.

Test: vp8, h264 decoding consumes less power on nocturne, results:
https://docs.google.com/a/google.com/spreadsheets/d/e/2PACX-1vSVUtbYO5JcdAeQbAh8K7he8Nwl52uJH2_5KkzS7ZzcJOITdx9OfpBoIWrKm7dEExjETQqZswvH_HvK/pubhtml#

[1] https://cs.chromium.org/chromium/src/media/gpu/vaapi/vaapi_video_decode_accelerator.cc?&g=0&l=1030
[2] https://cs.chromium.org/chromium/src/media/base/limits.h?g=0&l=24
[3] https://docs.google.com/document/d/1EAZUkmGcjyX9kmgDnArenNV3QPFP5AoRs9FRYVwpe4A/edit#bookmark=id.1hz04nk4zm6e

Bug: 910986
Change-Id: Ib15424b7f6b88c23f179cad2607a3298ad98396a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1576560
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658106}