[Password Manager] Skip decrypting passwords which look like plain text
On Android and ChromeOS, we have a mix of obfuscated and plain-text
password entries. Our decryption in POSIX knows to skip decrypting values
without the prefices "v10" and "v11". However, those are also possible
password values. Therefore, if decryption fails, we will presume that it
was a plain-text password, which was mistaken for an obfuscated password.
If the password starts with "v11", then it is definitely plain-text,
because Android and ChromeOS don't perform this kind of encryption.

Bug: 960322
Change-Id: I65b19b12ae896099c45acc6540d063b0c467d972
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1599181
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Christos Froussios <cfroussios@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Commit-Queue: Christos Froussios <cfroussios@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658129}