Cache the core information of the primary account in IdentityManager
SigninManager maintains the account ID of the primary account itself but
gets the primary account information from AccountTrackerService.
Meanwhile, AccountTrackerService removes its tracking of an account when
the refresh token for that account is revoked. These facts together
result in a buggy SigninManager API surface when the primary account
exists with its refresh token revoked: SigninManager will correctly
report that there is a primary account and will gives its account ID,
but will give empty account info if asked for the account info of the
primary account.

IdentityManager has inherited this behavior. This CL fixes the bug at
the level of IdentityManager by caching the CoreAccountInfo of the
primary account in IdentityManager, resulting in the IdentityManager API
surfaces behaving as consumers would expect (note that there are no
longer any external consumers of SigninManager).

This CL will result in behavioral changes for any consumers of
IdentityManager that are obtaining the account information of the
primary account when it is in a refresh token revoked state. These
behavioral changes should be *positive*, as the API now behaves as these
consumers presumably expect (returning the CoreAccountInfo of the
primary account instead of an empty CoreACcountInfo). One such consumer
is sync; the result for sync should be displaying more correct error
messages to the user in this case (notably, displaying that the primary
account has an error rather than that there is no primary account).

As a minor implementation note, this CL requires the addition of a new
observer callback from SigninManager to IdentityManager so that
IdentityManager can be informed precisely when SigninManager clears the
authenticated account ID; otherwise, clients observing the notification
that the primary account's refresh token was revoked as part of the
signout process can see inconsistent state (SigninManager no longer
thinks that there is a primary account but IdentityManager still does).

Bug: 862619
Change-Id: I07b0dac8c9c5a41be42b6974c68fd94cc2a888b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1581625
Commit-Queue: Pâris Meuleman <pmeuleman@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Auto-Submit: Pâris Meuleman <pmeuleman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658132}