Introduce AutomaticUsbPrinterConfigurer class
Introduces the AutomaticUsbPrinterConfigurer class. This class will be
owned by CupsPrintersManager. It observes CupsPrintersManager, and when
it sees a usb printer in the kAutomatic class, it installs the printer.

Eventually, we will wire up a notification to let the user know that
their printer is ready for use.

Bug: 951135
Change-Id: I79f2a447eef6adfebbbf8d16fb84549870703adc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1585472
Commit-Queue: Bailey Berro <baileyberro@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658274}