Re-enable layout provider unittests
This CL is re-enabling the layout unittest on windows.

Fonts size have been adjusted on MAC. They are now about
1 pixel larger than other platform, which make more sense
than the previous values. I cannot explain the choice of
the previous values.

These tests were flaky because some code in chrome tests were
modifying a system-wide flag that change the way fonts are
rendered.

We are assuming the default configuration: Anti-Aliasing is on.

Other unittests should not turn on/off the system-wide flag.
Modifying that flag may leave the developer computer or the
build bot in a bad state. Tests are likely to crash and may
not restore correctly the state; they may run in parallel and
have incorrect behavior.

This change needs to be landed after fixing flakiness:
  1) https://chromium-review.googlesource.com/c/chromium/src/+/1570527/
  2) https://chromium-review.googlesource.com/c/chromium/src/+/1569969

NOTE TO SHERRIFS:
 If any Layout test are flaky, please disable the test
 and assign the bug to me. I think I've found any cause of flakiness. If not,
 I'll dig more.

BUG=759870, 701241
R=tapted@chromium.org
CC=​​​​​​​​​robliao@chromium.org,asvitkine@chromium.org

Change-Id: Iab8f4adb58e565600bedecb1365e05bfdf931079
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1570653
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Michael Wasserman <msw@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658689}