viz: Fix presentation feedback frame token DCHECKs
Frame tokens are provided by the client and the only guarantee that's
enforced is that they're non-zero.  Therefore, we should not DCHECK
other invariants like not receiving two PresentationFeedbacks for the
same token.

Also add a deserialization error if the client sends a zero frame token
since we have DCHECKs for that anyway.

Bug: 958310
Change-Id: I4ffad483b46999c084f88ca62a0bb824522e6d37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1594070
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Commit-Queue: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658802}