Add scrollbar LatencyInfo type
In order to quantify the impact of moving scrollbar scrolls to the
compositor thread, we want end to end latency metrics that we can look
at during experimentation.

Add a new SCROLLBAR type to LatencyInfo to facilitate this, this
new type will be set on scrollbar-generated gestures so that they'll
flow to the end of the latency pipeline and be logged separately,
instead of as MOUSE latency as they are today.

The Scrollbar latency metrics are the same as the ones that are
currently gathered for wheel and touch scrolling, both for ScrollBegin
(the first scroll update in a scroll gesture sequence) and for
ScrollUpdate.

Bug: 954007
Change-Id: I362486be76a4427df271f1cf31e4d23c12857b40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1595296
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Timothy Dresser <tdresser@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658812}