Worker: Clear ResourceTimingNotifier on WorkerFetchContext::Detach()
This CL makes WorkerFetchContext::Detach() clear
WorkerFetchContext::resource_timing_notifier_ to cut off the cross-thread
persistent as soon as possible during worker thread termination.

Before this CL, I guess the notifier can outlive worker thread's
termination GC in WorkerBackingThread::ShutdownOnBackingThread(), and that
causes the failure on the parent thread as the issues say.

I tested this using the script in issue 960626, and confirmed the failure
doesn't happen with this CL. I also tried the script in issue 959508 but
couldn't reproduce it regardless of this CL. It will be tested by the fuzzer
instead.

Bug: 959508, 960626
Change-Id: I2663e5acddec0d9f88a78842c093c594fb57acb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1609024
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658989}