Orderfile: Update running benchmarks to catch any exceptions.
Added a catch to any exceptions produced by benchmarking as it is not
critical for the orderfile genaration but for assesing its impact on
memory and performance. Also, switched to use histograms to extract
Speedometer2.0 results as results-charts.json currently does not contain
any results.

Bug: 947012
Change-Id: I88d8367b2d70a0e716427088a4a784ee1841eb74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1609837
Reviewed-by: Matthew Cary <mattcary@chromium.org>
Commit-Queue: Monica Basta <msalama@google.com>
Cr-Commit-Position: refs/heads/master@{#659008}