Add kAccountManagerSubPage to Chrome OS settings pages
Account Manager is an "OS level" setting, not a "browser level" setting
and was somehow not in the list of OS level settings pages, causing the
crash linked in the attached bug.

Fix this by adding it to the list of OS level settings pages.

Bug: 962438
Change-Id: I46424cb8636b7600ad5d144561626fb6d83b378b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1609778
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Kush Sinha <sinhak@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659014}