Run unified consent tests on the bot ios-simulator-full-configs.
This CL runs the ios_chrome_unified_consent_egtests set of tests on
the iOS CQ and main waterfall. These tests were already run on the
downstream bots, but it makes more sense to run them on the Chromium
waterfall as it depend only on public code.

Bug: 960260
Change-Id: Ia396cef6109437e3f20176e7ba739d2581deaea8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1599011
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659015}