Improve NTPCatalog::create API and implementation
Change the API to receive a base::Value instead of the deprecated
class base::DictionaryValue (the implementation already check the
type of the passed value so does not assume base::Dictionary type).

Change the API to receive the base::Value by reference instead of
pointer as there is only one caller and it already check that the
pointer is non-null before calling NTPCatalog::create.

Use the base::Value::Find{String,List}Key methods instead of the
base::FindKeyOfType method as it reduce the complexity of the
implementation.

Bug: 646113
Change-Id: Icc3a484619786d5afcf0733ef01c07904635d481
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1609780
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Justin DeWitt <dewittj@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659053}