Fix inline reply opening with key press.
When opening the inline reply text view by clicking on the "Reply"
button, we get a LocatedEvent to start the ripple effect from. When this
is triggered via key press, after focusing the button, the event is not
a LocatedEvent and we hit a CHECK.
This fix shows the default ripple instead if we activate via keyboard.

Bug: 961233
Change-Id: I1f6488f9ce240b9c2cec07ff7342faa75781635a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1602507
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Richard Knoll <knollr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659054}