CacheStorage: Hold self-ref during cache operations.
The LegacyCacheStorageCache must be kept alive as long as there are
any outstanding disk_cache Entry objects associated with it.  This is
for two reasons:

1) If a partial write has occurred, then removing the cache will
   cause any outstanding runnables to be discarded destroying the
   ScopedEntryPtr.  The ScopedEntryPtr will cleanly close the entry
   causing it to remain in a truncated state.
2) A second LegacyCacheStorageCache object could be created immedately
   after and open a new backend that stomps on our existing Entry
   objects.

This CL solves the problem by making LegacyCacheStorageCache hold a
handle reference to itself whenever its performing a public operation.
This will in turn cause the owning LegacyCacheStorage to keep the
cache object alive.

Bug: 961752
Change-Id: Icf333341d51c4acd0688626977690d73408c43ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1606200
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659087}