Removing layout size calls from AnchorElementMetrics OnClick
The clicked element will no longer report its size (or relative size) to
the browser/UMA. Instead, we plan to see if it is in the Anchors
reported on load. Its size would only be valuable if we knew about the
anchor from OnLoad.

Bug: 957026
Change-Id: I59a98348fd0d58a4dd1a99bd86c2720b36f1af88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1602773
Commit-Queue: Ryan Sturm <ryansturm@chromium.org>
Reviewed-by: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659283}