Decouple AOM sparse attributes from the ax tree
This change removes the write capabilities of Element.accessibleNode,
for string properties only. Other attributes will be decoupled in future
based on type, and sparse attributes are not part of this change.

This is due to a simplification of the API to avoid computing the
Accessibility Tree in order to modify it. See:

https://github.com/WICG/aom/blob/gh-pages/explainer.md

BUG=746523

Change-Id: I63ccd52f7193f017559402d76d4c446423e27743
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1605322
Commit-Queue: Meredith Lane <meredithl@chromium.org>
Reviewed-by: Alice Boxhall <aboxhall@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659351}