media/gpu/VFMapper: Create VideoFrameMapperFactory with PixelFormat
Formerly, VideoFrameMapper is created without any info. It may be possible that
some unsupported pixel format is fed into VFMapper. We should create VFMapper
with VideoPixelFormat and the VFMapper should only accept the set pixel format.
Pixel format is not decided when VFMapper is created in VFValidator and
VFWriter. So this delays the creation of VFMapper in these classes.

Bug: 952147
Test: video_decode_accelerator_tests
Test: video_decode_accelerator_unittest
Test: video_encode_accelerator_unittest
Change-Id: I1baf80fa4392c74e0d82a4239f791363ef46fe2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1573311
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659374}