vk: Move //ui/keyboard to //ash/keyboard/ui
To make this CL as small as possible, we're just moving the directories
and updating any references.

We had to add a few DEPS from chrome to ash/keyboard/ui.

In future patches well consider what files should go where and whether
we want to keep keyboard code in its own self-contained module.

TBR=benwells@chromium.org

Bug: 961568
Change-Id: I27c140ab51709ebb89910fe43b96bf64a38100d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1605584
Reviewed-by: Ben Wells <benwells@chromium.org>
Reviewed-by: Oliver Chang <ochang@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659376}