Switch WebViewGuest to using UnescapeBinaryURLComponentSafe().
It was using UnescapeURLComponent(), which, unlike
UnescapeBinaryURLComponentSafe(), leaves certain characters escaped. The
options passed in left spaces, path separators, control codes,
"special characters", and UTF-8 sequences unsafe to display in the
omnibox escaped. Since the string passed in is created by
EscapeQueryParamValue(), however, none of that should be needed.

UnescapeBinaryURLComponentSafe() escapes all input characters, but fails
on path separators and control codes. Unclear if the failing behavior
matters here, but since the previous code left path separators escaped,
seems the safest option.

This CL is part of a project to clean up uses of
UnescapeURLComponent(), as it's been used for wildly disparate
purposes, and much of its logic is only suitable for cases where
we want to format a URL for safe display to the user. We want
to make it leave more weird characters escaped, but need to
eliminate usage of it in other contexts before we can do that.

UnescapeBinaryURLComponent() should be used instead in most cases,
particularly where URL_SPECIAL_CHARS_EXCEPT_PATH_SEPARATORS is in use,
as all consumers of that option seem to either be slightly buggy, or
don't actually need to leave any characters unescaped. The primary
cases where switching to UnescapeBinaryURLComponent() may cause
problems (without additional checks), are cases where unescaping
path separators or funky characters (like %01) is a problem. In those
places, we should generally fail instead of partially unescaping a URL.

Bug: 849998
Change-Id: I498e0471a95f183c18c34eec5e3732804dbc225a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1600085
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659514}