[LayoutNG] Don't force legacy layout on ::first-letter pseudos.
Due to the complicated behavior of ::first-letter elements, where the
placement of the ::first-letter DOM node in the DOM tree doesn't
correspond with where we place the LayoutObject in the layout tree, it's
tricky to figure out whether it's part of a subtree that requires us to
force legacy layout or not. So just skip checking. The ::first-letter
element itself can never force legacy layout on its own, so we'll just
do whatever the parent layout object does.

Solution developed together with futhark@chromium.org

Bug: 962090
Change-Id: Ie112561bbd58c0c6f822f500247dbbaa1c8a8118
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1611602
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659529}