[LayoutNG] Put line cache under a new runtime flag
This patch puts LayoutNG line cache under a new runtime flag
which is not enabled by default.

There have been discussions on this, for its cost/risk and
benefits. As we approach to ship, more risk cases were found.

Also with recent normal layout caching work by ikilpatrick@,
we're hoping to find better integration of line caching with
block caching.

It is known that this hinders the performance on appending
micro benchmarks and in editing. We will probably need to
revisit in post-phase 1, but we can polish it better with
the data and the experiences we've got so far.

Bug: 962008, 962141
Change-Id: Iee162be41b36ce5462a89efbbdd7abe796c3a14e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1611068
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659552}