[Extensions] Update GetEffectiveDocumentURL() to use precursor tuples
Opaque origins (such as about:blank and data: URLs) inherit the security
and privileges of their creator origins, which are tracked in precursor
tuples. ScriptContext::GetEffectiveDocumentURL() attempts to determine
the precursor origin of an opaque URL today by traversing the creation
chain and looking for the first non-about:blank opener or parent.
However, this is not always strictly correct for all opaque origins;
for instance, the precursor tuple of a data: frame may not be that of
its immediate parent in the case of nested iframes.

Update ScriptContext::GetEffectiveDocumentURL() to use
Origin::GetTupleOrPrecursorTupleIfOpaque(), which correctly tracks this.
Note that we still have to do the opener/parent traversal case because
content script matching needs to have the full URL (including path), and
not just the scheme, host, and port tuple.

To balance these, we simply return the document URL (which is
potentially opaque) in the case of a mismatch between opener/parent
traversal and the precursor tuple. This should be an exceptionally
rare case in practice.

Updating this method to use the precursor tuple is a prerequisite for
adding support for data:-URL content script injection.

Bug: 55084
Change-Id: I8f7d17b999f909cc32091b8fff06bd53a2370db7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1601663
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659599}