Log SCROLLBAR LatencyInfo's for main thread injected scrolls
Note: based on crrev.com/1595296

When scroll gestures are injected during the handling of an input on
the main thread, we want to dispatch the scroll gesture with the
LatencyInfo information from the original input event, but with a
different type. This allows the end to end latency metrics to be
computed correctly and classified separately as a specific type of
scroll (not just, e.g. as mouse latency).

Bug: 954007
Change-Id: I725eaa46f21883aff08ab209d73dc5d0398a85db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1603342
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Timothy Dresser <tdresser@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659630}