Attempt to close windows that get in the way of focusing a native window.
From time to time, a stray window is left behind on the Windows 10 bots
that interferes with interactive_ui_tests (it's usually an OpenWith.exe
dialog asking the user to pick an app to open some file/URL protocol).
This change augments ShowAndFocusNativeWindow so that it makes up to
five attempts to close any window that remains in the foreground despite
its best efforts to bring the desired NativeWindow to the foreground.

BUG=711256
R=sky@chromium.org

Change-Id: Iccb568f34c66b5d0571d6d4dc3648c7b06ac35de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1610766
Commit-Queue: Greg Thompson <grt@chromium.org>
Auto-Submit: Greg Thompson <grt@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659635}