Moved signin_test_util#SignOutAndClearAccounts to chrome_earl_grey
This change requires for /ios_internal repository(submit only after this cl)
https://chrome-internal-review.googlesource.com/c/chrome/ios_internal/+/1284283

This is part of EG1 to EG2 migration, which involves moving EarlGrey code from
app-side helpers into test code. For GREYAsserts, the policy is to take assert-
containing-helpers, modify them to return NSError* with localizedDescription
containing the assert failure (in the event of a failure), and then assert
on the result of the helper on the test side.

Bug: 922813
Change-Id: I42086810e5efe23139210fa1e198c14665f903bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1606248
Commit-Queue: Maksym Onufriienko <monufriienko@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659657}