Remove use of NetLog from Cast client code.
Chrome's network stack runs out of process in most places, so logging
actions to the in-process NetLog accomplishes nothing. Cast should be
grabbing a UDP socket using the NetworkContext interface, instead of
creating an in-process socket (Which will also get its network events to
appear in the out-of-process NetLog).

Bug: 767450
Change-Id: I3c742689f0b4a29138fd63669928b3a55d5ee2d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1612056
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Yuri Wiitala <miu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659747}