Ignore ancillary chunks in PNG header
When parsing a PNG header (i.e. up to the IDAT), skip ancillary chunks
that the decoder does not need (rather than passing to libpng, as was
done previously). This can speed up decoding (as well as reduce memory
allocations) when a PNG header contains large and/or many unneeded
chunks (e.g. tEXt). This is important because the header is parsed in
the main thread. (Note that for simplicity, this change still passes
unneeded chunks to libpng during decoding, which occurs in a background
thread.)

A resulting behavior change (side-effect) is that a PNG containing an
ancillary chunk with a CRC error prior to the IDAT would previously have
made the decoder Fail and no longer does. This is acceptable because
these chunks are ancillary and can be ignored. Note that an IEND chunk
before IDAT *is* passed to libpng, because that *should* result in a
failure.

Add a test to verify that we still handle tRNS correctly. Other chunks
are tested by web_tests.

Bug: 956259
Change-Id: I4d159597e99adcfcfe5cf81f9057624eae541de2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1599916
Commit-Queue: Leon Scroggins <scroggo@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Auto-Submit: Leon Scroggins <scroggo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660051}