[Password Manager] Disable uploading noisified info about special symbols
We may need to disable the uploading and merge it to late Beta. With this CL, I test that disabling doesn't cause any crashes or other issues. If we don't need to disable uploading, this CL will be reverted.

The uploading was introduced in https://chromium-review.googlesource.com/c/chromium/src/+/1567751.

Bug: 953668, 552420
Change-Id: I6e85bc047bcf1a263b60c5a0cbea19eb5e86f576
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1615023
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Commit-Queue: Maxim Kolosovskiy <kolos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660352}