webauthn: remove the CREDENTIAL_NOT_RECOGNIZED status
The WebAuthn spec says that during a WebAuthn get() request, if an
authenticator returns an error in response to the request, the browser
should simply ignore that device and let the user try a different one
until the user cancels or the request times out, at which point the
browser returns a "NotAllowedError" DOMException.

Chrome diverges from the WebAuthn spec a little in that errors from the
authenticator can cause termination of the entire request (depending on
the exact nature of the authenticator's response). Originally, this
resulted in the WebAuthn request's promise to be rejected with an error
immediately.

To allow the relying party (RP) to distinguish a request that was
terminated due to a security key not recognizing any of the presented
credential IDs from other errors (such as the user clicking "Cancel" in
the UI), Chrome therefore returned an InvalidStateError for this
particular case, such that the RP could then inform the user that the
security key that they tapped wasn't registered with the site.

Later, Chrome's WebAuthn implementation added an error to its own UI for
this case that informs the user that their security key wasn't
recognized. But Chrome still returns the InvalidState error with a
special message. This change aligns Chrome's behavior with the spec and
makes it return NotAllowedError in this case.

Change-Id: Ib2a3c3134555cfc87da451a5c5738b59df674139
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1602745
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660519}