Move application locale for fallback fonts to RenderText_harfbuzz
This CL is lifting the locale passed to font fallback into RenderText.

The locale is used to resolve fonts when they are localized.
The HAN unification rely on locale to decide which font to used.
Glyphs are not displayed the same depending of the locale.

Adding unittests for fonts fallback using on different locale.
Adding a unittest to check that RenderText is reading and passing
the locale to the fallback fonts code.

Bug: 944227

Change-Id: I5c2997e34174a1a32c354476e77a1e04b9656429
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1610569
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660579}