Fix Profile::IsNewProfile on first run
On some platforms (windows in particular), the installer may create a
preference file in the profile directory, before the profile is created.
This happens in first_run::ProcessMasterPreferences().

In this case, when the profile is created, the preference file is found
and the profile is assumed to not be new.
As a result, all the code that is intended to be called for new profiles
is bypassed, which leaves the profile in a bad state.

Bug: 728402, 959188, 145351
Change-Id: Ibfeeebbb72893cc48095ce810d8300718dc24e6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1614207
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660847}