Do not create PrefetchGCMAppHander in reduced mode (relandx3)
If in reduced mode, we cannot create PrefetchGCMAppHandler. Instead we pass the
prefetch service a closure to create the gcm app handler on demand.

This cl also changes the signature of the getter for PrefetchGCMHandler in
PrefetchService to require a profile to be passed in. This ensures that:
1) the getter can only be called in full browser mode (there is no profile in
reduced mode).
2) if PrefetchGCMHandler hadn't been created yet, then it can be
created on demand using the aforementioned profile.

> This is a reland, Original CL:
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1495737

This reverts commit a59ecdc7b32b83272c02a83a2ae0d2876d03b0ba.

Reason for reland: Fixed CQ tests being slow by mocking out InstanceIDAndroid
during tests.

This reland also moves away from deprecated call to PostAfterStartupTask.

Bug: 934337, 963688
Change-Id: I8d57b1f96f0b6152ba3271a9a7ef7e023f5f6988
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1613672
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660857}