overview: Allow dragging windows on nonprimary displays in tablet mode
In tablet mode, if you start to drag a window from the top of a
nonprimary display, before the present CL, the functionality is
completely broken. You do not even see the window. If you try to drag a
window from overview on a nonprimary display, before the present CL, the
drop target is created on the primary display. After the present CL,
dragging a window (either from the top or from overview) on a nonprimary
display shall be similar to dragging a window on the primary display.
Future work will address functionality where a window is dragged from
one display to another.

Test: ash_unittests OverviewSessionTest.*
Bug: 916856, 964063
Change-Id: Iddbafc7b289ab2f711db47ebaf18bfcfcda2539c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1617906
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661110}