Remove card elevation in Tab Grid
The elevation is rendered slightly later than the View, and this
causes flickering when entering the Grid Tab Switcher (GTS), or
switching between normal/incognito mode in GTS.

The rendering of elevation is hardware-accelerated, and cannot be
captured in View.draw(Canvas). Removing the elevation eliminates
this discrepancy, and make it possible to have a smoother transition
from the captured Bitmap to the real Android View. We'll need this
when making the Tab to GTS transition animation.

Bug: 964406
Change-Id: I725f12a8dd186d4e6f848ef7bdef9a1cc8b9556d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1617213
Commit-Queue: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661165}