ime: Change MODECHANGE key to show IME mode indicator if not visible
Currently, the MODECHANGE key just switches to the next IME (same as
pressing Ctrl-Shift-Space). However, because users may accidentally
press the key, we want it to be a no-op initially. So the user pressing
the key once just shows the current IME. Pressing it again while the
IME mode indicator is visible will switch to the next IME.

TBR=holte@chromium.org

Bug: 953901
Test: accelerator_controller_unittest.cc; manual test with fake key.
Change-Id: Ibf0cc0379d35c4aae50f5f41a2dceb1e14de3b34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1614997
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Oliver Chang <ochang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661239}