Use Mount() to restore an SMB share instead of using Remount()
The Remount() D-Bus method attmepts to remount the share with the same
mount_id. However, mount_id is like a file descriptor and not suitable
as a persistent id. Trying to restore a share to the same mount_id
causes the remount to unexpectedly fail in certain situations (i.e. if
the browser process needs to restart).

The alternative is to use the regular Mount() method and have it skip
attempting to connect to the share. That will give us a new,
non-colliding mount_id. Since there's no way to pass this data to the
SmbFileSystem, it is stored in a map in SmbService and looked up by
SmbFileSystem when needed.

BUG=961183

Change-Id: I40c858541fb065a6f88a51249fd72678bdba862d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1609753
Commit-Queue: Anand Mistry <amistry@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661264}