[Files app] Add role and aria-describedby for file list column headers
This allows screen reader users to know that the column header is
clickable and also what is the sorting order it will apply when
clicked.

Use aria-described to point to the text (with translation) that
describes the action performed when clicking in the column header.

Use aria role=button so screen reader recognizes column header as a
clickable element, Chromevox will say:
"Press Search plus Space to activate."

Add test that check for those 2 attributes in all column headers.

Remove 2 missing files from the test harness:
test_util_unittest.js and details_panel.js

Change cr.ui.table.TableColumn.headerRenderFunction to accept Element
in addition to just plain text, so we can customize further the column
header, in this case for add these ARIA attributes, but in the future
it needs a "ripple" effect when clicked too.

Test: browser_tests --gtest_filter="*fileListAriaAttributes*"
Bug: 888620
Change-Id: I6d56d94a2d59aad3b416133106f569cc495fb944
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1610669
Reviewed-by: calamity <calamity@chromium.org>
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661266}