service worker: Simplify deletion states on ServiceWorkerRegistration.
Before this CL, ServiceWorkerRegistration had three states:
- is_deleted
- is_uninstalling
- is_uninstalled

It's difficult to understand the differences. The tension that caused
the states seems to be that |is_deleted| was controlled by the lower-level
storage layer, and |is_uninstall{ing,ed}| by the higher-level API layer
like the implementation of unregister().

This CL aims to simplify things by removing the is_deleted flag. The
layers get a bit more blurry but the benefit is fewer states. There are
likely further cleanups and renamings possible to make this nicer.

Change-Id: Iee1102da0a865f571bc8906e184dd16551541c1e
Bug: 964201
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1616870
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661282}