blink/canvas: Fix shared image mailbox tracking in StaticBitmapImage.
The AcceleratedStaticBitmapImage can transition between a mailbox or
skia backed TextureHolder multiple times. When a mailbox texture holder
is converted to a skia texture holder, we import the mailbox to a
texture correctly and wrap it in an SkImage. But after this point, we
loose all information about the previous mailbox. If we now transition
from skia to mailbox texture holder, then we'll create a new mailbox to
associate with this texture using ProduceTextureDirectCHROMIUM.

The bug is that the state for whether the mailbox is associated with a
shared image is stored on AcceleratedStaticBitmapImage. After the
transition above, the mailbox held by the image is no longer a shared
image mailbox but the code would incorrectly assume it is.

The change here addresses this bug by storing this state on
MailboxTextureHolder so it is tightly coupled with the lifetime of the
original mailbox. Long term though, we should keep retain the original
MailboxTextureHolder on AcceleratedStaticBitmapImage since after the
transitions above, the new mailbox can not be used in any interop
scenario.

R=kbr@chromium.org, piman@chromium.org

Bug: 961777, 900706
Change-Id: Ibba933c87684fcda210194a65ac314c9f9f18142
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1618510
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Commit-Queue: Khushal <khushalsagar@chromium.org>
Auto-Submit: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661567}