Updated the orderfile generation pipeline to support call graph.
Updated the orderfile pipeline to support generation of an orderfile
based on the new instrumentation call graph information. To enable using
the call graph instrumention, set |--use-call-graph|. This CL uses the
counts of how many a pair of (caller, callee) was recorded to define the
distance between them. It then constructs the neighbors, and uses the
existing clustering algorithm to produce the orderfile.

Bug: 963833
Change-Id: Idd07a7a0aa5e00d87581fc68fe60374fe426d24c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1615022
Reviewed-by: Matthew Cary <mattcary@chromium.org>
Commit-Queue: Monica Basta <msalama@google.com>
Cr-Commit-Position: refs/heads/master@{#661708}