Adding windowed CPU metrics observation.
Adds tracking over a 30 second window of the cpu usage for all ad
frames and the page as a whole.  Keeps track of the peak usage number
for each frame prior to frame activation as well as the page (which
is never considered activated).  For pages where the navigation is
under 30 seconds, usage is still calculated based on a full 30
seconds (the overall time spent in the cpu is divided by 30 seconds
to find the percent usage).

Bug: 964021
Change-Id: Id7a40743c1bdaae3326f075265dbcc22bb00e682
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1612852
Commit-Queue: Eric Robinson <ericrobinson@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661740}