[DevTools] HeadlessShell::OnReadyState - Check the unique_ptr.
The reason this is nullptr is the conversion code
in Domain::HandleEvaluateResponse.
https://cs.chromium.org/chromium/src/out/Debug/gen/headless/public/devtools/domains/runtime.cc?q=headless::runtime::Domain::HandleEvaluateResponse&sq=package:chromium&g=0&l=349

So now we check the unique_ptr directly. I think if that's non-null,
then the GetResult() will succeed because for the domain-specific object,
result is a non-optional field so it's always set to something.
So I hope (since we have builders and all). So I'll leave out that check for now.

BUG=961299,958667,958666

Change-Id: Ic1ff77655f484e33d37d1d18bdcc1c28da15fd1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1622792
Reviewed-by: Alexei Filippov <alph@chromium.org>
Commit-Queue: Alexei Filippov <alph@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661921}