[FastBorderRadius] Re-land: Adjust rounded corner bounds when aggregating quads into a new target.
The rounded_corner_bounds field of SharedQuadState needs to be in the
space of the target render surface. When changing that target, therefore,
rounded_corner_bounds needs to be updated accordingly.

This re-land fixes a bug where we incorrectly applied a transform to rounded corners
on a surface draw quad, which is already in the space of the containing surface.

Bug: 960436, 947715

Change-Id: Ia5db2566392c1e6fc711dce18ca4ab18b63ff948
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1620108
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: enne <enne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661929}