WebUI: cr-link-row, add role and simplify aria attributes
As part of work to replace usages of paper-icon-button with
cr elements in the ChromeVox options, cr-link-row needed a slot for
a label and subLabel. This is due to how the ChromeVox page loads its
i18n strings. I added the link role to the button and defined static
aria-labellby and aria-describedby.

Here is the CL to replace paper-icon-button with cr elements:
https://chromium-review.googlesource.com/c/chromium/src/+/1618564

Bug: 960564
Change-Id: I7934a13911d4c7572db8701542560859e4af7d3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1622476
Reviewed-by: Hector Carmona <hcarmona@chromium.org>
Commit-Queue: Esmael El-Moslimany <aee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661945}