Fix backdrop filter capture rect for software_renderer
Previous to this CL, backdrop-filters with pixel moving filters
would cause the capture rect for the backdrop-filter to be
improperly located. In those cases, the capture rect was expanded
for the filter, but the bounds were not properly updated. This
is now fixed.

I added a unit test to capture this corner case.

Bug: 964371
Change-Id: I5e9833b7d8460feaafca9dd29fe4a469fb6a55f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1618304
Commit-Queue: Mason Freed <masonfreed@chromium.org>
Commit-Queue: enne <enne@chromium.org>
Reviewed-by: enne <enne@chromium.org>
Auto-Submit: Mason Freed <masonfreed@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661974}