Fix ReportMissingFormManager metric when old parser is disabled.
Old parser is disabled is final state of the launch of the new parser,
so everything should work for it. PasswordManager::ProvisionallySavePassword
sends ReportMissingForManager metric even if old parser is disabled (and
then it sends it incorrectly). This metrics is sent for NewPasswordFormManager
in PasswordManager::ProvisionallySaveForm.

Along the way extracted sending of SubmittedFormFrame metric to a separate
method and called it for NewPasswordFormManager.

Bug: 831123
Change-Id: If170658a030e2b34b8aa7b9bd70cc048d8de0fab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1624573
Commit-Queue: Vadym Doroshenko <dvadym@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662200}