[DIAL MRP] Don't send "stop" receiver action if a request is pending
When the DIAL MRP sends a request to stop casting to a receiver, it also
notifies the sender of the receiver action. This may cause some senders
to submit another stop request, causing a loop that continues until the
app is actually terminated on the receiver. This CL prevents that.

Bug: 958529
Change-Id: Ife55002096bd5d57b689021c4901fad4f4c4100c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1610531
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Commit-Queue: Takumi Fujimoto <takumif@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662260}