Need to wait another frame for spatnav.
As part of trying to remove the forced UpdateStyleAndLayoutTree on
parsing finished, this test started failing. Parsing the document would
finish and force a style update after display is set to none, but before
the promise is resolved through rAF. Add another rAF to allow for a
lifecycle update of the computed style instead.

Bug: 742413
Change-Id: I91d21b5cd871d82d81dd8fb156465762ad594318
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1627355
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662599}