Present the advanced sign-in settings from First Run.
This CL adds a temporary solution to present the advanced sign-in
settings when the user taps on settings from the FRE sign-in screen:
* is adds a command that open the advanced sign-in settings
* it calls this command from the FRE screen.
* it updates the failing earl grey test.

This is a temporary solution as the plan is to move the ownership of
the FirstRunChromeSigninViewController to the SigninInteractionCoordinator
(see crbug.com/965992). However it is preferred to only make this change
after the unified consent feature is launched as it is fairly complex to
do if we want to also support pre-unified consent flows.

Bug: 963406

Change-Id: Ie291dc7f935e5b5e0b100888b3f55c81fb64558e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1621176
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662615}