GMC: Make MediaNotificationView independent of MessageView
This CL adds a new MediaNotificationContainer abstraction to the media
notification so that the MediaNotificationView no longer depends on Ash
or MessageView. This will allow media notifications to be created that
don't need to be sent to the MessageCenter. Instead of
MediaNotificationView being a MessageView, it will be contained in a
MessageView in Ash when it needs to be sent to the MessageCenter.

This is part of a refactoring of the media notification to make it
reusable outside of ash/media.

Bug: 959478
Change-Id: Iac96211b3feb6e640b5e79a61832927c72a4fa88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1616953
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662617}