[ MimeHandlerView ] Block PluginStatus != kAllowed/kPlayImportantContent
Currently, a MimeHandlerView is created as long as there is a plugin
for it. This behavior undermines issues such as <webview> permissions
API.

This CL will avoid creating MimeHandlerView if the plugin is not
allowed. This change will also make the test
WebViewPluginTest.TestLoadPluginInternalResource pass with the flag
MimeHandlerViewInCrossProcessFrame.

Note that when the plugin is not allowed, no (frame-based)
MimeHandlerView will be created and the code will follow the ordinary
CreatePlugin code path.

Bug: 963694
Change-Id: I91b88956480139ffdb0df06077d099746b713751
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1622890
Reviewed-by: Ehsan Karamad <ekaramad@chromium.org>
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Ehsan Karamad <ekaramad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662641}