[Bluetooth] Create stability metrics for reconnection attempts.
Create metrics to capture the success rate of user-initiated
attempts to reconnect to an already paired device, for both relevant UI
surfaces, Settings and System Tray. The distinction of "user-initiated"
is key; it's a signal that the user expects that device to be present
and connectable, so we can trust with reasonable certainty that
most failure emissions to these metrics represent "true" failures or
errors in our Chrome or platform stack.

Bug: 958452
Change-Id: Ic8f551deff7c4a4cf83c6a974d3d83c5fcb423b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1626637
Commit-Queue: Ryan Hansberry <hansberry@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662864}