Corrected mirroring updates in ime_controller updates
There was a misunderstanding in how OnDisplayMetricChanged's
changed_metrics worked. The implementation in ImeController
assumed that the DISPLAM_METRIC_MIRROR_STATE will be true when mirroring
is occuring.

What actually happens is that OnDisplayMetricChaniged will have it's
changed_metrics bitflag set to true when the metric has changed value
(from true to false or false to true)

The implementation has been updated to reflect that and the function
documentation was updated to reflect that more clearly.

Change-Id: I110b5b1fbee11f1c761b2c5d715b7034d8c44cd6
Bug: 824656
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1600616
Commit-Queue: Keith Lee <keithlee@google.com>
Auto-Submit: Keith Lee <keithlee@google.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Shu Chen <shuchen@chromium.org>
Reviewed-by: Darren Shen <shend@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662872}