Use FrameTreeNode id, instead of WebContentsGetter, in WebContentsGetterRegistry
This is a preparation CL for https://crrev.com/c/1614723.

As per the review comment [1], we can use FrameTreeNode id, instead of
WebContentsGetter, in WebContentsGetterRegistry because we can get
WebContentsGetter from FrameTreeNode id, as well as to know whether the frame
is mainframe or not from the id.

WebContentsGetterRegistry is now renamed to FrameTreeNodeIdRegistry.

[1] https://crrev.com/c/chromium/src/+/1614723/12/content/browser/web_contents/web_contents_getter_registry.h#16

Bug: 623464
Change-Id: Ifab7727f6a6b74408f815e5b853d75d39c515f9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1623611
Commit-Queue: Hayato Ito <hayato@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662992}