Tracing: Defer PID book-keeping until connection ack
This adds an ack response to TracedProcess.ConnectToTracingService and
defers internally registering the PID with PerfettoService until the ack
is received. This ensures that an unresponsive or otherwise defunct
service cannot indefinitely block tracing for other services in the same
process.

Also changes chrome_renderer instance registration back to a mode where
it never reports a PID (as opposed to implicitly reporting its PID is
the same as the browser's). This adheres to subtle but valid
expectations in tracing code that  services reporting a given PID are
actually running in that process.

Bug: 962918
Change-Id: I71273f8ba954f580c1375e9b0ce2ab6c148b69b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1626991
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: oysteine <oysteine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663001}