🦁Cache default search engine favicon in omnibox
There is already code in the omnibox_mediator to load the default
search engine's favicon every time it changes or the consumer is set.
We also load this favicon every time the top suggestion is a search
suggestion. Caching the favicon would simplify all this loading.

Bug: 966771
Change-Id: Iaef03730b603cb1dff6af163f6ebcaa2d9c9fb1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1628611
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#663064}