overview: Call OverviewGrid::RearrangeDuringDrag() on one grid, not all
During dragging from overview, OverviewGrid::RearrangeDuringDrag() gets
called on all overview grids, not just the one where an item was dragged
from. Something like that might eventually make sense for purposes of
the behavior described in Comment 3 on Issue 961171 (see [1]), but for
now, it is nothing but disruptive, as shown in the videos linked in
Comment 6 on Issue 940710 (see [2]).

[1] https://bugs.chromium.org/p/chromium/issues/detail?id=961171#c3
[2] https://bugs.chromium.org/p/chromium/issues/detail?id=940710#c6

Test: manual
Bug: 940705, 940710
Change-Id: Id13b1e11b67ae04bd095cb047ca59632c3f3b7aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1628484
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663214}