[IndexedDB] Handle OriginState reentry during transaction abort
Aborting transactions can cause databases to be deleted. This caused
reentry in the OriginState while iterating the databases_ map in
AbortAllTransactions. This change grabs the keys to the map first, so
it can iterate the keys instead of the map.

R: pwnall@chromium.org
Bug: 966784
Change-Id: I42a8710aff9d25deb3d9eace9fa06654bba46507
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1629190
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663308}