Improve smooth scaling for reflected surfaces.
With OOP-D enabled the Chrome OS software mirroring and unified desktop
features use surface embedding to mirror content from one display to
another. This causes scaling artifacts when the source and destination
displays are not the same size.

Plumb through a bit with SurfaceLayer/SurfaceDrawQuad to indicate the
Surface is being reflected. In SurfaceAggregretor we don't merge
RenderPasses for the SurfaceDrawQuad if it has the |is_reflected| bit
set and it has a transform that will scale it.

With this CL the destination display will draw the source display
content into a RenderPass at its original size. The RenderPass will then
be copied with AA enabled into the final frame. This matches the visual
behaviour without OOP-D when Reflector and TextureLayer are used.

Bug: 601869
Change-Id: Ic794caa63fe62a66ac4af4cfc8379375bdbfa6da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1627931
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: weiliangc <weiliangc@chromium.org>
Commit-Queue: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663629}