[signin] Remove inconsistent url loader factories
Several classes were taking both a SigninClient and a URLLoaderFactory.
However, SigninClient has a GetURLLoaderFactory() method, and this
was leading to a situation where code had access to two instances of
URLLoaderFactory, and would use one or the other in a random fashion.

In production, these factories were the same in practice, but in tests
this was not always the case. As a result, it was not clear what the test
were doing, and it also made the code harder to understand.

This CL removes this confusion, by always using the factory from the client
when it is available.
It also removes the unnecessary url_loader_factory parameter from
SigninClient::CreateGaiaAuthFetcher.

TBR=eugenebut

Bug: 863347
Change-Id: Ie47b56226b126fbdc753cfd170c7fead65d6952b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1569137
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663696}