Remove kApplyContainerFlip geometry mapping flag
Previously in many geometry functions, kApplyContainerFlip controlled
whether the input (when mapping from local to ancestor) or the output
(when mapping from ancestor to local) is in flipped blocks direction
for now boxes. Many geometry mapping functions forced the flag by
default, while others allowed the caller to specify the flag, which was
inconsistent.

Now remove the flag and let the functions always deal with physical
coordinates, and let the callers to meet this requirement.

Change-Id: I02c0da9d27fca1a0e2849bde3213c69791e2ad2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1614448
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663835}