Removing External LocalSurfaceIdAllocation from WindowTreeHostPlatform
WindowTreeHostPlatform supported bounds changes where the source of
the LocalSurfaceIdAllocation was external. This was to support Mus,
where the Window Service would allocate new ids when triggering resizes.

On other Aura platforms there were DCHECKs that no other source was
setting these ids. As WindowTreeHostPlatform is intended to be the
root allocator.

With the removal of Mus this is no longer needed.

TEST=aura_unittests, manual testing of unified mode
Bug: 958239

Change-Id: I595a147cb38c454e3199e9cab9d180719adc9711
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1625801
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663847}