display: Retain content protection requests
ApplyContentProtection submits an ApplyContentProtectionTask that
aggregates existing per-client per-display requests, plus the new
request on the given display. The per-client per-display requests
are updated when the task succeeds. The ContentProtections passed
to ApplyContentProtectionTask may omit displays, which means that
protection is not requested and should be disabled.

This logic is flawed: submitting two successive requests enabling
protection on one display and disabling it on another one results
in protection being disabled on both. This notably happens when a
window with protected content is dragged across displays.

This CL fixes this by tracking requests as long as their client is
registered. It also adds validation for client IDs so that pending
callbacks don't run if the client is unregistered in the meantime.

Bug: 929449
Test: Check HDCP status with modetest in above/other scenarios.
Test: display_unittests
Change-Id: Id03702fd9673cee2b925b5ba043d086256b4e360
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1587655
Reviewed-by: Daniel Nicoara <dnicoara@chromium.org>
Commit-Queue: Dominik Laskowski <domlaskowski@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663849}