Introduce services GRD
There is a need to express localizable strings in //services code.
This CL introduces all the bits necessary to define translated resources
in a new //services/services_strings.grd file and moves a single string
(pertaining only to a service defined in //services) from the Chrome GRD
to the services GRD.

This is a precursor to simplifying expression of out-of-process service
registration by allowing service manifests to include a resource ID
directly in their definition.

Bug: 895615, 915806
Change-Id: I76d420266fb5528f407b624838172c267666654b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1615701
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#663919}