Call RegisterInProcessThreads for Android chrome browser tests.
RegisterInProcessThreads() is called to set up using single process with
child processes running in threads. It must be called for:
- Chrome unit tests
- Android Content browser tests
- Android Chrome browser tests

Currently it is called in ContentBrowserTestSuite and
ChromeUnitTestSuite which excludes the 3rd case. This adds a
call in ChromeTestSuiteRunner, which is used to run Chrome browser
tests.

Bug: 961849
Change-Id: I371eb8777ba1fbe8b1dd7f4426dc15144392cad9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1610717
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663945}