AW: addJavascriptInterface test helper method
No change to production logic, mostly a no-change refactor in tests.

This adds the AwActivityTestRule.addJavascriptInterfaceOnUiThread()
helper method, which avoids some need for posting to UI thread manually
in tests.

This replaces all uses of bare addJavascriptInterface(), most of which
are mechanical (NOOP) replacements. The non-trivial replacement is
AwContentsTest, which had multiple operations in the same UI-thread
section. Now we have multiple posts to UI thread, which is slower, but
should be more readable.

This is a follow-up CL to the test added to AwNetworkConfigurationTest.

Bug: 960165
Test: run_webview_instrumentation_test_apk
Change-Id: I1ad88c0cb37f288c7a3b43711ccaa204c02e7590
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1617649
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Changwan Ryu <changwan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664011}