Mark Device Orientation API event handlers as [SecureContext].
This CL concludes the work to restrict the API to secure browsing contexts.
It removes the `RestrictDeviceSensorEventsToSecureContexts` flag that could
previously be disabled manually to re-enable the API in non-secure contexts,
tests that verified the behavior with this flag disabled, as well as all
deprecation warnings.

Bug: 932078
Change-Id: I742ab162e97017e47b44dfe72e99ef9c24178b36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1632329
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664180}