Navigation: Initial history subframe navigations use BeginNavigation.
When navigating back to a frame that has subframes, we should try to
restore the subframes to their latest history item. This is currently
done by OpenURL and should be done by BeginNavigation instead.

Design doc:
https://docs.google.com/document/d/13Rqdg1HBmtfnYMUXvg7SpUl3bJisTElfyMc1Ze8hLdE/edit#

This is patch 1/2.
This patch only modifies the behavior for
IsPerNavigationMojoInterface off.

Bug: 962518
Change-Id: I40c38663f964d3b4e0b09dcbab887d91a2faf282
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1578740
Commit-Queue: Arthur Hemery <ahemery@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664216}