[Autofill Assistant] Allow talkback access to the website again.
Before this change, both the bottom sheet and the overlay would call
mActivity.addViewObscuringAllTabs(), which means that talkback always
thought something was obscuring the browser and didn't give access to
the browser.

With this change, responsibility for deciding whether the browser page
is accessible to talkback is moved to the bottom sheet coordinator and
controlled based on the state of the controller.

This allows access to the website using talkback whenever interactive
access to the website is expected.

Bug: b/133436371
Change-Id: I3e49d26985c44f534eb5d11a85b6a2c692e60424
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1634886
Commit-Queue: Stephane Zermatten <szermatt@chromium.org>
Reviewed-by: Jordan Demeulenaere <jdemeulenaere@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664304}