overview: Avoid ending overview because of window activation change while an overview item is being dragged
The present CL aims to address a problem similar to [1]. [2] addresses
[1] by modifying OverviewSession:OnWindowActivating() to avoid ending
overview while there exists a window in a dragged state. That solution
is fine for workflows where a window is dragged from the top in tablet
mode, but there is a similar problem (possibly introduced by [3], which
I have reverted but would like to reland) involving overview item
dragging, which is handled by OverviewWindowDragController and does not
put any window in a dragged state.

[1] https://crbug.com/955814
[2] https://chromium-review.googlesource.com/c/chromium/src/+/1626542
[3] https://chromium-review.googlesource.com/c/chromium/src/+/1627562

Bug: 961170, 967196
Change-Id: I059127b88b343ea56a995bd7b815d5f434877064
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1633109
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664370}