Move the supported BufferFormats field from GpuInfo to GpuFeatureInfo
crrev.com/c/1612041 landed the enumeration of supported BufferFormats
on Ozone and wired it to a GpuInfo field; however, for the purpose of
using this information in gpu::gles2::FeatureInfo::FeatureFlags, it's
better to store it in GpuFeatureInfo. This CL moves it to there.

Bug: 967905, 911754
Change-Id: I1279e36af7b2fb2d680b6916f364db71d99e26f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1632597
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664391}