Bluetooth Scanning API: Implement permission storage on desktop
This CL implements the permission storage for Bluetooth Scanning API on
desktop:

1. Implements permission storage suggested on the design docs listed below.
The default permission setting for a site is "Ask". Permission storage
is only needed for the current active tab by adding a |allowed_scan_filters_|
to the WebBluetoothServiceImpl class.

2. If user blocks a site to use the Bluetooth scanning API, this permission
preference is stored in the content settings and can be used in other tabs.

3. Adds two functions to ContentBrowserClient to store and retrieve permission
setting from content settings.

Permission in tab switching will be handled in a follow-up CL.

Design docs:
https://docs.google.com/document/d/18Rqg1HRGeDngPv60DAGoMnIwRB-Zjh5gOWmANmH8lTo/edit
https://docs.google.com/document/d/1k4WM56Dx7sCu2k2MjvOdXArNz2Dy8BA1vt2VYHA6v7A/edit

Bug: 953075
Change-Id: Ie3339981584f219baa43a42d7522acce79a62197
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1610221
Commit-Queue: Jun Cai <juncai@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664399}