Fix bug where tab hover card image is seen red in color.
While a tab hover card is animating between tabs the hover cards anchor view is in transition and the tab hover card's data is updated before the anchor is moved. Because of this, GetThemeProvider() was not using the correct tab, this cl gets the theme provider from the correct tab that the hover card is moving to.

Bug: 968055, 910739, 928954
Change-Id: Ide4f109ea66c206da1c6fb9c8c785282694ccadb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1635857
Commit-Queue: Caroline Rising <corising@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664479}