Create TabGroupId and use it instead of int to identify tab groups
The new TabGroupId is backed by a base::Token so it can be meaningful
across different sessions. This will simplify the sessions code.

A new class is used for the ID instead of just a base::Token to avoid
mixing it up with other base::Tokens and to avoid similar mechanical
changes later on.

Bug: 930991
Change-Id: Ib0869730bff26fe182a4cb9398f1bb643547b1b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1629043
Commit-Queue: Collin Baker <collinbaker@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Bret Sepulveda <bsep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664512}