Avoid direct TelephonyManager method calls on the UI thread.
WebContentsAndroid has a getTelephonyNetworkCountryIso() call in its
constructor. As far as WebContents is usually created on the UI thread,
there is a synchronous call which goes through the binder there.

In this CL TelephonyManager subscribes to the TelephonyManager updates
and stores obtained values.

Former WCA construction time was about 8ms for the first time and
about 4.5ms for the following times.
In this CL WCA construction time is about 13ms for the first time
and about 0.09ms for the following times.

Bug: None
Change-Id: I9933a1c16931b12ae8489c7e9a5e319ca655dc8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1617499
Commit-Queue: Alexander Yashkin <a-v-y@yandex-team.ru>
Reviewed-by: Paul Jensen <pauljensen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664687}