Fix an issue that OnDriverReady is not called in reduced mode.
This is missed from the earlier patch.
In reduced mode, only InProgressDownloadManager will be created.
As a result, we should call OnDriverReady() when it is initialized.
This doesn't affect existing code when reduced mode is not enabled
for DownloadService though. As we will always create DownloadManagerImpl
when creating the download service.

BUG=942770

Change-Id: Ic6b54ccc47b8764796bf54eb086099d1c638cc6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1636814
Commit-Queue: Min Qin <qinmin@chromium.org>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Auto-Submit: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664847}