Add SimpleFactoryKey in //content embedders with their own BrowserContext.
Some parts of BrowserContext (such as VideoDecodePerfHistory and
DownloadDB) depend on //components. The components they depend will going
forward need to depend on a component service that is build with a
SimpleKeyedServiceFactory, which requires a SimpleFactoryKey. For most
use-cases, this key can be retrieved from Profile in //chrome, and used
in all of //components, since the implementation of the key lives in
//components/keyed_service/core. However, that is not the case for code
in //content or other embedders of //content such as ChromeCast, headless
and Fuchsia.

This CL therefore updates the subclasses of BrowserContext in these
embedders to provide and own a SimpleFactoryKey:

CastBrowserContext
HeadlessBrowserContextImpl
WebEngineBrowserContext

TBR=blundell@chromium.org

Bug: 943170
Change-Id: I021b817c576c789b9ef059e882c6f8bfe1bd1b3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1616422
Commit-Queue: Salvador Guerrero <salg@google.com>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Sean Topping <seantopping@chromium.org>
Reviewed-by: Xi Han <hanxi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664889}