Create BatteryMetrics in BrowserProcessImpl after ServiceManagerContext
The ServiceManagerContext is set up after starting the browser threads,
and the BatteryMetrics uses it on creation, so it should not be created
before. This was working because ContentMain() creates a
ServiceManagerContext and passes it to BrowserMain() in the
StartupDataImpl. However Android browser tests do not do so (and doing
so induces flakiness in the NestedMessagePumpAndroid), so this exposed
a crash in these tests.

R=jam@chromium.org

Bug: 961849
Change-Id: I54b535b040ad3c10985b087aed0410ff0c731196
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1637682
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665146}