Use testdriver Action API in WPT pointerevent tests - Part 7
In order to run the web-platform-tests automatically, we will use
testdriver Action API in all the wpt to simulate inputs. Here we are
changing some pointerevent input tests which have touch scroll actions.

This is the seventh CL that changes part of the pointerevent input
tests.

Bug: 606367
Change-Id: If83fe7b4490e88fa113b06bdcc9e07af7f734149
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1622456
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665160}