Added user-bounce metric to Android consent dialog
A user-bounce metric is when the user clicks deny button on the consent
dialog or otherwise closes the dialog by clicking outside of it, only to
retry and click allow button this time. The metric is to count the
number of times this happens.

This cl has a no-op refactor. The ConsentFlowMetrics(Helper)
classes in Java/C++ respectively are basically updating the same
histogram. Metrics folks strongly discourage this. They would rather
have one function update a metric and have all other functions call
it. So, this cl replaces Java histogram calls with Jni ones.

Bug: 965538,962755
Change-Id: Ie7e526caa7d18310b2c9580411d41df6ae618b1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1638098
Reviewed-by: Bill Orr <billorr@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Commit-Queue: Suman Kancherla <sumankancherla@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665183}