Compile in Preference support lib
The Android Framework Preference classes are deprecated as of API level
Q in favor of the v7 Preference Support Library. Add the required
dependency so we can begin migrating.

The net effect of this migration is expected to be a reduction in
method count and binary size.

Bug: 966073
Binary-Size: New support library import
Change-Id: I6b1f2f762f131567db31a6d9c314c71142866cfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1605515
Commit-Queue: Natalie Chouinard <chouinard@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665184}