Fix mailto links on ChromeOS with Network Service.
When you navigate to a url with an unsupported scheme,
NavigationURLLoaderImpl will attempt to open the URL with an external
program. However, if you have a protocol handler associated with that
scheme, it will handle the request in addition to the external program
(see ProtocolHandlerThrottle).

On ChromeOS, the external program ends up being a new tab, which then
repeats this process and in certain situations can result in an infinite
loop of tabs being opened. This CL prevents the browser from opening an
external program if a protocol handler exists for the URL's scheme.

A browser test will be added in a followup CL.

Bug: 963133
Change-Id: I6346b92adb3ea508a037b65a49dd1b44dd7d7b8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1639077
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665249}