[Previews] Adding remote Optimization Guide URL HTTPS check.
This change also inserts the default URL for the remote Optimization
Guide URL if not provided by Finch or command line flag.

This also removes the hard CHECK on the provided URL for the remote
Optimization Service URL and defaults to the production server URL.

A unit test is implemented to confirm that GetOptimizationGuideServiceURL
will default to the production URL if an HTTP URL is accidentally set.


Bug: 932707
Change-Id: I17e661679caa762a318c89c9a6cfd3f10c540f40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1635078
Commit-Queue: Michael Crouse <mcrouse@chromium.org>
Auto-Submit: Michael Crouse <mcrouse@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Doug Arnett <dougarnett@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665378}