[Bluetooth WinRT] Load Ole32.dll for RoGetAgileReference
The RoGetAgileReference() function is only used when initializing the
Bluetooth library on Windows 10.

As Ole32.dll is linked as a delay-load library we can't assume that it
will be found by ::GetModuleHandle(). Instead this change explicitly
loads it with ::LoadModule() and releases it with ::FreeLibrary() after
it is no longer needed.

Bug: 943494
Change-Id: Ieec3f8a46e5b459048b68b70836dae429a1c75de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1638940
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665382}