[libusb] Remove dynamic library loading on Windows
Chrome already links against the libraries that libusb is loading
dynamically so this code can just be removed. This fixes both the fact
that libusb leaks the library handle it creates and also does not
properly retain references to the handles it uses so they could be freed
by other application components.

Bug: 943494
Change-Id: I96e3d14ea642d90074f277446ac4336d69ae2105
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1536696
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665391}