Android: Foreground tab check with |Tab.isHidden|
In many places, |FullscreenManager.from(tab) == null| is used to see
if a given tab is in the foreground or not. This CL replaces it with
|Tab.isHidden()|. They are equivalent in all situations - tab in
fg/bg, tab switcher, app in bg. This helps us step closer to deleting
the method |FullscreenManager.from(tab)| entirely.

Bug: 966272
Change-Id: I737fb59f94e27fc4e33838291e2263098821d6ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1636994
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665414}