Remove kChromeUISyncResourcesHost and cleanup Context Menu code.
This CL cleans up kChromeUISyncResourcesHost which isn't used. In doing
so, it removes a context menu check that makes the normal page options
not appear on chrome://syncresources, as well as a default implementation
that looks like it was intended to remove these options on all WebUIs.

Bug: None
Change-Id: Id7a8bce3ca68759d143178ebe50615d00dcaec13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1634593
Commit-Queue: calamity <calamity@chromium.org>
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665480}