Fix header view visibility.
When switching to inline settings, we want to hide all views except the
origin, not just the timestamp. This groups all views that need to be
hidden together so it's easier to hide and show them again.

The previous CL http://crrev.com/c/1565001 did not handle the summary
text view that is shown for list and progress style notifications.

Bug: 955410
Change-Id: I3190cf112ee926b6e8ca1edc0953e1a3cd494273
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1615467
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Richard Knoll <knollr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665482}