[Files app] Deflake addRemoveFolderShortcuts test
Add retry for "right-click" to show the context menu for "remove folder
shortcut" menu. The right-click is on directory-tree (LHS), there is an
event listener p1[ for file selection (RHS), which races with this
right-click and makes this flaky.

I've added a local log for a retry and I could see it failing in the
first run and succeeding in the second run, this is a stopgap fix,
there is still a chance to race between the retry and the click after
it, but the window is much smaller. As the TODO in [1], I'll try to
remove the event listener to eliminate this race.

[1] - https://cs.chromium.org/chromium/src/ui/file_manager/file_manager/foreground/js/file_manager_commands.js?l=362-366&rcl=051250e22e8533ec28b9baa56ba945e9deda3166

Bug: 936429
Change-Id: Ibcb1cbfec9d22cb451d74f481bf4af4255e0f752
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1640817
Auto-Submit: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665485}