[Autofill Assistant] Report the intent data url.
Before this patch we reported the committed URL to our backend in the
initial request. The committed URL can potentially contain PII and can
therefore not be kept.

This patch changes that and reports the data url from the intent as
passed to chrome by a caller. There is approval for that piece of
information to be logged.

Bug: b/133732376
Change-Id: Idb98a727d37f2b9e7488f2f84b8fd273c2f4240f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1632252
Commit-Queue: Mathias Carlen <mcarlen@chromium.org>
Reviewed-by: Stephane Zermatten <szermatt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665489}