[Autofill Assistant] Fix scrolling for elements within scrollable containers.
Before this change, when scrolling elements within scrollable container,
the element would be first taken to the desired position within the
screen, *then* scrolled within its container, which could result in the
element not even appearing within the window.

With this change, the element is scrolled into view within its container
before choosing how much to scroll the window vertically. At the end of
the scroll, the top of the element is always at the desired position
within its frame.

This change also switches the browsertest for scrolling to doing
synchronous checks, as we don't use smooth scrolling anymore.

Change-Id: I5bd090988ec8ca5441d15783d2b46ceb3982e57d
Bug: b/129046627
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1636138
Commit-Queue: Stephane Zermatten <szermatt@chromium.org>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665523}