[Switch Access] Create settings page
This commit creates a settings page for Switch Access behind the
Switch Access experimental flag. The page is not accessible to users at
this point except by typing in the direct URL
(chrome://settings/manageAccessibility/switchAccess) with the flag
enabled. The preferences set here are not yet connected to the behavior
of Switch Access.

Bug: 958081
Change-Id: I86951e2877c18c047053a4ec3fcb3b2f18dd46b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1600082
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Dan Beam <dbeam@chromium.org>
Reviewed-by: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665705}