Remove callback for injected scroll gestures
With the additional logic to not inject scroll gestures when the
computed delta would not end up scrolling, we no longer receive
calls to our callback that contain overscroll data. We can remove the
dummy callback we were passing in to suppress the DCHECK in
QueuedWebInputEvent::HandledEvent.

We can also set the event as non-blocking as well since the only reason
it was set as blocking was to ensure our callback was present in the
event that overscroll data was propagated.

I looked into suppressing this further within Blink for scroll gestures
with scrollbar device, but the location at which the overscroll
notification is created only has access to ScrollData, which has no
device information.

The test at
fast/scrolling/scrollbars/scrollbar-button-gesture-target.html exercises
the codepaths that used to cause the DCHECK to be hit.

R=bokan@chromium.org, tdresser@chromium.org

Bug: 954007
Change-Id: Ida793c2b590704aa26f25734ef131160b017fd03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1636859
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Timothy Dresser <tdresser@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#665769}