gpu/exo: Wait for Exo acquire fences on the GPU
Exo exposes zwp_linux_explicit_synchronization wayland protocol that
allows clients to request explicit synchronization for buffers on a
per-surface basis.

Exo protocol implementation currently waits for the fence on the CPU
before sending the surface content to the display compositor.


This CL propagates the in fence from Exo to the SharedImage
implementation on the GPU service side and inserts a server wait on the
actual GPU driver.


Bug: 865708
Test: wayland_explicit_synchronization_client --use-drm=i915 on atlas
Change-Id: I90f6eee8a959159a2342400918e3f75a8c38b783
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1611102
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Reviewed-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>
Cr-Commit-Position: refs/heads/master@{#665795}