Ensures fonts default settings for unittests
This CL is adding a test that will fail when the system doesn't have the
default UI settings. Chrome unittests often rely on these default values.

Previously, the EXPECT were running into:
  static void SetUpTestSuite() { ... }
which does not make the test suite to fail on any error.

We are proposing to add a separate test only for testing the computer
states.

R=tapted@chromium.org

Bug: 963868
Change-Id: I08fc33e0fb35d278562e4bcbb977e5376f3d415e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1637670
Reviewed-by: Trent Apted <tapted@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665924}