Reland https://chromium-review.googlesource.com/c/chromium/src/+/1635859.
Original CL description:

Implement AppLaunchManager for controlling app-under-test for EG2 tests.

No tests are added to smoke_egtest to test app relaunching, as it descends
from ChromeTestCase, which currently has logic which is required to run
every time the app-under-test is launched. Long-term, it is preferred if
this logic is moved to -setUp and executing before every test method. This
refactor will happen in a later CL as it requires some planning.

Confirmed that smoke_egtest does however only launch the app once with the
change to use AppLaunchManager.

Bug: 922813
Change-Id: I02d7e65cb47d86c35c09d049bc628c0220b4adb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1641468
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: ericale <ericale@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666024}