Ensure throttles don't access destroyed manager
This is a fix for a crash in
~DataReductionProxyURLLoaderThrottle() that is caused by a
use-after-free of DataReductionProxyThrottleManager.  The lifetime of
the manager vs. the throttle is unfortunately not easy to determine
precisely, with these types of objects created in distinct places and
the manager sometimes owned by ref-counted objects.  To work around
this, the manager will now inform the associated throttles about its
destruction so they can drop references and behave accordingly.

Bug: 968513
Change-Id: I301555716b8bdf905d699662ef66f0524d0338ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1638521
Commit-Queue: rajendrant <rajendrant@chromium.org>
Reviewed-by: rajendrant <rajendrant@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666064}