Restrict Crostini on experimental kernels
As part of the kernel uprev process devices may receive kernel support
for VMs. In these situations it would be nice to allow the user to
enable Crostini. However, while the kernel is in its experimental
rollout phase for a given device there is no guarantee the kernel won't
be reverted to a version without VM support. As a compromise, let's
still allow users to enable Crostini when supported, but guard it behind
an experiment flag so users are forced to acknowledge the possibility of
data loss.

BUG=b:133833983
TEST=start chrome with --kernelnext-restrict-vms, verify crostini is not
available. Enable enable-experimental-kernel-vm-support flag and verify
crostini is available.

Change-Id: If5727122e5d76f1a674ae96964d5eb6587eb5bf1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1639726
Commit-Queue: Jon Flatley <jflat@chromium.org>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666075}