Android: Animation effect in Ephemeral Tab favicon
This CL drives Favicon visibility with fade-in/out animator. Also
keeps the size of favicons consistent.

|EphemeralTabLayer::GetLocalFaviconImageForURL| is replaced with the
notification via |FaviconDriverObserver|. |GetLocalFaviconImageForURL|
provides favicons that are already available in local history db
only. Relying soly on it caused a bug not updating favicon for
newly visited pages. FaviconDriverObserver addressed the problem.

If navigated to a page from a domain different from the current one,

1) reset to a default favicon
2) show the new one as soon as we're notified of the new favicon.

Note that these operations do not happen if navigated to a page within
a same domain. This is checked by the host name of URL being visited.

Bug: 929727
Change-Id: I716f9c03b31c513a7eb00dd286018a7a200cbde3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1605381
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Donn Denman <donnd@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666123}