[FCP++] Image: int-multiply overflow
ImagePaintTimingDetector has an overflow issue in multiplying two
integers (height*width). In this patch we will use
rect.Size().Area() to compute the size. As the function has handled
the overflown issue properly, we can use it to fix the bug.

The fix has been verified with a local run of the Cluster Fuzz.

Bug: 969295
Change-Id: I67fda6339bbcbc8346b02397031d9a73c3c61f4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1643091
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666280}