Disable leaking content_browsertests on LSAN.
See crrev.com/c/1590404 and the linked bug for context -- the goal is to
fix the broken memory-leak checking in Chromium's ASAN
(AddressSanitizer) trybots, which are supposed to also run LSAN
(LeakSanitizer).

Currently-leaking tests must be fixed or disabled before leak checking
can be enabled in the ASAN trybot.

WARNING: This will result in a loss of ASAN coverage for the affected
tests given that LSAN runs as part of the ASAN trybot.

If that's not acceptable, we can wait until the leaks are fixed before
fixing leak detection in trybots, but keep in mind that newly-introduced
leaks Chromium-wide won't be caught until then.

Bug: 961029
Change-Id: I189cfe1a129ed7af4f87d256e3ecf1d9d84a392a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1638920
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Katie Dektar <katie@chromium.org>
Reviewed-by: Martin Robinson <mrobinson@igalia.com>
Commit-Queue: Caleb Raitto <caraitto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666310}