[iOS] Inset WebState view in BVC's |-displayWebState:|.
crrev.com/c/1529921 resizes the WebViewContentView to match the
CRWWebControllerContainerView's frame when it's added to a new window.
However, BVC resizes the container view to match the content view
bounds in |-displayWebState:|.  The frame was later reset by the
fullscreen resizer when BVC.broadcasting is reset to YES when the
animation finishes.

This CL updates |-displayWebState:| to account for the current
viewport insets so that the frame remains constant during the settings
dismissal animation.

Bug: 963108
Change-Id: Ifd44277f7c88bfea60e292387468483a521e5f2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1644419
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666326}