Fire UIA/MSAA menu start/end events when menu is shown/hidden
Menu events should fire when a menu is made visible or hidden, whether
or not the menu or something within it gets
focused, because some authors simply show the menu at first, and
only move focus if the user arrows into it. If the menupopupstart
doesn't fire when the menu first appears, then the screen reader
does not know to change to focus mode (virtual cursor off), and
the web page will not receive down arrow keys.

This also matches the behavior of Firefox for menupopupstart.

See the two bugs listed. One contains a general simplified test case
and the other describes how to reproduce the issue with Gmail.

Bug: 98981, 964546
Change-Id: I975513ba73f3b23e884a7f1b38e6b5b161028d17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1637057
Reviewed-by: Ian Prest <iapres@microsoft.com>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666478}