Convert //components/tracing over to use Mojo IPC.
This meant exposing a service BackgroundTracingAgent from child processes
that the browser processs can use to configure the reporting of background
tracing back to the browser via BackgroundTracingAgentClient.

As part of this change, background tracing is disabled for pepper plugins
(nacl and flash) as these are deprecated technologies, and we are not
deriving much benefit from background tracing of these processes.

Change-Id: Id940dbf4646c315c01f1ca311b66b721a6648706
Bug: 967834
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1621067
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: oysteine <oysteine@chromium.org>
Commit-Queue: Darin Fisher <darin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666509}