DevTools: Re-enable shadow-distribution.js test
This test was disabled when the IncrementalShadowDOM flag was turned on,
since it relied on the timing of the UpdateDistribution phase of ShadowDOM,
when dumping out the elements tree.

This change proposes that we force a style recalc in the test, to ensure
that the appropriate DOM domain events are sent and processed by the front end.
(e.g. ChildNodeCountUpdated, DistributedNodesChanged, AttrModified, etc).

Bug: 840238
Change-Id: I90207b8109c951e7c7b826cbce1b8f9e28143351
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1638851
Commit-Queue: Mike Jackson <mjackson@microsoft.com>
Reviewed-by: Joel Einbinder <einbinder@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666553}