[webnfc] Introduce NDEFCompatibility to replace NFCWatchMode
This CL switches the existing enum NFCWatchMode to NDEFCompatibility
which is newly defined at
http://w3c.github.io/web-nfc/#dom-ndefcompatibility
and updated the behaviors implemented in Device Service accordingly.

Now we just use it for watch operations, in the future we'll be using it
for push operations as well.

BUG=520391

Change-Id: I82cf736d1a57044e9fd59a350b748b99548f9a82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1628537
Commit-Queue: Leon Han <leon.han@intel.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666621}