[DL]: Ensure we can use debug printing with display locking.
Without this patch, we run into some DCHECKs when trying to dump the
layout trees as text or to dump paint records, since that code is not
aware of display locking.

This patch addresses that by adding display locking checks in spots
that recurse down into children of elements. Also adds a
"(display-locked)"
string to elements that are locked.

R=chrishtr@chromium.org

Change-Id: Id0f58a33c0d1df3374d9d496ea626efadc8a2f8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1645815
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666689}