Update UKM network annotation after Unity changes
The unified consent changes the settings required to enable UKM collection.
Unity adds a setting 'Makes searching and browsing better' that directly
controls whether UKM is enabled instead of relying on the sync state.

The UKM code has two different ways of determining the feature state
depending on whether Unity is enabled or not. Unity is going to be enabled
by default in M77 on all platforms but ChromeOS that still uses the old
model.

Bug: 951781
Change-Id: I0caa9503e4e9446f35b4d59940b19c1eb26a1a6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1642552
Reviewed-by: Robert Kaplow (slow) <rkaplow@chromium.org>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666790}