Ensure OpenVR always sends Axes data
While testing with OpenVR it was discovered that the WMR joystick data
was no longer being plumbed up, because the corresponding button was not
reported as supported (OpenVR reserves the joystick click on WMR). With
the button not supported, the GetAxesButtons helper was also ignoring
the (valid) axis data.

While WMR through OpenVR is not an expected/supported scenario, OpenVR
could do this to other headsets/controllers which are only exposable via
OpenVR.

Since the spec reserves a button spot for the secondary axes (and
expects a button with each other joystick), dummy button data is plumbed
up with the axis data.

Bug: 971478
Change-Id: I0269d944aefd325f2cce46939754bc547cc78c38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1646761
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Bill Orr <billorr@chromium.org>
Reviewed-by: Bill Orr <billorr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666791}