Reland "Implement the full Supervision Onboarding flow."
This is a reland of 53c05c25123627b91296a0be5298c96c732f79c3

We are removing the browser tests that navigate deeply in the supervision
onboarding flow. These tests are timing out not because the logic is wrong, but
because the navigation and setup for the OOBE webui is taking too long.
I verified this by running the tests with slightly higher timeouts (+10s) and
all of them consistently passed in my machine after 20 runs.

The idea is to cover these navigation scenarios in unit tests that are more
reliable, leaving only the start page being covered by browser tests.

Original change's description:
> Implement the full Supervision Onboarding flow.
>
> This change adds the "Details" and "All Set" pages to the flow, as well
> as a class to manage the (now multiple) onboarding states.
>
> Another change is that we stop removing some webview listeners when we
> stop loading the page. This was causing the webview to never finish
> loading subsequent pages, getting stuck in the first blocking listener.
>
> Bug: 958995
> Change-Id: Ib28f8d271263ffa832e05a5c6cefb7f7675b02a2
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1630223
> Commit-Queue: Lucas Tenório <ltenorio@chromium.org>
> Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
> Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
> Reviewed-by: Oliver Chang <ochang@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#666535}

Bug: 958995
Change-Id: I814392870714b173540e480125d833b199c32118
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1647196
Commit-Queue: Lucas Tenório <ltenorio@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666826}