Remove DRP code not used with network service
Network service is launched on all platforms that support DRP, and is
enabled by default in the code. It is now safe to remove code that is
not used on the network service path.

This change mainly deletes files that aren't used, and removes
references to those classes in other files. It also removes all
references to net/url_request objects in DRP code. A couple
non-mechanical changes to look at:
- DetermineCommittedServerPreviewsState moved from ContentLoFiDecider
  (now deleted) to previews_content_util.cc
- data_reduction_proxy_bypass_protocol_unittest.cc converted to not
  use URLRequest. A couple tests removed that should already be covered
  by browser tests.
- DataReductionProxyIODataTest.TestResetBadProxyListOnDisableDataSaver
  converted to use network service objects.

Bug: 934009
Change-Id: Ifff7d49d20ccfdd50c28e42c2250680f859075e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1644379
Commit-Queue: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666832}