Move DCHECK for existing ThreadPool in ScopedTaskEnvironment
If ScopedTaskEnvironment is created with MAIN_THREAD_ONLY we are not
creating a ThreadPool, so no need to DCHECK if one already exists. It
could be that the tests if explicilty setting MAIN_THREAD_ONLY becasue
it already created one.

Change-Id: Ib0514a240d8d06d2c50b03fef4da87f583113a31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1648335
Commit-Queue: Gabriel Charette <gab@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Auto-Submit: Carlos Caballero <carlscab@google.com>
Cr-Commit-Position: refs/heads/master@{#666858}