Enable site isolation on all sub-domains of a single origin.
1. Make GetMatchingIsolatedOrigin() respect wildcard isolated origins.

The matching behaviour for GetMatchingIsolatedOrigin() is changed such
that when iterating the list of isolated origins per site, if a
wildcard origin is encountered, the function returns immediately. This
prevents a non-wildcard isolated origin elsewhere in the set being
matched.

2. Add isolating wildcard domains to ChildProcessSecurityPolicy.

Provide an additional implementation of AddIsolatedOrigins which
accepts a list of IsolatedOriginPatterns, correctly preseving whether
or not they are wildcard isolated origins when converted to
IsolatedOriginEntries.

3. Add IsolatedOriginPattern, a container class for wildcard patterns.

In order to add support for isolating all sub-domains of an origin,
expressed by a double asterisk (e.g. https://**.foo.com) a new
container class is required. We are unable to use GURL or url::Origin
to hold these patterns as they escape the '*' character to '%2A', which
would make parsing brittle if this behaviour ever changes in the
future.

Instead we propose to hold user-supplied data in an
IsolatedOriginPattern instance, which correctly identifies and removes
wildcards. After removal of the wildcards conversion to url::Origin()
is possible, requiring only minimal changes to code that calls
ChildProcesssecurityPolicyImpl::AddIsolatedOrigins().

4. Add double-wildcard support to IsolatedOriginEntries.

Double wildcard origins are origins for which any sub-domain should be
subject to site isolation. By default the varaiable is initialised to
false; future commits will permit callers of
ChildProcessSecurityPolicy::AddIsolatedOrigins() to specify URLs of
the form https://**.foo.com to isolate any sub-domain that matches the
specified pattern.

Design doc for these chanes is available here:
http://doc/1nAL_ZH64weVOaTSpfA85MOdNW1RzTAZg2A3i1oixAuE

Bug=chromium:926997

Change-Id: I5259cf1e7fd69ad933ce741debb979c155d0a9bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1589080
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Andrew Stone <ajws@google.com>
Cr-Commit-Position: refs/heads/master@{#667030}