[IOS][Translate] Allows Compact Translate Infobar creation in error state
This CL changes the Compact Translate Infobar implementation so that the
infobar can be created in the error state. The infobar presents a Snackbar
notification in this state.

This is needed when the Translate infobar is manually force-triggered in a
situation where translation cannot be done (English to English translation)

Bug: 971832
Change-Id: I9d2a58da42a783a23a8a0c958823bd27293fec77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1648709
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667130}