GCC: FeaturePolicyParser ParseValueForFuzzer is not in anonymous namespace
Compilation fails because we are declaring ParseValueForFuzzer as friend method,
but we are declaring it is in anonymous namespace. Moving to global namespace
still fails (in this case in Clang).

So final solution is making it a public static method of FeaturePolicyParser.

Bug: 819294
Change-Id: Iea307cb6faef675b748d6eb5da2175dcbb17fdc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1645223
Commit-Queue: José Dapena Paz <jose.dapena@lge.com>
Reviewed-by: Luna Lu <loonybear@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667228}