Increase VR setting service delay
Increases the delay after sending an intent to the VR settings service
from 500 milliseconds to 3 seconds. Apparently it's possible for the
service to take multiple seconds to actually apply the requested
settings, in which case the test would run without the proper
settings actually being applied. Most commonly, this resulted in the
DON flow being skipped when we expected it to be present.

TBR=billorr@chromium.org

Bug: 972153
Change-Id: Iba9265c67cac5e5aee55a2dd85b2d2b0ab1170ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1650087
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667352}