Block-start of a float must be >= that of previous floats.
Final layout would eventually end up being correct anyway, but during
layout we need to calculate the inline-size of the float, before we can
tell whether it's going to fit next to the current line or not. This
piece of code forgot to check with the block-start of preceding floats,
so that it would temporarily be put at the wrong offset, triggering a
DCHECK failure during block fragmentation.

Move some code out of LayoutBlockFlow::PlaceNewFloats() into a new
utility method.

Note that the test included also passes without this fix (except for the
DCHECK failure).

Bug: 961516
Change-Id: I3bbd37c7affcb5fa7143e951f9f1bba78a2a4a0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1649477
Reviewed-by: Emil A Eklund <eae@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667467}