[Previews] Remove expired fetched hints from the store.
This implements the functionality for the HintCacheStore
to remove fetched hints that have expired from the store.

The next change will purge the expired hints after a batch of fetched hints
are stored by the PreviewsOptimizationGuide. Another change will record the
number of fetched hints that are expired from the store.


Bug: 932707
Change-Id: I553128a238f61c2e443094fabb7512aeecb305c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1647190
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Doug Arnett <dougarnett@chromium.org>
Commit-Queue: Michael Crouse <mcrouse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667599}