Fixes lost pinned tabs when launching links
When the browser closes or launches, pinned tabs session gets
saved in preferences. The current Chromium behavior is that it
doesn't show pinned tabs when it launches a URL externally via
App or CLI. Since this isn't a transient session, showing pinned
tabs when launching a URL is probably the right thing to do.

Bug: 812545
Change-Id: Ic50f048e99aba87fb302501a5ab36da6ca3559ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1648918
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667648}