app_list: No animation when take screenshot
When taking a screenshot from the embedded Assistant, it will take a
screenshot of Launcher instead of the content before opening Assistant
because the Launcher dismiss animation. This patch turns off the
animation so that Launcher will close immediately. In table mode, it
will set to the final state immediately. I.e. if there was a window on
top of launcher, it will take a screenshot of that window instead of
launcher. If user asks to take a screenshot in home launcher, since
the Launcher cannot be closed, it will take a screenshot of the home
launcher.

Bug: b/133248292
Test: manual
Change-Id: Ibe92cdb288aff7ef1539e4943f4e0b64a7f43e22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1649308
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Tao Wu <wutao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667748}