device/fido: cancel authenticators in ~FidoRequestHandlerBase
FidoRequestHandler is scheduled for removal at some point. Because of
that, some request handlers are derived directly from
FidoRequestHandlerBase. However, such request handlers were not
canceling requests on destruction because that was still in the higher-
level class.

Move that logic down. Also make it unconditional because even a
completed request handler should still cancel any authenticators that
are still pending. (Although there probably shouldn't be any.)

In order to keep the gMock tests happy we need to ensure that
superfluous |Cancel| calls aren't made. This is a nice-to-do anyway
because we previously had |FidoDeviceAuthenticator| keeping pointers
to operations and tasks after they were finished. Thus this change also
causes those pointers to be cleared up as soon as the callback is made.

Change-Id: I78a9ce1b4d0cdad8ebd03a720deb41a308b34247
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1646675
Commit-Queue: Adam Langley <agl@chromium.org>
Auto-Submit: Adam Langley <agl@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#667749}