Use long form of TEMP directory name
%TEMP% may be a short directory name which will be turned into
a long directory name when it is normalized. As we compare
path_a=%TEMP%\stuff with path_b=Normalize(%TEMP%\stuff) it is
necessary to ensure the prefixes of path_a and path_b are correct
to start with by finding the long path name once the containing
directory has been created.

      and manually tested fix, see linked bug for steps.

Test: reproduced bug reported by official.desktop/win-clang
Bug: 972558
Change-Id: I4d08dc65d5d2610daed75bd10a51296ba022e7be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1652013
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667771}