Ensure ARIA menu items emit ATK focus-changed events
We were treating selection and focus the same for menu items because in
browser UI we get selection-related events only, but need to emit ATK
focus-changed events. However, for ARIA menus we get active-descendant-
related events; not selection-related events. As a result, we were
returning early when the selection changed in ARIA menus such as those
found in Google Docs.

Bug: 959857
Change-Id: Ie5962a5c5ca246980b68fba591654e2126815d81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1652262
Commit-Queue: Joanmarie Diggs <jdiggs@igalia.com>
Reviewed-by: Martin Robinson <mrobinson@igalia.com>
Cr-Commit-Position: refs/heads/master@{#668011}