Split Cache: Network isolation key for subresources in the URLLoaderFactory
This CL adds a network isolation key as a member in URLLoaderFactoryParams
which is used as the key for the Http cache and is propagated to the cache
in every URLRequest.
This CL is scoped to create and fill the key for subresources
which is done in RenderFrameHostImpl when a navigation is being committed.
RenderProcessHost::CreateURLLoaderFactory() interface is changed to accept
the network isolation key as an argument.
Follow up CLs would fill it up for navigation requests and worker
subresources.
Added unit tests in url_loader_unittests.cc and for the new mojom traits.
The existing browser test WebContentsSplitCacheBrowserTestEnabled.SplitCache
tests this change and the helper function is modified to also test the http cache
key in the RenderFrameHostImpl.

Bug: 910708
Change-Id: I066d53df5c1ad1abe84ecce67697ddd2bb5ee11c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1610218
Commit-Queue: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Reviewed-by: anthonyvd <anthonyvd@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668057}