Pass --repository-absolute-path command line argument to Typ
The absolute path of the source directory will be passed as the value of
this command line argument. Typ will then use this value to derive relative
paths for files. One such path is the expectations file path. The relative
path of the expectations file will show in the full JSON results, which can
give a better idea of where the files are located in the chromium repository.

Bug: chromium:698902
Change-Id: I6374eee8e83b9a1d1b99c628be4368649e1b0d7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1652563
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Cr-Commit-Position: refs/heads/master@{#668167}