Add a ChromeOS-specific GN arg for generating Chrome orderfile.
This patch adds a ChromeOS-specific GN arg "dump_call_chain_clustering_order"
that is only intended to be used by Chrome OS ebuild files to specify a path
to a file used to save symbols ordered by Call-Chain Clustering (C3) by LLD.
The arg is by default to be empty so there's no impact on other platforms. And
when a path to a file is specified, it will turn on new pass manager in Clang,
and pass --print-symbol-order=<file> to dump the symbols to the specified file.
Again, only the ChromeOS toolchain team will turn on the arg to generate
orderfiles.

Bug: chromium:950627
Test: This patch is tested as a local patch in Chrome OS builder
https://ci.chromium.org/p/chromeos/builders/general/Try/b8911280322967308192

Change-Id: I87a8cceca918db1f2b89c7244c4e3dc6f48ad5ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1650618
Reviewed-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Tiancong Wang <tcwang@google.com>
Auto-Submit: Tiancong Wang <tcwang@google.com>
Cr-Commit-Position: refs/heads/master@{#668200}