[quickview] Make openQuickViewSniffedText test verify the mimeType
CL:1653610 added a tool to read <files-metadata-box> field values. Use
it to verify the displayed mimeType of the plaintext test file.

Minor: fix comment in the tool and the <div> containing the value is a
direct child of the #value element: use a child CSS selector to better
match the proceeding comment.

Test: QuickView/FilesAppBrowserTest*
Bug: 772600
Change-Id: Ifab947e535c2e87f897fc1d41dc6adfcab82ff38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1654668
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668262}