Remove WebPushSubscription and merge it into blink::PushSubscription
This struct is no longer used outside of Blink, so we can remove it from
the public API and merge it along with blink::PushSubscription, which
represents the same kind of information, but from a point of view that
is closer to the DOM and V8 than the struct being removed.

To do this, though, we needed to adjust a few places in different points
of the PushMessaging module (e.g. PushSubscription constructor, callback
definitions...), but this will also help removing other WebPush* types in
follow-up CLs, so it should pay off in the end.

Bug: 939943
Change-Id: Ibdf7eff146dc4a0be92dd7bc246841ea7bec5a98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1645780
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668362}