Add detection of high usage of HFONT in PlatformFontWin
This CL is adding code to detect when GDI Handles (HFONT)
are going to high because too many gfx::Font exists.

There is a plan to fix the incorrect uses of gfx::Font
and these dump will help tracking the source of these leaks.

Bug: 972689
Change-Id: I326938c3751766e01b3c318b5397ef8f4d4f9ae9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1648398
Reviewed-by: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668462}