Compositor threaded scrollbar scrolling: AutoScrollCurve implementation
This CL adds a new mode to the ScrollOffsetAnimationCurve that can be
used to create a cc auto-scroll animation when interacting with the
scrollbar. The curve is linearly increasing and its duration depends on
the auto scroll velocity. A new linear timing function has also been
added.

Bug: 969203
Change-Id: I648cb3ca289c0457699d221ebd89124a6e8853fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1639561
Commit-Queue: Rahul Arakeri <arakeri@microsoft.com>
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668563}