Add DiscardableSharedMemoryManager::Get()
There's always a single instance of this class in practice, shared by the
whole browser process. Previously it was exposed via a method on
BrowserMainLoop, but BrowserMainLoop has other usage constraints which make
it an inconvenient way to expose the DiscardableSharedMemoryManager in some
environments.

This adds a static Get() method to DiscardableSharedMemoryManager and removes
the discardable_shared_memory_manager() accessor from BrowserMainLoop. This
is a precursor to lifting the instance itself out of BrowserMainLoop and into
ContentMainRunnerImpl so that the the system Service Manager Connector (which
also exposes the DiscardableSharedMemoryMananager) can be brought up earlier
and without BrowserMainLoop.

Finally, this also deletes an unused and unnecessary public content API
to access the DiscardableSharedMemoryManager, given that any dependents
(in, e.g., other non-chromium content embedders, assuming some dependents
exist) can just call the static Get() now.

Bug: 904240, 968147
Change-Id: Ifb0f9936a2f852f9ce3ffe5447f778c7fad3b7d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1651496
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Peng Huang <penghuang@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668597}