Set restore bounds when ARC window enters PIP
Restore bounds is used to specify the ideal position for PIP, but
currently this is only set in completeDrag() for ARC PIP. So,
When a ARC window enters PIP, its restore bounds is not set until
the user drag the PIP window.

For Chrome PIP, the restore bounds for the initial position is set
in DefaultState#setBounds, so the similar thing should be done in
ClientControlledState as well for ARC PIP.

UI windows before the window is drag moved.

Bug: b/129534057
Test: Confirmed an ARC PIP window properly interacts with system
Change-Id: Ie85bd8e15c49160802f3173fa9a72b18ca84d246
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1655629
Commit-Queue: Kazuki Takise <takise@chromium.org>
Auto-Submit: Kazuki Takise <takise@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668714}