[Android] Reassign default text color to input chips
This CL restores the primary text color of Input Chips. It used to be:
Gray 900 for the primary label and
Gray 700 for the secondary label.

Recently, this must have changed (presumably when generalizing this code
for night mode) to using Gray 700 everywhere.
For Input chips, this regressed for the keyboard accessory since primary
and secondary label are impossible to distinguish.

So although we don't have any fixed default for these chips, I propose
the mentioned colors as first step so it is absolutely clear that they
need different text styles.

Bug: 973763
Change-Id: Ia0288974367d8862a9efd86e06d0d0b4be5d1f4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1655476
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Becky Zhou <huayinz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668764}