[Mac] Stop pretending that browser_tests/InProcessBrowserTest are bundled.
After 28bc65ff4fd69fabb622eb8dc2b8efb8704c8646, browser_tests no longer
launched child processes via the bundled helper applications. But the
browser process maintained the charade that it was running as the main
bundle executable, including by calling SetOverrideAmIBundled(true).

The AmIBundled flag will be used in the future to choose whether to
launch specialized helper apps with macOS code signing entitlements, or
to re-launch the existing binary (as in tests). As a result, it is no
longer possible to pretend to be a bundled app if the test executable
does not launch bundled child processes.

Bug: 850193
Change-Id: Idc46cfe032e888719c14badec97d708613059635
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1652665
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668785}