Eliminate PrimaryAccountPolicyManager inheritance
This CL eliminates PrimaryAccountPolicyManager inheritance from
PrimaryAccountManager. The latter will now hold an unique_ptr to the
policy manager and call Initialize() (renamed from
FinalizeInitBeforeLoadingRefreshTokens()) when it's not null.

Additionally, split PrimaryAccountPolicyManager into an abstract base
class and an implementation in PrimaryAccountPolicyManagerImpl, which is
not built on ChromeOS.

This is the last step to merge SigninManager and SigninManagerBase as
motivated in this design doc:

https://docs.google.com/document/d/15y-Db27BV08vrIyelHB-3CwiAfDYh-FigNKGXrqSWto/edit#heading=h.8jjdy95t6p7x

and described concretely here:

https://docs.google.com/document/d/15y-Db27BV08vrIyelHB-3CwiAfDYh-FigNKGXrqSWto/edit#heading=h.mbkrv9nkb93w

Bug: 952766
Change-Id: I47f53a6dd4bec810b6ebb78624228f082d2bd693
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1645225
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: John Wu <jzw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668788}