Remove BrowserMainParts::ServiceManagerConnectionStarted
This is no longer necessary since the Service Manager is initialized
before browser threads are even created. None of the implementations
of this override needed the ServiceManagerConnection specifically,
but only its Connector. Additionally, with recent changes to Service
Manager initialization the PostCreateThreads() method is effectively
equivalent in timing to what ServiceManagerConnectionStarted was.

As such all prior implementations of ServiceManagerConnectionStarted
have been moved into a PostCreateThreads override, and references to
the connection argument have been replaced with GetSystemConnector().

Bug: 904240
Change-Id: I70a445f5dcf9d16df654b95d971672f8c611e108
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1652798
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669025}