vk: Move keyboard layout delegate impl to AshKeyboardController.
VirtualKeyboardController is a class for observing changes to input
devices to decide whether to enable the VK. It currently also implements
the keyboard layout delegate interface (for working with displays),
which is not really its responsibility.

This patch moves the logic into AshKeyboardController, which will be
the ash impl of the virtual keyboard interface, so it makes sense that
the code is there.

Bug: 845780
Change-Id: Idf2e1805fbb3b3788dce27ee0ba0b76a9d4b2989
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1657511
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669069}