OnionSoup: Remove ScopedMockClock from tree
This is the last step of the work started in crrev.com/c/1518518. After
replacing all the usages of ScopedMockClock by TestMockTimeTaskRunner we can
safely remove the former. The reason why the latter is preferred is because it
does not need to globally override the time functions, and thus removes the
dependency with Blink's wtf/time.h header.

Bug: 919383
Change-Id: I1c6c1954182f8dcdc86657fd1b8373336d07282b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1593257
Commit-Queue: Sergio Villar <svillar@igalia.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669178}