OnionSoup: Remove SetTimeFunctionsForTesting
After removing ScopedMockClock in crrev.com/c/1593257, the last user of
SetTimeFunctionsForTesting() is gone so we are safe now to completely remove it
from the tree. This means that global time overrides are gone now.

We cannot remove Blink's wtf/time.h though because there are still a few utility
functions there which are widely used in the source tree. They will be removed
in follow up patches.

Bug: 919383
Change-Id: Ic592138c772cfc7f2626c7426f177fd9c7fccaee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1593372
Commit-Queue: Sergio Villar <svillar@igalia.com>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669195}