Create KioskNext Browser factory in Chrome Os.
KioskNextBrowsers are customized browsers. A factory class is required to ensure the correct parameters are used to construct the browsers. The factory class also keeps track of the number of browsers already created to ensure that the number is below a configurable maximum upper limit.

Bug: 965120, 968253

Change-Id: Ie15fe164810a9ef4e50d7069499ad2e96a93b89b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1625650
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Commit-Queue: Yilkal Abe <yilkal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669288}