Fix Supervision Onboarding presentation.
This change fixes an issue where the supervision onboarding would not
be rendered since it had width and height set to 0.

This was caused by a bad interaction between changes:
https://chromium-review.googlesource.com/c/chromium/src/+/1646076
https://chromium-review.googlesource.com/c/chromium/src/+/1659469

One of them fixed a bug in display_manager.js when calculating the
size of a screen.

The other change inadvertently set the screen to an empty state
during initialization so its size was calculated as (0, 0).
This was not detected since the bug fixed by the previous change
caused a JS exception right before the code the code that sets
the size of the inner container[1].

This change just sets a reasonable size for the supervision
onboarding screen by showing the loading state at initialization.

[1]: https://cs.chromium.org/chromium/src/ui/login/display_manager.js?rcl=c0937666bedbe5f4bbbc13bf67931b2715476f9f&l=787

Bug: 958995
Change-Id: I5e2c43a6cbc7a400e959a9d360da597ec43a3d60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1658912
Commit-Queue: Michael Giuffrida <michaelpg@chromium.org>
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Auto-Submit: Lucas Tenório <ltenorio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669388}