[CrOS Cellular] Fix network name strings for Cellular networks.
When a Cellular SIM is locked, the SIM cannot be used to scan for a
mobile network, so the associated carrier's name cannot be fetched. When
a SIM is unlocked, the carrier name is used as the network's name, so
in the locked-SIM case, an alternative name needs to be used. Before
this CL, we defaulted the name to the Cellular Device's name (generally
"wwan0"), which is confusing to users since it's unclear what it means.

This CL leaves the Cellular network name empty when a true name cannot
be determined, and leaves it to the UI to determine the correct name in
this case. When it cannot find a network name, CrOnc.getNetworkName()
defaults to using the network type as the name, so that's what the UI
does with this CL in place.

However, in our UI, we have been using the term "Mobile data" as a
general term that encompasses both Cellular and Tether networks, and
Product/UX specifically would like to use the term "Cellular" in this
case. Thus, I also updated the CrOncStrings so that strings are
available for each of Cellular, Tether, and Mobile, and I made sure to
maintain the "Mobile data" string in <network-summary> and
<internet-subpage>. Note that we have future plans to remove the "Mobile
data" terminology from settings altogether in the near future, so
eventually this edge case will be removed.

Bug: 953230
Change-Id: Icf5cead55b81e709b9f5ff12cfc3f5b5242c845e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1660263
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669434}