Add custom element for 'xf-button'
This CL adds a custom element that implements a button for the
feedback panels.

Included are SVG assets for the different button decorations.

Note, the custom element template includes a 'paper-ripple' child
which does nothing stand-alone but triggers a ripple effect on the
button if combined with the Polymer 'paper-ripple' element in an
application.

Bug: 947388
Change-Id: I345c300349008cbcb1bd96cd294d96886afaf3ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1657380
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669602}