[Fuchsia] Revert to Import/ExportToken to embed GPU-rendered imagery.
Embedding GPU-rendered imagery as a full-fledged View attached within
the top-level View for each window poses the risk that a compromised
GPU process could intercept/observe input events.  We therefore return
to using the Import/ExportToken mechanism.

This is a functional revert of
https://chromium-review.googlesource.com/c/chromium/src/+/1649181
with the exception of retaining some of the function and variable
renamings, which were made to improve readability.

Bug: 974826, WEB-50
Change-Id: I6fc3ebab01c79e93828e870150e5e11276de250f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1661389
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Auto-Submit: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669717}