Quota: Remove vestigial member causing memory leak.
This change removes a member of QuotaTemporaryStorageEvictor that is no
longer used. The member, in_progress_eviction_origins_, is now growing
to hold millions of items for some users and results in an unexpected
consumption of memory for a short period of time. The member is no
longer used for tracking eviction exceptions, so we are removing it to
eliminate the memory leak.

Bug: 969900
Change-Id: If221e329336086d048d86d0c558a338747266af0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1658660
Commit-Queue: Jarryd Goodman <jarrydg@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669723}