[ElementTiming] Add intersectionRect for text entries
This CL adds support for intersectionRect for PerformanceElementTiming
entries referring to text elements. The rect is different from the one
used by Largest Contentful Paint because it does not get projected to
the top-level frame, as that is not acceptable from security/privacy
perspective. However the rect is only computed when needed.

Bug: 942033, 969174
Change-Id: I1f09288b17a73c84daf8cc55996b78e44739a042
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1658630
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Liquan (Max) Gu <maxlg@chromium.org>
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669777}