Use PageMessage to send VisibleViewport sizes to OOPIF renderers.
This CL adds PageMsg_UpdatePageVisualProperties and causes it to be
sent to all renderers for a page when the viewport is resized. At
present, the mainframe renderer continues to act on the viewport size it
receives via RenderWidget::SynchronizeVisualProperties, but any oopif
renderers receive it via RenderViewImpl::OnUpdatePageVisualProperties().

This CL also removes ResizeVisualViewport from the WebWidget API and
moves it to the WebView API.

Bug: 939050, 599688
Change-Id: I04097ea936e8d87792155278c9c1b8823f6aaf45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1621868
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670135}