Include inline JS and onclick content in Chrome RIND report.
Elements with the tag-name "a" or "img" with "onclick" attributes
will be included in the RIND report.

The inline JS is stored in the inner_html field of HTMLElement.
This string is not modified or inspected before the report is generated.

The capturing of inline JS and onclick content is disabled.

http://gpaste/5722548836761600

Bug: 959964
Change-Id: I8c59cfb83fdb211dd5bc8486d4d21eb0378037df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1594034
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Reviewed-by: Luke Z <lpz@chromium.org>
Commit-Queue: Archana Simha <archanasimha@google.com>
Cr-Commit-Position: refs/heads/master@{#670221}