[Previews] Add a flag to override the HintsFetcher timer.
Add a commandline flag, kHintsFetchOverrideTimer to force a hints fetch
without specifying a list of hosts. This allows testing of the
TopHostProvider and HintsFetcherTopHostBlacklist without the timer
delay for fetching hints.

A future change will implement browsertests that utilize this flag to
test the TopHostProvider, blacklist, and blacklist state.


Bug: 932707
Change-Id: I309daedec53be4e250adf2abcc2215dccf8dac51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1657392
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Doug Arnett <dougarnett@chromium.org>
Commit-Queue: Michael Crouse <mcrouse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670226}