Consume portal when disconnecting the content frame.
This CL fixes the HTMLPortalElement so that when the Frame is
disconnected by the ChildFrameDisconnector we also consume the
portal. This prevents anything being executed between the
disconnection of the Frame and the portal being consumed
(which is where we fire focus events and unload events).

Bug: 971702
Change-Id: Ic4cedb1c48df35984ee8a1b02111762aab780089
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1655381
Commit-Queue: Lucas Gadani <lfg@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670256}