Print Preview: Add query parameter indicating locale for GCP requests
If the user's preferred language for the selected Google Cloud Print
(GCP) account does not match the language for the Chrome UI, the names
and descriptions of cloud printers will appear in the account's
preferred language, mismatching the langauge of the rest of the UI.
This also means these printer names/descriptions are not searchable in
the Chrome UI language, which is confusing. Send the locale as a query
parameter in cloud print requests so that the GCP server sends printer
names and descriptions in the appropriate language.

Note: the Accepts-Language header is ignored by the GCP server.

Bug: 974869
Change-Id: Ia63432d2ef83206a7985f38bf83f88ae6a24b63b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1663400
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670353}