Move Windows UI Fallback code to using Skia implementation
font_fallback_win implemented two fallback paths, one own code path
based on IDWriteFontFallback and IDWriteTextAnalysisSource, and one
GDI-based one drawing to an Enhanced Metafile and extracting used fonts
from that (needed on pre-Windows 8.1). The latter was to be replaced
with a code path based on IDWriteTextLayout, which is less crazy then
extracting the fonts from the EMF image.

Investigating the Skia implementation it turns out that this is the
fallback logic that Skia follows: Use IDWriteFontFallback when
available, otherwise use IDWriteTextLayout.

Move the Skia implementation that is already used on Fuchsia and Android
to an implementation file and borrow that for our Windows implementation
as well - the Skia backend on Windows then follow the code paths
explained above.

Tests adjusted to not special case and test the removed
Uniscribe/EMF/GDI backend anymore.

Bug: 947059
Change-Id: I5a26ea690ba19a5edb3407505babf800e7c6983b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1663504
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Auto-Submit: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670447}