Add lock screen media controls view
There is now a feature flag for the lock screen media controls UI. The
flag is disabled by default, but when enabled with the media keys
feature flag (and user preference),a media controls view will be added
to the main view of the lock screen.
The design and shape of the view will change later on, but screenshots
of the current placeholder can be viewed in the corresponding bug.

Bug: 965563
Change-Id: I16aee057a82d994814a602f18e4afe73a87b6ab9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1661177
Commit-Queue: Mia Bergeron <miaber@google.com>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670530}