[CrOS Network] Fix <network-config> password functionality.
When configuring a Wi-Fi network and choosing the security type, the
password field would sometimes not display when it should have; e.g., if
the WEP security type is chosen, a password field would not always
appear.

This issue was caused by an issue with computed bindings; specifically,
the HTML template for this element defines a dom-if using:

<template is="dom-if"
    if="[[configRequiresPassphrase_(type, security_)]]" restamp>

The configRequiresPassphrase_() function would then check |this.type|
and |this.security_| to make a determination of whether to show the
password field. The |this.security_| property was modified by another
element on the page, and each time the property changed, we would
re-assess whether the password section should be shown.

However, when this function was run, |this.security_| was actually the
*previous* value, from before the property was changed. According to
[1], configRequiresPassphrase_() should actually be taking the two
properties as parameters instead of relying on |this.*| within the
function itself. I converted to this format, and it fixed the issue.

Note: There are a lot of functions throughout WebUI which follow the
same pattern which caused the bug, but they don't seem to cause the same
bug here. It's possible that this is some sort of race condition in
Polymer's data binding code.

[1] https://polymer-library.polymer-project.org/2.0/docs/devguide/data-binding#annotated-computed

Bug: 976385
Change-Id: I0449da9465a08f2ce107ac43c32644921b5e368d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1666508
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Auto-Submit: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670563}