Fix WebXR samples viewer pose height for inline sessions
Make WebXR samples viewer pose height correct for inline sessions on
both WMR and Android. When using WMR, it does not support a local-floor
reference space for inline sessions. In that case, the tracking space
origin in world space must account for an estimated user height.
Android supports local-floor for inline sessions, so what really needs
to be checked when determining whether to manually apply an estimated
user height is the reference space type instead of the session type.

I also verified that the behavior is correct when using Oculus.

Also fix an issue with teleportation.html that caused onSelect in
inline mode to stop working after entering and exiting an immersive
session. This requires keeping track of separate viewer spaces for
both sessions.

Bug: 954388
Change-Id: I302666c9c8ce2faaf73454a04dcde425a11bc042
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1666302
Reviewed-by: Brandon Jones <bajones@chromium.org>
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Commit-Queue: Jacob DeWitt <jacde@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670661}