arc: Extends charts till the end in tracing tool
Some metrics are captured relatively rare. For example GPU frequency,
when it is actually changed. That may lead to case when corresponded
chart is not extended till the end. We can automatically extend these
charts because we know it does not change at this interval.

TEST=Locally + unit test
BUG=b/135617781

Change-Id: I9bea72d9571bb2e584e9abb8b8ec219f27836f0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1666305
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Yury Khmel <khmel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670669}