Fix the issue that the shelf has wrong bounds after drag
When pressing the AppList button by mouse during the drag on shelf by
gesture, the RootView of ShelfView will reset the gesture handler. As
a result, the ShelfView cannot receive the gesture end event. It leads
to the wrong bounds of the shelf after drag. In order to fix it,
explicitly cancel the drag in progress if any in this scenario.

Bug: 968768
Change-Id: I023ba730fafcb601c5abba2bc3f1c74fe9b2fd84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1664206
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670675}