Code to auto-generate the seed corpus for the wayland fuzzer
For the wayland fuzzer, we can enumerate sequences of calls which we
will use as the seed corpus. The enumeration strategy is fairly simple,
first we make a dependency graph showing which interfaces are needed
to call certain methods, and which methods create which interfaces. So
a valid sequence is the sum of all paths from the chosen method to the
root wl_display interface.

The coverage is ~25% for exo using only the partial list of wayland
specs (as per BUILD.gn, only wayland and xdg_shell)

Bug: 961564
Change-Id: I0da406003fa61fec1bb21543253f33e3266435f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1654678
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670798}