Move observers of access token-related events to OAuth2AccessTokenManager
This CL is a part of moving access token management to
OAuth2AccessTokenManager.
It splits OAuth2TokenService::DiagnosticsObservers into
the observers of access token-related events and the
observers of refresh token-related events. The former
is moved to AccessTokenDiagnosticsObserver and the
latter is maintained in
OAuth2TokenService::DiagnosticsObservers.

The classes which need to observer access token-related
events are inherited from AccessTokenDiagnosticsObserver
and are added through
OAuth2TokenService::AddAccessTokenDiagnosticsObserver.

Bug: 967598
Change-Id: I25e42dd75cc14fdadf5c1c8f09ee9069717dfd5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1666683
Commit-Queue: Julie Jeongeun Kim <jkim@igalia.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670847}