[AA] Fix a drop out metrics enum collision.
Before this change, failing to install the feature module or failing to
start due to no initial scripts were mapped to the same drop out reason
enum value.

Thich patch changes the enum value and the name for the DFM entry. It is
not due to the DFM being cancelled, it is triggered when a DFM
installation failed alltogether.

R=arbesser@google.com

Bug: b/132424979
Change-Id: Ic235f4b503f059ecb1ee21b5cb1bc213e578df1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1669351
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670869}