Unfriend ukm::UkmRecorder and TestAutofillClient
There are basically two ways of addressing this task:
1) Having a test own a ukm::TestAutoSetUkmRecorder and injecting that
into the TestAutofillClient, or
2) Having the TestAutofillClient own the ukm::TestAutoSetUkmRecorder.

Approach 2) does have the risk of a second class creating a
ukm::TestAutoSetUkmRecorder as well. But on the other hand it was
way less work to take this path (there are many instances and derived
classes of TestAutofillClient) and we can refactor the
ukm::TestAutoSetUKMRecorder out of the TestAutofillClient if we
actually need that. Therefore, I went with 2).

Bug: 962485
Change-Id: Ib339c5c2f048656c7f5cb5c8c25b321c010efcb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1662412
Reviewed-by: Robert Kaplow (slow) <rkaplow@chromium.org>
Reviewed-by: Roger McFarlane <rogerm@chromium.org>
Reviewed-by: Maxim Kolosovskiy <kolos@chromium.org>
Commit-Queue: Maxim Kolosovskiy <kolos@chromium.org>
Auto-Submit: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670919}