Convert FTP auth cancellations to net errors
This CL is part of converting auth requests handed by LoginHandler to
committed interstitials. The general pattern is to cancel requests
upon receiving an auth challenge, commit an interstitial, and show the
login prompt on top of it. For FTP auth, unlike HTTP auth, there is no
error page from the server to commit, so we instead want to commit a
blank page on top of which to show the login prompt.

Currently, when you cancel an FTP auth prompt, you get an empty
download. This is a bit weird and also doesn't provide us with an
error page to commit, so this CL instead converts auth cancellation to
an ERR_FTP_FAILED error. A future CL will convert this error page to a
blank page with a login prompt on top of it.

Bug: 972188
Change-Id: I7ab9a52cf4a43eabfc8cba7772f20efe1b8e450a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1663586
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Emily Stark <estark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670976}