Canvas: improved the speed of drawing unused canvas
In our current implementation, a transparent image is created every time when
GetSourceImageForCanvas is called, so drawing to an unused canvas element is
much slower. After this cl, when draw is called for the first time, a
transparent image is created and saved, i.e., no re-initialization needed
when draw is called repetitively to un-used canvas.

Change-Id: Ibb795fd5e2d93e4df53218a8340c8b723d8ee01b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1663169
Commit-Queue: Yi Xu <yiyix@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671007}