Refactor saving the login password in the kernel keyring
Refactors the code that sends the login password to session manager to
store it in the kernel keyring. The goal of this refactoring is to
allow a simple addition of more services that want the login password
saved.

To add a new service, add it to the PasswordConsumingService enum and
call VoteForSavingLoginPassword as soon as the service knows whether it
needs the login password (usually right after parsing policy). Note that
this call must happen before user_context_.ClearSecrets() in
UserSessionManager::FinalizePrepareProfile(), which wipes the login
password in Chrome.

In this CL, there are no behavior changes. The next 'user' will
be the Kerberos daemon in a subsequent CL.

BUG=chromium:952240
TEST=Have a network that requires the same password as the user's login
     password. Set a OpenNetworkConfiguration using a ${PASSWORD}
     placeholder. Make sure network authentication still works.

Change-Id: Ic6f72f2876446bf8762eeae489700517545ad59a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1656569
Commit-Queue: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: May Lippert <maybelle@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671079}