Make OfflineContentAggregatorFactory a SimpleKeyedServiceFactory.
OfflineContentAggregatorFactory is currently a singleton in Android
but a per-Profile object in other platforms.

This CL changes it to be per-Profile in all platforms by transforming
it into a SimpleKeyedServiceFactory, which can be used in the absence
of a BrowserContext.

This is motivated by an integration test (crbug.com/959957), but is
a cleaner design too.

Bug: 974825,959957

Change-Id: I125edac1c3192bdefdf62564f36e3b2e6487a1c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1660919
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: Rayan Kanso <rayankans@chromium.org>
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Xi Han <hanxi@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671130}