[Navigation Timing] Replace TAO restrictions with same-origin ones
The NavigationTiming specification changed [1] to replace TAO based opt-in
to get information about unloadStart, unloadEnd and redirectCount with
same origin restrictions.

Our implementation was not aligned with the old spec language, but
isn't aligned with the new one either.

This CL aligns the implementation with the new language.

Change-Id: Ic4f3c7061e7b9a40f960f0bf26ef26823d5fe743
Bug: 977147

[1] https://github.com/w3c/navigation-timing/pull/108

Change-Id: Ic4f3c7061e7b9a40f960f0bf26ef26823d5fe743
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1669355
Commit-Queue: Yoav Weiss <yoavweiss@chromium.org>
Reviewed-by: Nicolás Peña Moreno <npm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671229}