[getUserMedia] Restore arbitration for EC selection.
This CL reintroduces the final arbitration in SelectBestEcMode which was
removed in crrev.com/c/1626414. However, we reimplement the arbitration in a
simplified way. Firstly, we have added the possibility to prefer system echo
cancellation also when the experimental bit is set. As a result, the
arbitration, can only select between AEC3 and disabled EC.

BUG=976646

Change-Id: I81b9a936e17a2ba08902c7fa811047751af83559
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1669299
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671236}