Add DB cannot be opened error state for loading credentials.
Bug 976143 shows a case when the token database could not be opened.
My suspicion is that another process had a handle on this file and
that forbids Chrome for using it. However debugging it was fairly
complex.

This CL separates this state to make debugging easier.

Bug: 976143
Change-Id: I2106201c3ae17299a3ce3d2e4675bb90bab1f062
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1670223
Reviewed-by: Thomas Tangl <tangltom@chromium.org>
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671344}