LazyLoad: Mark "loading" attribute support as experimental.
This CL marks "loading" attribute support as experimental, so that tests
run with this attribute supported. Note that this only turns on lazy
loading of images and iframes marked with "loading=lazy", and does not
turn on any automatic lazy loading of images/iframes that aren't marked
"loading=lazy".

Bug: 937986
Change-Id: Iccf1c4614f648f0596198e9c76eafac818daac23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1636639
Commit-Queue: Scott Little <sclittle@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671465}