Make feedback panels flex end align.
The feedback panels host two feedback DisplayPanels, one for
completed tasks, and the second for activities in progress.

This CL changes the layout of their parent hosting container
to make sure that if both are visible and have different
widths, that the panels align a the flexbox end side (i.e.
aligned on the rigth hand side for LTR layout).

This also introduces a vertical gap between the two sets of
panels so they don't abutt when both are visible.

Bug: 947388
Change-Id: I78a780cb88a81bd3f24c355f483d125d57cb0852
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1673565
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671586}