Make web apps with a supported related app intent to the store for unified install.
This CL allows Chrome OS to query ARC for whether a PWA has a supported
related app available and installable. If it does, the installation
action redirects to the Play Store. A test is added to ensure correct
functionality.

This follows on from https://crrev.com/c/1642753, which implemented this
functionality for BookmarkAppHelper. This CL replicates the
functionality for the unified install pipeline, which is replacing
BookmarkAppHelper.

BUG=969560

Change-Id: I54e5392037cd5cc5fe247c180384a12000629870
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1666650
Commit-Queue: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Auto-Submit: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671593}