Create ChromeSigninManagerDelegate JNI
This introduces the Jni counterpart of the java
ChromeSigninManagerDelegate. This interface will provide access to android
specific dependencies on //chrome/browser to the componentized java
SigninManager.

This first iteration creates the new objects and implements
stopApplyingCloudPolicy, fetchAndApplyCloudPolicy, isAccountManaged and
getManagementDomain.

As discussed in CLs on the java counterpart (bug 963400) this object
will hold pointers to singletons such as the ActiveUserProfile.
The native object is created by the java object.

details in document:
https://docs.google.com/document/d/18887XeZNJ9pmoTdJducssk5_yVU2CQ3EqaRFoMqWCCk/

Bug: 963402
Change-Id: Iccbb6b30d23f26fc5960b5d9db9c122df3e589f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1642547
Commit-Queue: Pâris Meuleman <pmeuleman@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671684}