Use ModalDialogManager for missing Android permission dialog
Currently, we use AlertDialogs for displaying the dialog that shows up
when there is a missing Android permission. This CL replaces this
approach with ModalDialogManager for displaying said dialog.

This CL also adds support for generating a touchless version of the
dialog model for touchless devices.

Bug: 974271
Change-Id: I786e2a7f7e2e48f57abd5d3a4674a51c902be51f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1667748
Reviewed-by: Theresa <twellington@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Becky Zhou <huayinz@chromium.org>
Commit-Queue: Mehran Mahmoudi <mahmoudi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671728}