ChromeVox: Report SetSize of Expanded Popup Buttons.
This change allows popup buttons to take SetSize values, which is
inherited from the wrapped menulist popup, if one is present.The
SetSize is then reported to the user in ChromeVox if the popup button
is expanded.

Tests have been added to ensure correct behavior in ui/accessibility
and ChromeVox.

Bug: 914666
Change-Id: I8f46b3d0fab297e12d5f70a8100984455f1e0078
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1666779
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671825}