Expose SW notificationclick's interaction state to extensions.
This CL exposes ServiceWorkerContextProxy to ContentRendererClient's
DidInitializeServiceWorkerContextOnWorkerThread, so that extensions/
can query the state of AllowWindowInteraction() through it via
ServiceWorkerGlobalScope.

This allows Service Worker based extensions to call extension APIs that
require user interaction to be present, in onnotificationclick handler.

This CL adds test for this feature.

Bug: 977629
Change-Id: I28df9ee0724dbc76f5525d9111416e9180ad3d5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1673253
Commit-Queue: Istiaque Ahmed <lazyboy@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671916}