Specify the Windows version for chromium.webrtc.fyi bots
Otherwise they're running on a random machine with arbitrary OS version, including the bots that are explicitly called like "Win8 Tester".

Right now the main problem with this is that Chromium fails to build on Windows 7, and due to random selection of machine the build appears "flaky".

Bug: 976184
Change-Id: I5de22cf1072d003856b89ef152f35d22436de991
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1672849
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Oleh Prypin <oprypin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671974}