Show WarningStatusString on ChromeOS for dangerous downloads
Currently, for dangerous downloads on ChromeOS, we show a notification
with an empty message, and the progress_status set to the
WarningStatusString. But for a blocked download, we don't show the
progress bar, so the WarningStatusString is ignored. This CL sets the
notification message to the WarningStatusString.

Previous behavior: https://screenshot.googleplex.com/GOtztKVUrJe.png
New behavior: https://screenshot.googleplex.com/hHu62tdx6vx.png



Bug: 973000, 807334
Change-Id: I48b060d57d1bc0ad26e5c02d0cdfce2b3664d9c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1663081
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672199}