Add checkbox to the Add Share dialog to save SMB credentials
SMB shares will save the username/password, and those saved credentials
will be used when the share is restored on next login. If the user
explicitly removes the share, the saved credentials will be deleted.

In order to accomodate the extra field, also increase the height of the
share dialog by 40px (total height of the checkbox field).

BUG=891462

Change-Id: If902f23a8b4f4da2ec8d9055ed2492ca1ff424f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1666427
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Anand Mistry <amistry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672394}