Factor SourceType enum out of AudioDeviceFactory class
.. and move it to Blink exposed API, so it can be used by both
content and blink code while audio_device_factory.cc|h is not Onion
souped.

Eventually, the enum can be either moved back to AudioDeviceFactory
class, if it ends up being part of the Blink API, or left as is,
given that there is not plans to Onion soup the chromecast/media/audio code.

This is a preparation CL for execution phase 3.3.10 on the
design document [1].

[1] https://docs.google.com/document/d/1rHJGi1U72qZsOGIctIf7GKRaY8d7BgtVW8MaMYumzYY

BUG=704136
R=guidou@chromium.org, haraken@chromium.org

Change-Id: I1b57f4b1493994fb1438df23a34cdb02c2d42208
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1677387
Reviewed-by: Kenneth MacKay <kmackay@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Antonio Gomes <tonikitoo@igalia.com>
Cr-Commit-Position: refs/heads/master@{#672593}