Keep all SL Fragments in preferences
When this rule was previously added, it was narrowed to Preference
Fragments, but there are other fragments in preferences (also
instantiated via reflection), which extend the support library Fragment
class directly.

Tracking this change's binary size impact as part of the original
regression tracker for this -keep flag.

Bug: 978754, 972581
Change-Id: I9c5fe9c9bf621bad8755ecd8f677c840ed5ed32b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1678941
Reviewed-by: Sam Maier <smaier@chromium.org>
Commit-Queue: Natalie Chouinard <chouinard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672638}