[iOS] Move UI delegate and observer out of OverlayPresenter.
This CL creates a separate file for OverlayPresenter::UIDelegate
(renamed to OverlayPresentationContext).  This was done because the
presentation context will need to support observation so that the
presenter can be aware of when it has a coordinator in which to present
overlay UI.  The class was renamed because it seems like an anti-pattern
for a delegate to support observation.

This CL also moves OverlayPresenter::Observer to a separate file for
consistency.

Bug: 941745
Change-Id: I92208a28a84a450d3a00aa08c7dfbf546bb7ff98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1671028
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672660}