Releasing ids for hovering touch move events
For smart boards, the devices are receiving some touch move events when
there is no finger touching on the screen, like hovering. For those
touch move events, we do not release their ids. We should check if the
touch pointer are in contact with the screen, if not, release the ids
for these touch move events as well.

Bug: 874948
Change-Id: I61f7cb75f92eb600b0dbbbd8f9fb7184c6c0bda1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1670159
Commit-Queue: Lan Wei <lanwei@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672694}