Convert ScopedBlockPopupsPref to be compatible with EG2.
App calls made by this test class are moved into ChromeEarlGrey and made
to compile with EG2, so that ScopedBlockPopupsPref can continue to work
when moved into test-side code.

ChromeBrowserState was removed from ScopedBlockPopupsPref as I was not
sure if it would be passable over EDO safely, and all callsites of this
class use GetOriginalBrowserState(). Therefore, I moved getting the
ChromeBrowserState using GetOriginalBrowserState() to app-side in
CEGAppInterface.

Bug: 922813
Change-Id: I954d28309360b8b805a9943c7fde77df69f7df95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1668512
Commit-Queue: ericale <ericale@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672710}