Add cc::MirrorLayer[Impl]
This new type of layer can be used to mirror contents of another layer.
When a mirror layer is set to mirror another layer, the latter is forced
to create a render pass so that the former can add a RenderPassDrawQuad
referring to that render pass, mirroring its contents.

Note that this CL does not add damage rect handling for the mirror
layer. That would be done in a follow-up CL.

The first use case for the mirror layer would be creating docked
magnifier on CrOS. The current implementation of docked magnifier does
not work with OOP-D.

BUG=947565

Change-Id: I1609edf53eb8c1ae8faff5345ee41dc547a916a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1652202
Reviewed-by: enne <enne@chromium.org>
Commit-Queue: Mohsen Izadi <mohsen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672744}