Add Net.SSLHandshakeDetails metric
Between TLS 1.2 (and earlier) and TLS 1.3 and the various optimizations
applied to either, we have many different handshake modes with different
performance and sensitivity to system state. E.g., losing a TLS 1.2
resumption to False Start is less of a big deal to losing one to a full
handshake.  Add a metric that breaks down the various scenarios.

This is relevant for both double-keying the TLS session cache and 0-RTT.
As part of testing this, add ALPN support to SSLServerSocket.

Bug: 641225, 974910
Change-Id: I9b24ae361f21444c3217b7c0e1619ab11e8eccee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1676987
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Steven Valdez <svaldez@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672761}