Introduce TestWithLocaleAndResources for iOS WebView unit tests
So tests needs to have access to localized resources. This requires
setting up the resource bundle before creating the test BrowserState
and then cleaning it up during shutdown.

Instead of copy and pasting the code in each tests that need this
(and having a hack to call the method before the creation of the
instance variable), introduce a class TestWithLocaleAndResources
that tests can inherit from that takes care of the registration
and cleanup in constructor/destructor.

Bug: none
Change-Id: I8086b38ec26797fb48631e8c1a60cff5e52cb3c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1678320
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: John Wu <jzw@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672864}