Use ElapsedTimer in FileReader to fire progress events
FileReader is using a double to store the last time a progress event was fired
in order not to notify more than once per 50msecs interval. This looks like the
perfect candidate to use base::ElapsedTimer instead. By switching to
ElapsedTimer we get rid of double <-> base::Time conversions and an extra
multiplication to convert to milliseconds.

Bug: 979137
Change-Id: I4d8f2236f6c8fe56c9796df87fce33b202493089
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1679758
Commit-Queue: Sergio Villar <svillar@igalia.com>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672945}