[BrowserSwitcher] Save external greylist to cache.dat
When the new external greylist policy is set, the contents of the
greylist will be concatenated with the regular greylist in cache.dat.

This CL also turns sitelistcache.dat into empty files. When the IE or
external sitelist changes, it is written to cache.dat instead.

Bug: 958659
Change-Id: Ia9193397f58a35dc2614308d337d6cc2458e184b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1675915
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672982}