Adding CORB protection UMA metrics (part 2): range requests.
Adding a UMA histogram that reports if the server supports range requests on
sensitive-seeming resources. We check for the accept-ranges header. If it is
present we know the server supports range responses, and otherwise we assume it
does not.

We measure this because multipart range requests would allow bypassing CORB.

Bug: 815290
Change-Id: If39bd78cfaf28bcf73eb4acb40a983147da3c375
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1671127
Commit-Queue: Kristina Nelson <krstnmnlsn@google.com>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672996}