[Context menu redesign] Add extra bottom padding to the list
This change required us to put the ListView in a FrameLayout and use
the FrameLayout as the background view. The reason for this is that the
background is a 9-patch and gets extra padding around the view
automatically, requiring clipToPadding to be true. Moving the background
to the parent view allows us to set clipToPadding to false to prevent
the extra bottom padding from cutting the list items.

Bug: 655359
Change-Id: I3e05f758d0d85adc323c4038216539aacb4a0570
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1678939
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Reviewed-by: Theresa <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673017}