Scope UIA accessibility dumps to requested node
AccessibilityTreeFormatterUia::BuildAccessibilityTree() was ignoring the
start node it was given and dumping the accessibility tree from the host
HWND. For existing HTML tree dumps, this just means we have an extra
node in UIA dumps compared to IA2 dumps. But for PDF, starting from the
HWND would mean capturing UI elements in the host document, which would
be out of scope for testing PDF accessibility.

This CL scopes UIA dumps by constructing a RuntimeId corresponding to
the start node and locating that node within he client-side tree.
Existing HTML UIA dumps are updated - the affected line count is large,
but the only actual change is removal of the fragment root node, which
results in every descendant node moving up one level of tree depth.

Bug: 974397
Change-Id: I8cfaae95d482fca02e066f79f0195bddf81dbc5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1674196
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#673204}