Do not call TextInputClient::InsertText if provided text is empty.
There are differences in the behaviors in Textfield::InsertText and
RenderWidgetHostViewAura::InsertText when the provided text is empty.
TSF uses TextInputClient::InsertText to commit current composition and
expects all implementation classes to have same behavior. The text will
be committed only if the provided text is not empty in
TextField::InsertText. However, in RenderWidgetHostViewAura::InsertText,
composition will be committed with current composition text if provided
text is empty.

As a short term fix for Windows platform, only call |InsertText| if text
is not empty. Use crbug.com/978678 to keep track of long term solution.

Bug: 977820
Change-Id: I3eb1bbc9ed6f98f8356fda01cd9cc2ee91a79a67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1675465
Commit-Queue: Siye Liu <siliu@microsoft.com>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Shu Chen <shuchen@chromium.org>
Reviewed-by: Yohei Yukawa <yukawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673392}