Dedupe UpdateFrameColor/UpdateFrameColors BrowserNonClientFrameView methods
BrowserNonClientFrameView::UpdateFrameColors() existed along side
BrowserNonClientFrameViewAsh::UpdateFrameColor(). These methods achieve the
same intended outcome but were named differently.

This CL merges them and fixes a bug where PWA windows on Ash were not
updating their frame colours in response to changes in the page's
<meta name="theme-color"> tag.

Screencast after CL: https://bugs.chromium.org/p/chromium/issues/attachment?aid=398086&signed_aid=YEVIswBhmEeXyH5qxXIAMw==&inline=1

Note that Ash will animate any changes in the frame colour which conflicts
with pages that attempt to animate the frame colour. The effect of this is
the overall animation stalls due to the ease-in nature of the Ash
transition timing function.
This is a known issue that is not being addressed in this CL.

Bug: 872121
Change-Id: I11bbf4d278e54faadcecd96c7d8cb61f694c5371
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1681266
Reviewed-by: Bret Sepulveda <bsep@chromium.org>
Commit-Queue: Alan Cutter <alancutter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673641}