Connect up confirm-password-change dialog
Wires up the confirm-password-change-dialog so that it fulfils its basic
functionality - when password scraping fails during an in-session
password change, the user is prompted to type their passwords in to
the dialog.

Adds a system dialog around chrome://password-change, and a handler
so that when save is clicked, cryptohome password is changed.

Still lots more TODOs to give this proper polish.

Bug: 930109
Change-Id: I0eadb73963c78b3409a7608c4ef708ca825bf669
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1679701
Commit-Queue: A Olsen <olsen@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673691}