Allow for non-fast scrollable regions on non-scrolling layers
This patch refactors the impl-side hit testing code to allow for
non-fast scrollable regions on all cc::Layers/cc::LayerImpls, rather
than just on scrollable ones. This is useful because non-fast regions
can be in transformed layers that move, and can become stale. This
patch prepares for PaintNonFastScrollableRegions which will fix this
by setting non-fast regions on non-scrolling layers.

Some background on scroll hit testing is useful in understanding this.
LTHI::FindScrollNodeForDeviceViewportPoint starts with a hit layer and
looks up the associated ScrollNode. Each ancestor of that scroll node is
then checked by calling LTHI::TryScroll. Before this patch,
LTHI::TryScroll would look up the layer associated with each ScrollNode
and check if the layer's non-fast region was hit. This skips
non-scrolling layers which required that the non-fast regions be
accumulated onto the nearest scrolling layer.

This patch starts by looking up all layers that hit a non-fast region in
LTHI::FindScrollNodeForDeviceViewportPoint. Then, the set of ScrollNodes
for each hit non-fast layer are collected; this can be considered the
set of non-fast scroll nodes. When iterating up the scroll tree in
LTHI::TryScroll, a scroll is considered non-fast if one of the nodes is
in the non-fast scroll node set.

This patch may affect performance because an additional layer list walk
is needed. This should be inexpensive in the common case when the layer
has no non-fast regions.

An end-to-end prototype of PaintNonFastScrollableRegions is at:
https://chromium-review.googlesource.com/c/chromium/src/+/1619208
A design doc is at:
https://docs.google.com/document/d/1IyYJ6bVF7KZq96b_s5NrAzGtVoBXn_LQnya9y4yT3iw/view

Bug: 864567
Change-Id: I7f56e2e0f6bbbbc4c8b072b091f019b7a4b9a5da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1683358
Reviewed-by: enne <enne@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673821}