[BGPT] Don't set transform_changed when rebuilding cc property trees
Previously we set the flag even if the transform didn't change,
causing excessive surface damages.

Cc will check if the transform of a layer really changed after a
property tree rebuild, so we don't need to set the flag.
cc::PropertyTreeBuilder doesn't set the flag either.

We still need to set the flag on direct transform updates.

Bug: 980165
Change-Id: I55f71ecda8661c97648c9aeca636b5e9226e9405
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1683666
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673838}