Make AwURLLoaderThrottle work on UI thread
We are in the process of migrating NavigationURLLoaderImpl to the UI thread,
which includes migrating all throttles. This makes the AwURLLoaderThrottle able
to be created on either the UI or IO thread. Since the headers are protected
with a lock in AwResourceContext, it should be fine to access them on either
thread.

Bug: 978907
Change-Id: I78c4063f3e48dcde395737bd7a267c35a7131445
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1680981
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674066}