Use the Delegate instead of PrefServiceBridge for PaymentRequestImpl
In the C++ code, the PaymentRequestDelegate will return a prefs
service, and tests replace the system service with one of their own.
The Java bindings through PrefServiceBridge do not allow injecting
a replacement prefs service on the C++ side (for now?).

So instead of returning a PrefServiceBridge from the delegate, or
changing the global PrefServiceBridge.getInstance(), have the
PaymentRequestImpl ask its Delegate if CAN_MAKE_PAYMENT is allowed.
Then the production impl can ask PrefServiceBridge, and a testing
override can specify the value that it wants.

R=danyao@chromium.org, rouslan@chromium.org

Bug: 961849
Change-Id: Ic8bd089c197b94f05c7c2ac57224110ee6759fdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1680994
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Danyao Wang <danyao@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674086}