Change SkiaRenderer blink_web_tests to GL
Originally we deployed a skia_renderer_blink_web_tests with
SwiftShader. However on the Chrome-GPU fleet there is a deadlock
between the real Xorg on :0 and Xvfb on :99.

So we do not want to continue with this config on these bots.

This change introduces skia_renderer_gl_blink_web_tests which makes
use of "--no-xvfb" and GL renderering. To prevent this deadlock
from being encountered on the GPU bots.

A new expectations file has been created for the feature
"enable-gpu-rasterization" to encompass known faliures.

TEST=skia_renderer_gl_blink_web_tests

Bug: 966064
Change-Id: I1d908c454c2c2e67bceb33f5734f4593f8ec1b73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1679985
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674099}