Don't show feedback checkbox on profile errors on non Google builds
If Chrome started up and had a corrupted profile store, it would show a
dialog box about this and on non-ChromeOS platforms there would be a
checkbox for filing feedback. This checkbox was always there, even on
builds that wouldn't allow sending of feedback (so it was a no-op). This
disables that checkbox on those builds.

Bug: 755680
Test: Manually verified
Change-Id: I7a37e57652aa3456999c99a9d4b69b72965318b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1684354
Auto-Submit: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Cr-Commit-Position: refs/heads/master@{#674153}