Move movementX/Y calculation for aura pointer-locked state to Blink.
This change moves movementX/Y calculation for aura pointer-locked
state to Blink.
On aura, pointer lock use "move to center" to make cursor stays in
the window. With the calculation done in blink side, the move to
center event is marked as synthesize move, so that we can update
the blink side states and do not dispatch the event.

The change is under content feature flag kConsolidatedMovementXY

Bug: 802067
Change-Id: I05360dadd18a2f41481a0de9ef78a05199493857
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1618306
Commit-Queue: Ella Ge <eirage@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674237}