media/gpu/test: Fix video decoder thumbnail test.
On boards that don't support import mode, which is required for the frame
validator to work, we render all frames to a large thumbnail and calculate its
checksum. CL crrev.com/c/1677770 fixed a bug causing some resources to not be
cleaned up. Unfortunately this change broke the thumbnail test, since we need
an active GL context when destroying the decoder.

TEST=./video_decode_accelerator_tests --disable_validator on eve

BUG=None

Change-Id: Icd15a92f80260fc8851cace5a833b3039a2c9888
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1681418
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674280}