identity_utils: Internalize knowledge of pref name
In a quick followup to
https://chromium-review.googlesource.com/c/chromium/src/+/1682630, this
CL folds the knowledge of the pref name that specifies the allowed
username pattern into IsUsernameAllowedByPatternFromPrefs() rather than
requiring clients to pass it.

This CL also builds identity_utils_unittest.cc, which I just noticed
was not previously in the build -- d'oh!

Bug: 964235
Change-Id: Ia85565a336c8940a5d0f098cf6da1026c1a866bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1683718
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674384}