Open the "Update password" bubble if the generated password can't be presaved.
The generated password is first presaved with the detected username. If it
causes a conflict then the username is cleared. If there is still a conflict
then the update bubble is shown. Dismissing it stops the generation flow.
Accpeting it allows the generated password to be presaved and prefilled.

Bug: 936011
Change-Id: I49d4d1cb8b3b5072674a47e8e12964f512e9fa93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1679699
Commit-Queue: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674437}