Make GPU Gold tests output artifacts
Makes the Gold version of the GPU pixel tests output the Gold triage
link as a link artifact. This will be picked up by the recipe and
displayed in Milo so that users don't have to dig through logs in order
to triage images.

Bug: 968276
Change-Id: I0e2b9492ac4e54de8f39709fc49c47b9dc5b4787
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1686540
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674542}