[WebView] Fix RootScrollOffset for scale factor
The SetSynchronousInputHandlerRootScrollOffset API used by WebView
expects to set a root scroll offset specified in document coordinates.
i.e. The same offset should go to the same part of a document regardless
the current zoom level.

In https://crrev.com/f93bf42468f108274e3ea97853679047465a8412 I changed
this method from LayerTreeImpl::DistributeRootScrollOffset to
Viewport::ScrollBy to reduce the number of scrolling code paths. This CL
accidentally missed the fact that Viewport::ScrollBy expects the delta
to be in "screen" coordinates and so it scales the delta by the scale
factor inverse so that content appears to stick to the finger. This
means the SetSynchronousInputHandlerRootScrollOffset was dependent on
the zoom level and broke cases like those in the linked bug.

This CL fixes the issue by having
SetSynchronousInputHandlerRootScrollOffset apply the scale factor before
calling ScrollBy so that the desired delta gets applied to the document.

Bug: 973771
Change-Id: I704a44334db915f8a14a1d903315fafa5c1349cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1686335
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Auto-Submit: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674559}