Portals: Handle delayed touch event acks from predecessor
This CL handles two cases. The starting steps for both cases are the
same, a portal is activated in a touch event handler and finishes
activation before the event handler finishes running (resulting in the
touch event ack being delayed). Activation results in the RWHV for the
predecessor page being destroyed (and replaced with a RWHVCF), and
the TEAQ being cleared. The cases differ as follows:

1) We receive a touch ack after the predecessor page has been adopted
   as a portal.
2) We receive a touch ack after the predecessor page has been adopted
   and reactivated, and is now the main page again.

For case 1, the RWHVCF cannot process a touch event ack, so we need
to prevent it from receiving one. For case 2, we could have new
touch events pending ack, and receiving an ack for an event preceding
activation can put things in a bad state, so we also need to prevent
the RWHV from receiving the ack.

To solve both cases, we extract the events that will get acked when
the predecessor RWHV is destroyed (before it is actually destroyed)
and then tell the appropriate RWHs to ignore any acks they receive
for these events.

Bug: 973647
Change-Id: Ic8c5260e1f28575974ea36d0f1b27436269578c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1656808
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674841}