[LCP] BgImage refactor: a vector map replaces double-key map
ImagePaintTimingDetector used to use double-key map to store the image
record of background image. This design, however, posed a challenge
when we need to search for image records with half of the key
(node_id). To address this issue, we change to use a vector map to
store the image records of background images.

The vector map is a map with the value being a vector of image-records,
and the key being DOMNodeId.

Note that this is a pure refactoring.

A follow-up CL will combine the handling logic of background image and
normal image.

Bug:972146

Change-Id: Ia9a1785ef086878e1dad09b43ad6b70fa06da9fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1688245
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674871}