Set Referrer String + Referrer Policy + Debugging Info
This is a re-land of "Set ResourceRequest's referrer string and policy
separately" https://crrev.com/c/1601813. That CL caused crashes
exclusively related to stale-while-revalidate requests, which have
proven impossible for me to reproduce locally. This CL does not fix the
crash, but instead adds information to the stack when Chrome is about to
crash letting us know the path that the erroneous ResourceRequest took.
This CL will be reverted when related crash reports are collected, and
the CL will be re-landed with an actual fix next time.

The original CL's description is as follows:

> This should be the final CL for 863769. Before this CL, there
> are situations where we eagerly set a ResourceRequest's `Referer`
> with ResourceRequest::SetHttpReferrer, sometimes invoking
> SecurityPolicy::GenerateReferrer(...). These calls are equivalent
> to the Fetch Standard calling into "determine request's referrer"
> algorithm.

> After this CL, we opt to set a ResourceRequest's referrer policy and
> string members separately as much as possible, minimizing the number
> of calls into ResourceRequest::SetHttpReferrer(SecurityPolicy...). The
> Referer header will finally be generated when ResourceFetcher calls
> into SetReferrer (equivalent to "determine request's referrer"
> algorithm).

> This CL also adds more CSS security tests regarding the `Referer` header.
> The only observable effect from this CL is that some shared worker referrer
> policy tests start passing because we invoke SetReferrer at the right time.

R=kinuko@chromium.org, yhirano@chromium.org

Bug: 863769
Change-Id: I35752faad95a97827ddbb4df9e2372d0cf9f1ec5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1690692
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675181}