Use IdentityManagerBuildParams in PO2TSBuilder
This is an intermediate step in the process of moving all of
ProfileOAuth2TokenServiceBuilder code into IdentityManagerBuilder. To do
this, the building of PO2TS cannot use Profile nor any code from
//chrome/browser.

In this CL, ProfileOAuth2TokenServiceBuilder is modified to use
IdentityManagerBuildParams to obtain the AccountTrackerService,
AccountConsistencyMethod, NetworkConnectionTracker, SigninClient and
PrefService.

Bug: 974198
Change-Id: I76045d59b5ef3e056f52f574c98ada70b3f6cc8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1688960
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675185}