[Autofill Assistant] Declare and forward direct action arguments.
With this change, user actions can be passed experiment ids and
parameters to be included into the GetActions and GetNextActions RPCs,
in addition to the global ones.

This allows making arguments passed to direct actions, called through
Controller::PerformAction, to be forwarded to these RPCs and used in
scripts.

This change also extends DirectActionProto to allow declaring arguments.
The information is forwarded to UserAction, so it'll be available when
building direct action declarations, but not otherwise interpreted by
the client.

This change also fixes a mistake in the tag number for
DirectAction.names; it's ok for it to be broken for a little while until
the server is modified.

Bug: b/134741508
Change-Id: Id405ad7364f2678bdb2cc37fe92b5b4c6796db49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1679695
Commit-Queue: Stephane Zermatten <szermatt@chromium.org>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675213}