[iOS] Add dark mode to recent tabs
Recent tabs appears twice, once in a standard appearance, and once
always dark in the tab grid. This has to be dealt with specially.

The CL also reverts to UITableView defaults in a lot of areas
(separator color, highlight color) to get dark mode support for free.

Bug: 976666
Change-Id: I762c88aa64e1cc766a2e2ca5338526283ed12563
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1685376
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675226}