[iOS] Only call NTP animation completion if all animations are finished
This CL changes how the completion block of the NTP opening animation is
called. In particular, if a new animation starts before the first one is
finished, the completion block of the first one isn't called.

self.foregroundTabWasAddedCompletionBlock isn't called when the first
animation finishes, as it is overridden at every animation.
For example, it is set to nil in line 4652. So if we duplicate this
code in the first animation completion, we would set the completion
block to nil when the first animation completes.
Also, the completion block of the last animation would be
executed when the first one finishes.

Bug: 978417
Change-Id: I368a0302e099fc0c4517fe5632140f95849b05b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1688915
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675263}