Update autofill automation_action.mm to be more EG2-compliant.
automation_action is a class that will live on the test side of the
EG2 test/app process divide, therefore it should not have any calls
to app helpers. In this CL:

* Instances of ExecuteJavaScript are changed to use ChromeEarlGrey's
version of CEG (which will send the JS to execute over GDO to the app
process).
* base::test::ios::WaitUntilConditionOrTimeout is changed to
GreyCondition.
* App-side headers that are no longer used are no longer imported.

Outstanding issue:
* tapOnTarget still relies on chrome_test_util::GetCurrentWebState()
to create GREYMatchers and execute GREYActions. I am not sure how to
refactor functionality of this method to properly split between
app and test side.

Bug: 922813
Change-Id: I565defe4e7b2f9b4419bcc967d2b1921e06cfadd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1688438
Commit-Queue: ericale <ericale@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675401}