RtcVideoDecoderAdapter: Store spatial layers info to media::DecoderBuffer::side_data
This CL is the first step to enables VP9 k-SVC decoding with HW decoder on
chromeos intel platform. The feature is currently disabled by default and can be
 enabled by --enable-features=Vp9kSVCHWDecoding. Webrtc stack sends hw decoder
 one SVC frame that is composed of multiple frames without superframe index. So
 that HW decoder can detect each frame, webrtc attach the data size info of each frame in
webrtc::EncodedImage. We propagate the info to HW decoder by storing it as
side_data in media::DecoderBuffer.

Design doc for this change is go/VP9-k-SVC-Decoing-VAAPI.

Bug: chromium:935411
Test: media_unittests
Test: Hangout Meet with VP9 SVC stream
Change-Id: I713de5bdbf1e7da85f531573ccf73791f2646099
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1688642
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675430}