Switching the release channel for users on secondary domains
Only users with _exactly_ the same domain as the enrollment domain was allowed to switch the release channel on the enrolled devices.
The corresponding pieces of code just performed the exact string comparison of the user's e-mail domain and the enrollment domain.
This made it impossible to switch the release channels when the secondary domains are used: e.g. the enrollment domain is example.org, and the user's domain is foo.example.org.

Now we allow all "affiliated" users to switch the release channel (including public managed users) provided corresponding policy allows that.

Bug: 674561
Change-Id: I64f9fdb0fe9a4907b0828724502a48425f2e592c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1689737
Reviewed-by: Ivan Šandrk <isandrk@chromium.org>
Reviewed-by: Hector Carmona <hcarmona@chromium.org>
Commit-Queue: Petr Smirnov <petrsmirnov@google.com>
Cr-Commit-Position: refs/heads/master@{#675561}