Update ownership of the RenderViewHostImpl.
The RenderViewHostImpl is kept alive as long as it exists one
RenderFrameHostImpl or one RenderFrameProxyHost depending on it.

The FrameTree maintains a map: SiteInstance →  RenderViewHostImpl for
reusing existing RenderViewHostImpl.

# Before this CL:

The refcount for the RenderViewHostImpl was handled by the FrameTree.
RenderFrameHostImpl and RenderFrameProxyHost had to manually call
functions on the FrameTree to update it.

# After this CL:

The refcount is handled by the RenderViewHostImpl itself, it becomes a
base::RefCounted. RenderFrameHostImpl and RenderFrameProxyHost are now
holding a base::scoped_refptr and the RenderViewHostImpl are
automatically deleted.

# Why this CL:

1) Less code. No more manual refcount. No more risk of leaking a
RenderViewHostImpl.

2) The FrameTree is no more in charge of "deleting" the
RenderViewHostImpl. This allows them to move freely. The
BackForwardCache will have to contain the set of used RenderViewHost
for a page. This CL is a small step in this direction.

Bug: 511340
Change-Id: I1afb30c919dc07cf8e693c28a34fa6f476b45a59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1665936
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675603}