Make SVGSMILElement::NotifyDependentIntervals() more explicit
The method now takes in an argument of which interval to look at.
This makes it clearer what piece of state is being used in
the method and it makes it easier to make future changes to
the code base, where dependencies are tangled.

This is the second step in a series of commits to be able
to separate the timing logic from the updating logic. This
aims to make it easier to make upcoming edits.

No changes should be made with this commit in how the
application functions or the runtime.

Bug: 981522, 641437
Change-Id: I6be50e1d6812d0238a127a42f582fbc44c82b265
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1692575
Commit-Queue: Edvard Thörnros <edvardt@opera.com>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Auto-Submit: Edvard Thörnros <edvardt@opera.com>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#675631}