👈 Disable failing junit tests on Touchless builds.
New Tab Page and Webapp Disclosure junit tests are failing on touchless
builds. For the Webapp Disclosure these tests are not relevant - we
would never show a disclosure on a touchless device. For the New Tab
Page tests, they should probably be updated to test the touchless
specific functionality, but I'll leave that to their team. For now the
tests are just being disabled.

Bug: 981871, 981870
Change-Id: I3c88fbd26cad3e67328f4c80c1101b9b7edc65fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1692164
Reviewed-by: Theresa <twellington@chromium.org>
Commit-Queue: Peter Conn <peconn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675659}