[LayoutNG] Do not use out-of-flow cache if in inline container
This patch changes |NGOutOfFlowLayoutPart::LayoutCandidate()|
not to use a cached result if it has an inline container.

|NGLayoutResult::OutOfFlowPositionedOffset()| has an offset
relative to the inline formatting context block. When an out-
of-flow object has inline contianer, this offset may change
without setting |NeedsLayout()| to the out-of-flow object.

Bug: 981686
Change-Id: Ib8376f791af0b95f6c2b6e5c2361b6a887c74e69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1692107
Commit-Queue: Emil A Eklund <eae@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675686}