Enable User Activation v2 as RuntimeEnabledFeatures.
The feature was enabled in M72 through content_feature kUserActivationV2, see
crrev.com/c/796938.  This means the RuntimeEnabledFeature gets enabled through
content in general but this is not true for few unit tests.  This CL fixes the
problem with RuntimeEnabledFeature and fixes the affected unit tests.

Bug: 696617
Change-Id: I50c3b22a5dfc6eb145ac00892bb3120b7c02e9ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1689038
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Mustaq Ahmed <mustaq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675791}