Use testdriver Action API in WPT pointerevent tests - Part 10
In order to run the web-platform-tests automatically, we will use
testdriver Action API in all the wpt to simulate inputs. Here we are
changing some pointerevent input test which taps on the elements inside
an inner frame.

This is the tenth CL that changes part of the pointerevent input
tests.

Bug: 606367
Change-Id: Ib90c92e8267f17fed74a86d70ca6d739b5ccedb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1638914
Commit-Queue: Lan Wei <lanwei@chromium.org>
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675891}