media: Allow for exporting a VA surface as a NativePixmap.
This CL creates VaapiWrapper::ExportVASurfaceAsNativePixmapDmaBuf()
which exports a libva-allocated surface as a DmaBuf in the form of a
NativePixmapDmaBuf.

This allows VaapiImageDecoder to expose the result of an image decode as
a DmaBuf so that the rest of the Chrome rendering pipeline can use it
without making copies.

For now, we only support IMC3 and NV12 surfaces, but there is no
fundamental limitation for supporting other formats later as needed.

A test is added to check that the result of exporting a YUV 4:2:0 JPEG
decode is correct. Note that we adapt LocalGpuMemoryBufferManager to be
able to convert the DmaBuf returned from the VAAPI (in the form of a
NativePixmap) to a GpuMemoryBuffer that can be mapped using minigbm.

Test: Some basic testing is added.
Bug: 868400
Change-Id: I3e5132ab1ff1a688d332ca19a641d146ffb4e7c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1648830
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675915}