cc: Make GpuImageDecodeCache request accelerated decodes when supported.
This CL adds functionality to GpuImageDecodeCache to request
hardware-accelerated image decodes when supported. The initial use case
is for Chrome OS + JPEGs using the VA-API [1]. The typical flow of
events for this path is expected to be as follows:

1) The "entry point" is GpuImageDecodeCache::CreateImageData(): here, we
decide if an image will be hardware decoded. This involves checks such
as whether the image is supported according to a list of profiles
supplied through GPUInfo.

2) In GpuImageDecodeCache::GetImageDecodeTaskAndRef(), we don't generate
a decode task for hardware accelerated decodes. Instead, we will wait
until the upload task runs in order to request the decode.

3) In GpuImageDecodeCache::UploadImageIfNecessary(), we request the
hardware decode. This ultimately involves sending a non-round-trip IPC
message to the GPU process, but this is handled by
RasterInterface::ScheduleImageDecode(). The decoded image never makes it
back to the renderer. Instead, we get a sync token that we can use to
"wait" on the decode to finish and we use the transfer cache entry ID to
refer to the decoded image.

Note that hardware accelerated decoding is currently supported only when
using OOP-R.

[1] https://tinyurl.com/accelerated-jpeg-decoding-v2

Bug: 868400
Change-Id: Ie57fe3a0cc30d039214a60a9da437b32f61a4ebd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1565568
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675949}