service worker: Remove "allow set registration" flag on clients.
This was needed in non-S13nServiceWorker, per the comment. The idea was
to prevent a client from being claimed before its URL had settled.

Before this CL, setting the controller registration was disallowed by
ServiceWorkerControlleeRequestHandler while it was finding the
controller service worker for a main resource request.  The flag was
checked in ClaimedByRegistration(). But that function is only called on
clients for which is_execution_ready() is true, which means the response
was already committed to the renderer so
ServiceWorkerControlleeRequestHandler cannot be operating on the
provider host.

Bug: 926114, 824858
Change-Id: I78fc13696fe72f9027da4a20d55e1f73dd44901f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1692235
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675950}