[LCP] Record's secondary ranking key is the insertion order
Currently, LargestTextPaint and LargestImagePaint uses node_id as the
secondary ranking key. This means when two records have the same size,
the node_id will be used to decide which one is the largest. However,
the order of node_id can be affected by factors out of LCP. So it's
potentially indeterministic. To make the result more deterministic, we
can use the insertion order as the secondary ranking key.

This CL reflects the above change.

Bug: 959984

Change-Id: Id42a32f05593f649f59fcb74fd3c57fa0e2f5fa1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1693036
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Reviewed-by: Steve Kobes <skobes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676057}