[NTP] Check timestamp for custom bg color calculation
The color calculated for a custom background should only
be updated if the image hasn't changed in the meantime.
Check this by saving the timestamp when a background is set
and checking that it still matches once the color calculation
is complete.

When possible use the thumbnail url |fetch_image| for the calculation,
the full-sized image was always being used.

Fix a copy-paste error in the definition for onlocalbackgroundselected.

Bug: 982215
Change-Id: Ib02f000add27f10eb23fb50051b44744a4b53323
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1693389
Reviewed-by: Gayane Petrosyan <gayane@chromium.org>
Commit-Queue: Kyle Milka <kmilka@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676069}