Reland: Add UMA for status codes of intercepted responses with null input streams.
Records the custom response status code for the intercepted requests where
input stream is null. The data is recorded only when the status codes and
reason phrases are actually valid. This is recorded regardless of whether
the network service is enabled or disabled.

The UMA is useful to see if there is any significant usage for this type
of interception. This type of interception has been originally mentioned
in crbug.com/965608, where it results in behavior different to the old
(non-network service) code path. Also see crbug.com/974273 for more
details.

Note on reland:
The original patch caused issues in dev/canary due to a side-effect of
AwWebResourceResponse::GetInputStream, which returns a unique_ptr and
which also closes the original java stream when destroyed. Current patch
avoids using unique_ptr, and instead introduces a new HasInputStream()
method, which does not have side-effects. Verified with the gmail app
(as in crbug.com/980223).

BUG=974273,965608

TBR=isherman@chromium.org

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1660657
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Tim Volodine <timvolodine@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#673362}
Change-Id: I3032c1b0485f2aa5a4495e7ba783dfda0f1fd737
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1689798
Cr-Commit-Position: refs/heads/master@{#676072}