[LayoutNG] Fix ShapingLineBreaker when last_safe < first_safe
This patch fixes |ShapingLineBreaker| to handle an edge case
to reshape the line end.

When reshaping the line end is needed, and the line overflows
due to the reshape, |ShapingLineBreaker| tries the previous
break opportunity until it fits.

It is possible that the opportunity is moved back more than
|first_safe|, in which caes |ShapingLineBreaker| calls
|HarfBuzzShaper| with |start > end|. This patch avoids such
case.

This is a speculative fix.

Bug: 981308
Change-Id: I3020ad79e925af5bd5d78bdf3202893117a9f279
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1694884
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Commit-Queue: Emil A Eklund <eae@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676077}