views: fix MenuClosureAnimationMac cancellation
The design intent of MenuClosureAnimationMac is that it should cancel
automatically on destruction. This does happen on the normal path,
since callbacks are posted either via base::OneShotTimer or
gfx::Animation, both of which cancel on destruction. However, on the
test path, a vanilla PostTask() is used, with an unretained receiver.
This means that this code is not safe:

  animation->Start();
  animation.reset();

when views::test::DisableMenuClosureAnimations() has been called.

This change makes MenuClosureAnimationMac use a WeakPtr to itself
as the receiver for the test animation completion callback, which gives
the desired cancellation behavior.

This change also makes views::test::WaitForMenuClosureAnimation()
more robust: now it uses Run+QuitClosure to exit its RunLoop rather
than RunUntilIdle.

Bug: 965601
Change-Id: Idb8a75449f79c44cfdb55a168e7c1bc9537785c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1687347
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676130}