Use AnimationDelegateViews for PaginationModel page transition
The page transition in ash::PaginationModel is used to provide
the animation effect on several views in ash, but currently
this is the default gfx::Animation which runs independently from
the compositor.

It would be plausible to run with the compositor by using
AnimationDelegateViews. The use of UnifiedSystemTralModel is
still TODO, so the use of the compositor is still optional.

Bug: 980722
Test: ash_unittests
Change-Id: I34a1cb2f8b179354a99bf5bcaad07249ca252fda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1692976
Commit-Queue: Jun Mukai <mukai@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676255}