Move RefreshTokenIsAvailable and GetAccounts to DeviceO2TS
This CL is a part of moving access token management to
OAuth2AccessTokenManager.

It is specifically a step toward folding
DeviceO2TSDelegate into DeviceO2TS now that the latter
is no longer an O2TS subclass.

It moves RefreshTokenIsAvailable and GetAccounts from
DeviceO2TSDelegate to DeviceO2TS not to refer to
|delegate_|.

Bug: 967598
Change-Id: I75faec6e9aa747dc908f9d9c2782dd01e426b847
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1692382
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Julie Jeongeun Kim <jkim@igalia.com>
Cr-Commit-Position: refs/heads/master@{#676317}