Move the refresh token encryption and decryption to DeviceO2TS
This CL is a part of moving access token management to
OAuth2AccessTokenManager.

It is specifically a step toward folding
DeviceO2TSDelegate into DeviceO2TS now that the latter
is no longer an O2TS subclass.

It moves the refresh token encryption and decryption
from DeviceO2TSDelegate to DeviceO2TS. The sequence
is that if the system salt is available, it gets the
encrypted refresh token from pref and decrypted it
using the system salt. Previously DeviceO2TSDelegate
had it. It's moved to DeviceO2TS with this CL.

Bug: 967598
Change-Id: I668704ce0ac655c64ddfe493c4ee5f0cc6fa1559
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1695281
Commit-Queue: Julie Jeongeun Kim <jkim@igalia.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676357}