Move the token validation from DeviceO2TSDelegate to DeviceO2TS
This CL is a part of moving access token management to
OAuth2AccessTokenManager.

It is specifically a step toward folding
DeviceO2TSDelegate into DeviceO2TS now that the latter
is no longer an O2TS subclass.

It moves the code for the token validation including
StartValidation() and RequestValidation(). DeviceO2TS
inherits from gaia::GaiaOAuthClient::Delegate and
implements methods. When StartValidation() is called,
it creates |gaia_oauth_client_| with |this| and gets
errors or responses.

Bug: 967598
Change-Id: I268e1a7f2909f77170279e178c95c354a45ca1a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1697126
Commit-Queue: Julie Jeongeun Kim <jkim@igalia.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676368}