Partially fix calc() length unit handling in media queries
When a calc() length is passed to media queries, the current code
assumes that there's always a DoubleValue(), and the unit type matches
TypeWithCalcResolved(), which is incorrect.

This patch partially fixes the issue when the calc expression can be
resolved without type conversion. Handling type conversion is more
complicated and hence left broken by this patch (as it's already broken).

Bug: 982542
Change-Id: I258b3d1aeec69fb65fbadd86981e97974239289b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1696191
Reviewed-by: Emil A Eklund <eae@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676452}