Remove the usage of Tab from ChromeActivityItemThumnailGenerator
The main problem was that a weak tab was provided to the existed method
to generate the thumbnail from Tab. This can't be done with webState as
by default WebState* is a weak ptr but it doesn't get invalidated after
the webState is destructed.
The solution is to change the thumbnailGenerator to a class instead of a
method. An object from this class will contain a webState and will
implement WebStateObserver, so it can invalidate the pointer when the
webState is destructed. Then change all users to use the object instead
of the block to generate the thumbnail image.

Change-Id: Ia918723c9190ce9d1c1977fe595a1f02bbc7ab5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1697051
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676490}