webauthn: show warning in console if userVerification option not set.
Several sites have not set a value for userVerification in either create
or get calls. This means that they end up with “preferred”, which is
generally not what they want.

This change adds a warning to the console in this case. The warning can
be silenced by explicitly setting userVerification to any valid value.

BUG=980972

Change-Id: Ide02de81f654c7101744bbc36d66b7c0dee00cda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1696170
Commit-Queue: Adam Langley <agl@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#676524}