[FractionalScroll] Fix DevTools web test
Turning on FractionalScrollOffsets allows scroll offsets to have
floating point precision. This means scroll into view on an element
might land mid-pixel. In the case of this test, the height of the
element meant that the scroll offset expectations all changed by 0.5.

This CL changes the height of the scrolled-into-view element by 1px to
avoid landing on a fractional pixel. This causes the test to pass when
the fractional scroll offset flag is turned on as well as keeping it
passing until we do.

Bug: 414283
Change-Id: I4239c8b424df7343cd671a760327093402826aeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1697304
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Andrey Kosyakov <caseq@chromium.org>
Auto-Submit: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676554}