[Web Payment] Verbose message when failed to download manifest.
Before this patch, a failure to download a payment manifest would be
printed in developer console, but would not be returned back to the
merchant. This made debugging production problems difficult, because
real world users have different configurations that can be hard to
predict and replicate in a local development environment.

This patch forwards the errors from the downloader to the merchant when
rejecting the promise returned from PaymentRequest.show(). If multiple
errors occur, only the first one is forwarded. If PaymentRequest.show()
is not being rejected, then the downloader error is still printed in the
developer console, but not forwarded to the merchant.

After this patch, the first failure to download a payment manifest is
forwarded to the merchant in the rejection of the promise returned
from PaymentRequest.show().

Bug: 977145
Change-Id: Ide9021293099a9c1488ac3d1bd2c96ce9afe2e46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1685429
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Sahel Sharify <sahel@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676602}