Signin: Move //components/signin/core/browser/webdata to public/
As part of finalizing the code structure of IdentityManager in the
codebase, this CL moves //components/signin/core/browser/webdata/
to //components/signin/public/webdata/. This change follows the
structure proposed here:

https://docs.google.com/document/d/1p_kahITGFgv7PehQ61-thRVR3Hrdaf3tLryMytTRPYs/edit#

This change was entirely scripted.

Using NOPRESUBMIT=true because an overly-aggressive check for the
presence of NULL detects its presence in a string in a moved file; I
don't want to change the string as part of this code move.

NOPRESUBMIT=true
TBR=eugenebut@chromium.org

Bug: 952788
Change-Id: Id86b0248f31ebfa8efddbbebbddb7f4fda933566
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1688848
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676902}