Skip extra thumbnail capturing when GTS animation is off
When the transition animation to Grid Tab Switcher (GTS) is disabled,
the thumbnail of the current tab is already updated when showing the
grid view. There is no need to force update it again after GTS has
finished showing. However, this is still needed when the animation is
enabled because the downsampled thumbnail is not written to cache.

This CL makes non-animation tests less flaky. Animation tests are
still flaky because the 2nd capturing might be skipped if the 1st one
takes too long.

Bug: 964406, 983448
Change-Id: Iabc8d360707fe78c3dea2b6e02c687c7d5b3bfca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1694929
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Commit-Queue: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676987}