Supersize now understand shortened resource paths
Supersize the tool that is used by the binary size trybot is now able to
deobfuscate the resource paths for APKs and AABs that enable resource
path shortening. This should allow the binary size trybot to blame
resource size diffs on the correct directories.

Bug: 753402
Change-Id: I396fd36b284a184bb5b07abf67d6b342920f0bc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1693172
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677082}