Add methods to blink::InterfaceRegistry for Pending{Associated}Receiver<T>
Similarly to what has been done already for blink::InterfaceProvider[1] and
service_manager::InterfaceProvider[2], in order to help with the transition
to the new Mojo types, let's add three more templated methods to the class
blink::InterfaceRegistry that enable calling the AddInterface() and the
AddAssociatedInterface methods passing a Pending{Associated}Receiver<T> as
parameter.

This will be useful in cases relying on the InterfaceRegistry's API to add
interface factories using the new Mojo types, as it's the case for instance
for the FindInPage class (see [3]).

[1]https://cs.chromium.org/chromium/src/third_party/blink/public/platform/interface_provider.h?l=28
[2]https://cs.chromium.org/chromium/src/services/service_manager/public/cpp/interface_provider.h?l=98
[3]https://cs.chromium.org/chromium/src/third_party/blink/renderer/core/frame/find_in_page.cc?l=55

Bug: 980151
Change-Id: I994d0d28c884f5500ac4f82a5592f16c824f8bff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1697486
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#677118}