SplitSettings: Move OS language settings to sub-page
The top-level settings row shows the current UI language and current
input method name, with a link to a details page.

The os-languages data model moves up a level so it can be used for
the section label.

The details page contains the list of languages and input methods,
which are always visible (we don't need to collapse them because we
have enough vertical screen space).

See bug for link to mock and for a screen shot.

Bug: 982299
Change-Id: I3548be4f420b22f0ab9314dbcc271b43240757d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1693094
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677145}