[threadpool] Do not run BEST_EFFORT tasks during startup (reland #2).
This is a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/1642732.
Since the revert, incorrect priorities were fixed at a few places:
- https://chromium-review.googlesource.com/c/chromium/src/+/1680760
- https://chromium-review.googlesource.com/c/chromium/src/+/1693184
- https://chromium-review.googlesource.com/c/chromium/src/+/1699372

With this CL, a task posted with:

base::PostTaskWithTraits(
    FROM_HERE,
    {base::TaskPriority::BEST_EFFORT, ...},
    base::BindOnce(...));

Will not be allowed to run before startup is complete.

The signal used to know that "startup is complete" is the same as
AfterStartupTaskUtils. It is not perfect. It will be improved in a
separate CL.

This CL aims to be the smallest change possible to allow seamless
migration of BrowserThread::PostAfterStartupTask() to
base::PostTaskWithTraits().

TBR=avi@chromium.org

Bug: 887407
Change-Id: Ie957af08849c4da1907d2d461965a1205f6a3dbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1699373
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677327}