[PM] Add the SiteDataCacheFacade class
This class serves as an interface between the SiteDataCache instances living
on the PM sequence and the UI thread. This is meant to be used by a
BrowserContextKeyedServiceFactory to manage the lifetime of the SiteDataCache
associated with user profiles.

It's a new component that doesn't exist in the current RC implementation of
the DB (the data caches currently live on the UI thread in the RC
implementation).


Bug: 961336
Change-Id: Ie799b30d27e169371105eea6a586cea6f00414e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1660410
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Commit-Queue: Sébastien Marchand <sebmarchand@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677334}