Remove NOTIFICATION_COOKIE_CHANGED_FOR_EXTENSIONS
It had only one notifier (ExtensionCookieNotifier) and one observer
(CookiesEventRouter). ExtensionCookieNotifier does nothing else; it need
not be a standalone class and is therefore merged into
CookiesEventRouter.

It is slightly complicated by the fact that CookiesEventRouter needs
to be notified of changes to both the original and OTR profile, and
thus has to listen to two CookieMonsters (over mojo). The
CookieChangeListener interface doesn't make it apparent which
CookieMonster a cookie came from, so an extra thunk class is introduced
to differentiate.

This also allows us to remove ExtensionSystem::InitForIncognitoProfile()
which is nice because it's one less direct dependency from
OffTheRecordProfileImpl to extensions code.

TBR=seantopping@chromium.org

Bug: 268984
Change-Id: I0fbbf82b6180c7a3ab30367ab41bdea2b431c2e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1699223
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677401}