oopd: Implement OutputSurfaceUnified.
Add a new OutputSurface type for the unified display on Chrome OS. The
unified display is a fake display that is the concatenation of multiple
physical displays. With OOP-D each physical display handles drawing so
the unified display can skip that work.

Add OutputSurfaceUnified that is basically a stub. It does as little
work as possible. It's also backed by a SoftwareRenderer to reduce
memory overhead, since we'll never actually use the renderer
implementation for this display.

Bug: 601869
Change-Id: I5184f502452035f40248cd54a355d6d42e0938ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1689955
Commit-Queue: kylechar <kylechar@chromium.org>
Reviewed-by: weiliangc <weiliangc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677439}