Add a GPU watchdog unit test GpuSwitchingToForegroundHang
This test covers OnBackgrounded() and OnForegrounded() functions and it catches
a hang after Chrome switching from the background to the foreground. The task
running time used for reproducing the hang is now set to 6 second long.

The running times of the other tests are also increased. In addition to the
expected timeout duration, at least extra 3 to 4 seconds are added. If the
schedule of any threads is delayed, the extra 3 to 4 second can prevent the
flakey test result.

Bug:949839, 983543

Change-Id: I9482205ddf392fa3c4d9d9db5490865361332eb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1700786
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677455}