Test HatsService Functionality and Remove hats_unittest.cc
This change provides browser tests for HatsService that better simulate
conditions found in production.

The test also revealed an issue with loading Hats only on an active window. This
change also modifies the behavior to show hats on the last active window instead
of requiring the last active window also to be currently active.

BUG=979530

Change-Id: Id47764af9e4b4d90c861b95e2382d771887eb99f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1698997
Commit-Queue: Robert Liao <robliao@chromium.org>
Reviewed-by: Wei Li <weili@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677511}