Remove is_main_frame and return value from ScriptCommandCallback
The fourth parameter |is_main_frame| in ScriptCommandCallback is not
necessary, callers can use WebFrame to check if the message is sent
from main frame. The Boolean return value of ScriptCommandCallback is
also not necessary, because it's only used by CRWWebController to print
error log which is not very helpful. Users of ScriptCommandCallback
should DCHECK and print error log in their own callbacks.

Bug: 956516,881816,881811
Change-Id: I789c7c0b3ad4e67a75ce03d42b770e30c5e6df64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1698463
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: John Wu <jzw@chromium.org>
Commit-Queue: Yi Su <mrsuyi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677667}