Unlocking for locked chall-resp user sessions
Implement unlocking from the Lock Screen for users who authenticate via
challenge-response protocol (e.g., via a smart card).

This adds functionality into the lock screen implementation that
asynchronously polls the list of currently available cryptographic keys
from all certificate providers (e.g., polls all smart card middleware
extensions to get the list of certificates from cards). If the key that
is needed for the user is present, public information about it is passed
to a CheckKeyEx request to cryptohomed.

The UI is not finalized yet, therefore the current implementation is
temporarily reusing the existing UI and the codepaths for unlocking by
entering a password / a quick unlock PIN. This will be fixed in
follow-ups.

Bug: 983103
Change-Id: I30084096455087a6b773754fe49b9178ae788e7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1689189
Reviewed-by: Steven Valdez <svaldez@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677771}