Open modern settings app for Manage Printers on Windows 10 RS1+.
On Windows 10 there's a new Manage/Add Printers UI in the modern
settings app that is much more user friendly than the legacy control
panel UI. On RS1+ there's a simple protocol launch that can be done to
redirect users to it. This change follows the model of
https://chromium-review.googlesource.com/c/chromium/src/+/1642818 and
uses the new protocol launch to give users a better experience.

Bug: 983687
Change-Id: I25d494e4acaacece18fe7a296b89ea645e7885ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1700331
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Cliff Smolinsky <cliffsmo@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#677974}