Ensure group contiguity when creating a tab in an existing group.
In this patch, TabStripModel::AddWebContents(), which is already
documented to alter the target index of a new tab in some cases, will
now move a new tab to enforce contiguity in a group if the requested
point of insertion would create a gap.

Forked from crrev.com/c/1703291, since dfried@ is on vacation.

Bug: 984112
Change-Id: I15224efc85c0565db7474cfd7dad7f0b4849cf9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1704935
Commit-Queue: Bret Sepulveda <bsep@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678017}