Update metrics for download foreground notifications.
Deleted metrics:
1. Android.DownloadManager.NotificationLaunch:
The relaunch happens for less than 0.0001%, this metric should be deleted,
and we can consider to delete the mitigation for the flicker issue as
well, since the occurrence is too low.

2. Android.DownloadManager.NotificationsCount.ForegroundDisabled:
This is related to the feature flag, the feature is shipped.

3. Android.DownloadManager.NotificationsCount.ForegroundEnabled:
Not actionable, we know users may have multiple download notifications.

Other metrics are kept and is treated as important stability metrics.


Bug: 975207,975228
Change-Id: I0f841c2e09aa475c8bd66da0ce4e8f0be12cd1ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1699096
Commit-Queue: Xing Liu <xingliu@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678020}