Reland "[MFill Android] Use password in PSL labels"
This is a reland of 04d2f0db8136441078db64421890839f13700873.
Diff to Patchset 1:
The change was crashing when tapping a field after it was filled
sometimes. The reason was that the second time a field was tapped,
there is an (currently invisible and unused) special item like
"Clear form". Since there was no label created for this, it ran
out of bounds and crashed.
This CL now reuses the previous (usually empty) label for these
items which prevents the crash.

Original change's description:
> [MFill Android] Use password in PSL labels
>
> By adding the password to PSL labels, we make more clear that we fill
> not only username but also passwords when tapping a chip.
>
> Bug: 982784
> Change-Id: Iea3fb6a537a47b16a32ea39c4ca4baa435885191
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1695269
> Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
> Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#676491}

Bug: 982784
Change-Id: I3504e792bc4051fa7626cac1c9a5adc4ffa77155
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1703406
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678242}