Views/Aura/ui: Replace the "always on top" bool with a z-order enum.
On the Mac, translate the z-order enum to the native window level
values. On other platforms, treat everything other than "normal"
as an undifferentiated "always on top" leaving open the possibility
of implementing full window level support on ChromeOS later.

There are six basic window levels. The "security sensitive" one
is unused in this CL but will be used in a follow-up CL to fix
a bug.

This is a relanding of e8151bacdcf51ba941b244d1ad99461670c0c452.
It was reverted due to incorrect handling of the z-order value
on Windows; this is fixed in this version.

BUG=865287, 869486

Change-Id: I41b277a3a67e13948e72da656d37a724c325fdbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1682667
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678339}