[Bluetooth] Create GattBatteryPoller
Class that gets the battery level of a connected Bluetooth device
through the standardized GATT Battery Service. Polling is done
periodically and updates the corresponding device::BluetoothDevice.

Instances for this class are expected to be created and owned by the
GattBatteryController class (to be implemented in a separate CL), for
each connected BLE device.

Design doc at go/cros-bt-battery.

Bug: 785758
Change-Id: Ic0de3e9b3d813537610378480cac6daaf8e25204
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1674716
Commit-Queue: Alberto Herrera <ahrfgb@google.com>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678369}