[OWNERS Tags] Inherit ancestor's tag if missing from dir.
This changes the component scraping to use the nearest ancestor's
component/team/os tag if the current directory does not define one.

Also, allow multiple teams mapping to the same component, and modify
presumbit to use the same criteria, warning instead of failing when such
situation arises.

Note that the 1:1 component to team mapping will continue to be present
in the mapping output, but it will show the team that is defined at the
topmost OWNERS file in the tree that is tagged with such component.

R=chanli,dpranke

Bug: 981713
Change-Id: Ie6f51f0beeb4ce63cedf302f4f15b2e2af25b451
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1693362
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678386}