[LayoutNG] Optimization - speculatively place floats.
This adds an optimization to speculatively place floats based on either:
 - The previous layout's "forced" BFC offset.
 - The previous layout's "optimistic" BFC offset.
 - The current layout's BFC block-offset.
(in that order).

This turns out to be an important optimization for sites which always
trigger a 2-pass layout when we can't determine the "exact" position
of them during the 1st pass (washingtonpost.com was one of these sites
for example).

If the "optimistic" BFC block-offset isn't set, this shouldn't change
the control-flow of the current layout algorithm.

This also saves some more stack space by removing a:
Vector<NGUnpositionedFloat, 1> from the stack.

Change-Id: Iae2ce8684e00ab166d810e875d7d0c420202a194
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1619047
Reviewed-by: Emil A Eklund <eae@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678392}