Add Feature and policy to control built-in cert verifier rollout
Add a base::Feature UseBuiltinCertVerifier to be able to control
built-in cert verifier rollout through experiments on Chrome OS.
Add a policy to fall back to the legacy certificate verifier.

Evaluate the feature and policy state in ProfileNetworkContextService
and request the built-in cert verifier accordingly.

Sign-in screen profile: Built-in cert verifier is active iff the
UseBuiltinCertVerifier feature is enabled.

User profile: Built-in cert verifier is active if
(*) the UseBuiltinCertVerifier feature is enabled and
(*) the UseLegacyCertificateVerifier policy is not set to true
(*) the RequireOnlineRevocationChecksForLocalAnchors policy is not
    set to true.

Bug: 939344, 978069
Test: browser_tests --gtest_filter=ProfileNetworkContextService*
Change-Id: I14d769ef30e9a6f0e65d11150b761f9b85a1f8f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1679689
Commit-Queue: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678474}