Record the number of third parties that read/write cookies on pages
What: Record the number of third party origins that read or write cookies
on a given page load.

How: TabSpecificContentSettings already had this information. So I've
rerouted the signal to to WebContents so that others can get it
too. TabSpecificContentSettings was already a WebContentsObserver so
this is a a small change for that class. I routed the cookie
read/write signals to PageLoadMetricsObservers as well and created a
new ThirdPartyMetricsObserver to host the new metrics.

Bug: 976866
Change-Id: I5917187c8152084c470b2bfe0d22dc48ba563ec9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1666303
Commit-Queue: Josh Karlin <jkarlin@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678495}