[mojo] Introduce "idle" interface state
This introduces a new ability for Remote endpoints to set an "idle"
handler callback to observe when the corresponding receiving endpoint
considers itself to be idle.

If set at all, a remote's idle handler must be set before any messages
are sent to the receiver. Setting an idle handler sends a control
message to the receiver to tell it to enable idle tracking. This has
several related consequences:

  - The Remote will expect every sent message to be acknowledged by
    the receiver using a new MessageAck control message.
  - The Receiver will in turn begin acknowledging every interface
    message it receives from the remote, using said control message.
  - The Receiver will set up a ConnectionGroup for itself and any other
    receivers dispatched in its messages. Any time the ConnectionGroup's
    ref-count goes to zero, a new NotifyIdle control message will be sent
    by the receiver back to the remote.
  - The Receiver will also retain a strong ConnectionGroup ref for every
    pending message reply on its own interface.
  - If a Remote receives a NotifyIdle message AND has no outstanding
    unacked messages in-flight, it will invoke its idle handler.

This will allow Content's ServiceProcessHost to watch primordial service
interfaces for an idle state and potentially tear down the corresponding
service process only once it is safe to do so. This is part of a series
of changes to eliminate Service Manager. See
https://docs.google.com/document/d/1wFKCupPftQUH4OAbLHHYcDJkjWX5a2fi9Kk7c3lnCs8/edit

Bug: 977637
Change-Id: I7a9b4b4d0d30d2eff4a7b6345daf95ea690c387d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1703305
Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#678702}