Add use counter for discard element
There is an animation element called "discard element" which
is not supported in Firefox and not really part of most peoples
daily internet usage. Removing it would simplify the code about
the SVG SMIL animations a lot and looks like a good idea seen from
and engineering point of view. This use counter is to check if it's
a good idea for the users as well.

There are actually two different UseCounters added, one which checks
for the existence of a "discard" element and one that checks for
if they are actually triggered. So breakage should be between those
two values.

The only thing added in this commit is two UseCounter, nothing else
should have changed.

Bug: 984947
Change-Id: I779fd2c05c20ac6912d7d9adb1e1ea65ab900bd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1708928
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Commit-Queue: Edvard Thörnros <edvardt@opera.com>
Auto-Submit: Edvard Thörnros <edvardt@opera.com>
Cr-Commit-Position: refs/heads/master@{#678707}