NaCl Browser: Call GetExtensionSystem on the UI thread
GetExtensionSystem uses a KeyedServiceFactory that must be called on the
UI thread (we have been calling in on the IO thread and blocking pool threads).
Previously the NaClBrowserDelegate was getting the extension info from
the profile directory on every call to MapUrlToLocalFilePath or IsNonSfiModeAllowed.
Instead, we now get the extension system on the UI thread once (after the
NaClHostMessageFilter is installed) and pass that into the NaClBrowserDelegate.

TEST= all existing tests pass with the inclusion of the DCHECKs in https://codereview.chromium.org/2775753002

Change-Id: I9dff9e41babdd80424d6a92cbaf3e89582dda3e1
Bug: 701326
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1708751
Auto-Submit: Derek Schuff <dschuff@chromium.org>
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Reviewed-by: Bill Budge <bbudge@chromium.org>
Commit-Queue: Derek Schuff <dschuff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678898}