Rename |vertical_separator| to |bottom_separator| in pdf/draw_utils.
Currently the draw_utils::Get*FillRect() functions refer to the
separator between two vertically adjacent pages as a vertical separator.
However, in pdf/pdf_engine.h it is referred to as a bottom separator.
This is a more specific name and cannot be confused with a horizontal
separator.

This CL renames |vertical_separator| to |bottom_separator| in the
draw_utils::Get*FillRect() functions and related constants in
draw_utils/coordinates_unittest.cc.

Change-Id: Ic088e4657e7056d86b9110d097e21d66186ff39b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1709636
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Jeremy Chinsen <chinsenj@google.com>
Cr-Commit-Position: refs/heads/master@{#678912}