Ignore I/O errors during Xbox gamepad shutdown
When a gamepad that supports vibration effects is removed and a
vibration effect is still playing, Chrome tries to issue a command
to stop vibration. If the device is already disconnected, the stop
vibration command will fail.

On Mac, Chrome implements an internal XInput driver for Xbox
controller support. This driver detects when an I/O operation
fails and signals the data fetcher to remove the gamepad. This can
cause a loop where Chrome tries to repeatedly issue the stop
vibration command.

To fix this, the XInput driver will suppress I/O errors when the
device is already shutting down. This CL also adds DCHECKs to
ensure the gamepad's Shutdown method is called exactly once.

BUG=982902

Change-Id: I960066f9b6da9e5135903cbd88d1447720eec9f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1709006
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Matt Reynolds <mattreynolds@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678943}