tcmalloc: Use the default TLS model on arm-gcc, always
Initially, the TLS model for the thread local heap pointer in
ThreadCache was only set to default for arm-gcc in ChromeOS in
https://codereview.chromium.org/2640263003

But this issue is corrupting PlatformThread CurrentId thread_local
storage in Linux, subsequently making PlatformThreadId calls unreliable.
So, instead of disabling it specifically for ChromeOS, let's disable it
for all GCC users.

Bug: 973767
Change-Id: Icf5908622ec128634f46412a682bdd6ea6b89cfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1699810
Reviewed-by: Primiano Tucci <primiano@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: José Dapena Paz <jose.dapena@lge.com>
Cr-Commit-Position: refs/heads/master@{#679063}