Wait for serial port closure in chrome.serial.disconnect()
With the introduction of a Mojo pipe between the chrome.serial API and
the SerialIoHandler it became easier for script to notice that the
serial port was not closed at an operating system level before the
callback passed to chrome.serial.disconnect() was executed.

This change ensures the file handle has been completely closed before
signaling completion.

Note that while this fixes issue 964683, r678396 is the better tactical
fix as it undoes the specific change that introduced the issue.

Bug: 964683
Change-Id: Ib5c65d3e341e7efd024a5d85cfd4900e822de433
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1705160
Reviewed-by: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679119}