[iOS] Fix omnibox legacy crash from inconsistent number of rows
From the UIKit documentation, -beginUpdates and -endUpdates should be
used when adding or removing cells. We don't do that, so the calls
aren't necessary.

It looks like the -beginUpdates and -endUpdates were somehow informing
the tableview/shortcuts collection view that it needed to change
height. With just the changes in legacy_base_view_controller, the
initial presentation of the shortcuts was the wrong height.
Invalidating the collection view layout fixed this.

Bug: 985736, 978004
Change-Id: Ia4aa2a01349a6e124c124fce599c5292e316aae2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1709323
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#679155}