[Mac] Create a new helper process variant for the GPU.
Swiftshader, the fallback software renderer, needs to perform JIT,
which is not allowed in the default helper variant. In order to expedite
a fix, do not use the allow-jit entitlement because the code has not
been updated yet to use MAP_JIT. Instead, grant the GPU helper variant
the allow-unsigned-executable-memory entitlement.

Bug: 977951, 850193, 985816
Change-Id: I1ea81949ab86dbf257f15c8883b69837dfb6028e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1709936
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679280}