[ios] Disables FallbackCoordinatorTestCases that fail on iOS 13 iPad.
On iOS 12, the keyboard could be docked/undocked by swiping on the
"dismiss" button, but that no longer works on iOS 13.  This CL disables
two tests that attempt to undock the keyboard and prevents tearDown from
calling the DockKeyboard() method when running on iOS 13.  DCHECKs are
added to DockKeyboard() and UndockKeyboard() to prevent them from being
called on iOS 13, since they no longer work as intended.

BUG=985977

Change-Id: I3346663ff3020239151aec5807fee4c68b4ab4f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1709696
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679382}