WebApp: Rename InstallOptions struct to ExternalInstallOptions.
ExternalInstallOptions is tightly coupled with PendingAppManager:
Many options like start_url are not used in InstallManager.

InstallManager and WebAppInstallTask should not work with
ExternalInstallOptions and ExternalInstallSource.

We will change InstallManager, WebAppInstallTask and InstallFinalizer
to work with WebappInstallSource and new InstallManager::InstallParams
struct.

Note that ExternalInstallSource is a strict subset of
WebappInstallSource.

InstallFinalizer::Source struct will be retired and replaced with
universal WebappInstallSource.

TBR=khorimoto@chromium.org

Bug: 860583
Change-Id: Ifee5ae36700cbf5026c22a66c8dba821795457fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1709315
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Commit-Queue: Alexey Baskakov <loyso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679440}