Don't open the editable combobox dropdown on click in the text field.
The current behavior creates problem for the context menu. In Chrome two
menus can't be open at the same time. Because the dropdown is almost all the time open
the user can't access the context menu for the text field.
Even if the dropdown is closed the following happens:
- user makes a right click in the text field.
- EditableCombobox::HandleMouseEvent is called.
- it opens the dropdown.
- the textfield's menu is cancelled.

Bug: 972069
Change-Id: Ib77cae34355049f4b0dfe287422d58c62cd00a09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1702250
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679518}