[omnibox] Create mechanism for joining matches and submatches
We have some new match types which pair with other matches and must
follow them in the list of suggestions (for aesthetic reasons). Keeping
them together, and pairing them with the best matching suggestion is
non-trivial. This CL adds a secondary sorting field to force them to
follow their origin.

It's first submatch type is pedals, which required some tweaks to
become consistent with the way the other matches are handled.

Change-Id: I2b924dcf4c3ee586d4a7fa6ba24c0f27d5a3012f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1592636
Commit-Queue: Kevin Bailey <krb@chromium.org>
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679569}