Navi: Only allow chrome://welcome if navi onboariding is enabled.
We are currently only showing the old welcome pages if someone manually
navigates to chrome://welcome and onboarding isn't enabled. This is a
dead top level chrome domain in that case because users won't get ANY
onboarding if navi isn't enabled. This CL makes sure it's no longer
accessible so we can continue to delete the code to the former welcome
page.

Bug: 956680
Change-Id: If6e1876e46b8daac40bdf8610d37ee9c78e6c8c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1711147
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Commit-Queue: Hector Carmona <hcarmona@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679641}