Rename CupsPrintJob::State::Error
Renames CupsPrintJob::State::STATE_ERROR to State::STATE_FAILED to
make it clear that the print job is complete and not recoverable. This
is so that once we have some notifications that convey errors that are
recoverable (eg "your printer is out of paper") we can differentiate.


Bug: 985020
Change-Id: I550c3aa8c22f4324c71b1b04ab00f0d0296273de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1707232
Commit-Queue: Bailey Berro <baileyberro@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679734}