[SmartLock] Do not build regular service in off-the-record context.
Prevent EasyUnlockServiceRegular from being built with an incognito
context (AKA "off-the-record"). My current working theory to explain 969135
is that the same EasyUnlockServiceRegular instance is being handed to
two separate Profiles (the on and off the record versions) with assumptions
of unique ownership, and when Shutdown occurs, they step on each others'
toes.

I unfortunately cannot repro this hypothetical situation of
EasyUnlockServiceRegular being handed to an OTR Profile, but in any case
this change makes explicit our desire that this never happen.

Also add temporary CHECK to find source of crash in service: CHECK to
ensure that EasyUnlockService is shutdown before it's deleted. This
is added because the crash stack trace in 969135 implies that a
deleted member variable of EasyUnlockService is being touched during
EasyUnlockService::Shutdown().

If my theory above is correct, this temporary CHECK will not trigger, and
the original crash of 969135 will no longer be reported.

Bug: 969135
Change-Id: Iaab415efdf47ac449539c2f91693e8002d99526d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1705014
Commit-Queue: Ryan Hansberry <hansberry@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: James Hawkins <jhawkins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679775}