cc: Ignore frames from different source-ids.
The source-id for BeginFrameArgs can change during a sequence (e.g. if
the gpu-process restarts). In such cases, ignore the frames from the
new source, so that the metrics aren't computed incorrectly.

Instead of ignoring the new source, it would be possible to update the
trackers to be able to process the begin-frames from new source. But
this is a rare event, and would introduce complexity in the code. So
instead of processing them, they are ignored.

This removes a couple of DCHECK()s previously introduced in
crrev.com/c/1710754, and adds a couple more DCHECK()s to catch issues
like this earlier in the pipeline.

BUG=985847, 985359

Change-Id: Id717e7a19b7099cad8199148160f84a84f8b9ecb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1711052
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679776}