assistant oobe: expand clickable area of show/hide buttons.
This change enlarges the clickable area of show/hide buttons on the
assistant third party oobe page, from a small rectangle area around
the buttons to the entire row, in order to: 1) be consistent with
the performance of other buttons in the same style, e.g. one in the
Languages settings page. 2) have a better user experience.

Bug: 981830
Test: manually.
Change-Id: I3f8228f8dd19d5491de62d1dc9b6b20fef290565
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1694089
Reviewed-by: Yue Li <updowndota@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Meilin Wang <meilinw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679777}