crostini: refactor export import notification
This change refactors the export import notification, such that all
string handling is internal, and such that there is a more defined API
for setting the many status variants.

This change facilitates two incoming changes: one that adds the status
CANCELLING and CANCELLED to the notification, and another that adds another
variant on the failed status to do with concurrent operations.

BUG=chromium:978318

Change-Id: Icc101378ce2f6650f1dc646d06af868684bd56b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1702346
Commit-Queue: Julian Watson <juwa@google.com>
Auto-Submit: Julian Watson <juwa@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679821}