Support signed exchange interceptors on UI thread
We are in the process of migrating NavigationURLLoaderImpl to the UI
thread. This CL allows the SignedExchangeRequestHandler and
PrefetchedNavigationLoaderInterceptor to be used on the UI thread
when the NavigationLoaderOnUI feature is enabled.

The main difficulty here was dealing with reading blobs, since blobs
still live on the IO thread. Those changes are mostly concentrated in
CrossOriginReadBlockingChecker and
PrefetchedSignedExchangeCacheAdapter.

Bug: 824854
Change-Id: Iaed622313bfb7a8d00c34a2e18382d7c3c5b7457
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1707249
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679870}