Enable VR DFM conditional install for DD-ready devices
The CL uses the new 'conditional modules' feature to auto-install the
VR DFM upon updating Chrome on supported devices. This should provide a
much better experience when launching Chrome from the DDH for the first
time on devices where the cost of having the VR DFM unused is minimal.

This is only used on monochrome_bundle builds.

Bug: 961796
Change-Id: I79b095205b5dd947484738034bf50a2a0c4c8ca8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1714144
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Commit-Queue: Will Cassella <cassew@google.com>
Cr-Commit-Position: refs/heads/master@{#680060}