x11: Make PlatformEvent-related methods optional in XEventDispatcher interface
XEventDispatcher implementation *can* optionally be associated to a
PlatformEventDispatcher, however in most cases it is not the case and 3 methods
must be overridden with a empty/stub implementation. This CL changes
XEventDispatcher, adding default empty implementation for PlatformEvent-related
functions, making it easier and more intuitive implementing it.

Bug: 789065
Change-Id: Icdb501133817038a536b108f8a0d8556ff3389f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1710242
Commit-Queue: Nick Yamane <nickdiego@igalia.com>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680107}