Configure the insecure and secure parts of the async resolver separately.
We can now set the secure dns mode to AUTOMATIC and turn the insecure part of
the resolver off, which will result in the async resolver issuing DoH requests
and falling back to the system resolver without attempting to send plaintext
DNS requests itself.

In a follow up cl, the secure dns mode part of the configuration will be read from
prefs. For now, we continue to set the mode to AUTOMATIC if there are DoH servers
specified and to OFF otherwise.

Bug: 985589
Change-Id: Ia40cbae1fdd57e19e925e125dcf1e82289220242
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1709928
Commit-Queue: Katharine Daly <dalyk@google.com>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Eric Orth <ericorth@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680161}