Add helper function GetInsetSizes() to PDFiumEngine.
Currently PDFiumEngine::EnlargePage() and PDFiumEngine::InsetPage()
have common logic for retrieving the inset sizes given the current
layout and position of the page in the document.

This CL adds PDFiumEngine::GetInsetSizes() which extracts the logic for
retrieving insets sizes and puts it into a helper function. This helps
manage the complexity of PDFiumEngine.

Change-Id: I2aabba1f08770dbc16e033c64607627675ff0b58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1715904
Commit-Queue: Jeremy Chinsen <chinsenj@google.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680188}