Add PalmDetectionFilter and 2 initial implemenations.
This adds a palm detection filter that interacts with the concept of
"held" items. The filters will be used in TouchEventConverterEvdev to
either hold or cancel touches, in addition to the behavior from device
firmware.

The two implementations are:
1. An always "open" detection, that doesn't mark anything to hold/cancel.
2. A simple heuristic that is based on interaction with stylus touches.

Both implementations have unit tests that cover 100% of the code.


Bug: 982118
Change-Id: I9f50b98be434ba5c767d9392c0a470262cf5e603
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1708148
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Rob Schonberger <robsc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680248}