Update BLE authenticator display name when pairing mode changes
Currently if an authenticator is being added to the BLE device chooser
dialog but it is not yet in pairing mode, it will use a generic display
name. This CL causes the display name to be updated to the correct
device name when the device enters pairing mode.

Bug: 906750
Change-Id: Ic86abf9a055746634b64b5e9043fa37640076d14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1708917
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680278}