Wait when sending first synthetic gesture.
Tests can often start sending synthetic events before the renderer has
fully initialized and produced a compositor frame. Before that, the
browser doesn't yet get the geometry information from the renderer and
the gestures  This leads to flaky, tough-to-debug, tests.

This CL makes the first gesture sent by the SyntheticGestureController
wait until the renderer confirms it has produced and presented a
CompositorFrame. By that time, it should be ready to receive events.

This change should make the synthetic gesture API more intuitive and
easier to use. It should hopefully also fix several flakes in addition
to the one linked in this CL.

The fix in scrollbar-button-gesture-target.html is required because the
autoscroll timer begins 250ms after the mouse down. Prior to this CL,
this would pass by accident because the first generated compositor frame
in web tests appears to take a long time (~200ms). Since that would
occur after the MouseDown this test would pass. Now that we force a
frame before the MouseDown the MouseUp happens in 100ms, before the
autoscroll timer expires. We fix this by making the delay exceed the
autoscroll timer delay.

Bug: 897520,951638,985374
Change-Id: Idd63e1f2b999253e0392ae37be4180956119a814
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1566750
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680301}