[Dolphin] Add parameterised histograms and more metrics.
Since we have more than one Dolphin model running, it would be useful to
be able to separate out our metrics per-model. This CL adds
parameterised histograms to do that.

At the same time, it increases the number of things we report metrics
about. In particular, we now keep counts of how often each function of
the RecurrenceRanker is used.

Bug: 921444
Change-Id: I100830883158c071bbf299efc204425b802f0f5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1713149
Commit-Queue: Tony Yeoman <tby@chromium.org>
Reviewed-by: Charles . <charleszhao@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680331}