[css-properties-values-api] Accept any value through setProperty.
According to a recent spec change, the syntax of a registered custom
property must be ignored until computed-value time. This means that
setProperty may not inspect the PropertyRegistry and reject values
based on that.

This CL also removes the PropertyRegistry parameter in many places, to
avoid incorrectly using it in the future.

Bug: 641877
Change-Id: I99c58775e800500862378420118d81fe3186e739
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1714590
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680388}