viz: Create Separate Type for OutputSurface as Overlay
When using a BufferQueue, the OutputSurface is treated as an Overlay,
and is made into an OverlayCandidate. That overlay candidate is then
pushed in as the front of the candidate list. This CL changes it into
its own type instead of checking for a bool on OverlayCandidate.

R=rjkroege

Bug: 963597
Change-Id: I05588ba4d3b1ac03ed01ac0fff25d9ae03fc28c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1696168
Commit-Queue: weiliangc <weiliangc@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680431}