Rename postMessageToFrame to postMessageToMainFrame
WebContents#postMessageToFrame is only used to post message to main
frame, all the call sites passed the |frameName| = null. The internal
implementation of native MessagePortProvider::PostMessageToFrame is not
for all frames either.

Renamed postMessageToFrame to postMessageToMainFrame to make that
explicit and get rid of the |frameName| parameter.

Bug: None.
Change-Id: Ib31bf02ad131bc91d10b8936c51e54a03fdf9ea6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1711053
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Commit-Queue: Shimi Zhang <ctzsm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680521}