Re-write test pointerevent_pointermove_in_pointerlock.html
The test was failing in wpt because it passing non-integer value to
test driver. This CL changes it to use "origin" instead.
This CL rewrites the test to makes it cleaner, instead of adding event
listener inside event listener.
This CL also changes the test to send the "move" events after receiving
pointerlockchange.

Bug: 986277
Change-Id: I0f51523c3c90791339877b2e51afa8e9a4cc9d4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1717195
Reviewed-by: Lan Wei <lanwei@chromium.org>
Commit-Queue: Ella Ge <eirage@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680563}