Remove conflict detection between test expectations which contain globs
This validation check should not have been added. According to dpranke@,
expectations with globs should be allowed to exist with
other expectations which have longer globs or have the exact test
name. This change also makes it easier to add test expectations with globs
to a test expectations file.

For example if we have the expectation below.

[ win ] test_suite_one.test_abc [ RetryOnFailure ]

and we wanted disable all tests inside the test_suite_one on windows bots
except for test_abc then we would have to add an expectation for each test
in test_suite_one. If we remove this validation then we can add the following
expectation.

[ win ] test_suite_one.* [ Skip ]

Bug: chromium:973936
Bug: chromium:976505
Change-Id: I58f93f2c170f4c97bbed38b04bf6b534abbb2cd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1711895
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680649}