Use a dictionary pref for UMA demographics
This CL switches from individual syncable int priority prefs for the
birth year and gender to a syncable dictionary priority pref (which,
in turn, will contain the birth year and gender values.

This CL also separates the handling of the syncable pref from that of
the (only locally stored) birth year offset in order to retain the offset.
If the user resumes syncing, causing these prefs to be recreated, we don't
want them to start reporting a different randomized birth year as this
could narrow down or even reveal their true birth year.

Bug: 983303
Change-Id: Iada3714d315320c19881b51d1446623dc40c3b1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1715468
Commit-Queue: Roger McFarlane <rogerm@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680862}