Don't send pixel-based display info for MacOS Hosts
This change is a temporary solution while we investigate and fix
the problems being reported by MacOS users around mouse cursor
offset issues and problems switching displays with multi-mon.

With this change, users will receive the same experience they had
with the Chrome app (and still have for iOS and Android clients),
meaning they won't be able to switch monitors but the issues above
will no longer occur.

I would like to get this change into the M77 branch so we can
provide this build to users who are blocked by the current issues
while we continue to investigate / fix the underlying problems.

Once we have the fix checked in, we can remove this workaround
and merge it into M77 along with the fixes for the user issues
reported.

Bug: 987513
Change-Id: I2d3553fcc4ab018d8bb9edc067be3a38bf6ab446
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1717756
Auto-Submit: Joe Downing <joedow@chromium.org>
Commit-Queue: Gary Kacmarcik <garykac@chromium.org>
Reviewed-by: Gary Kacmarcik <garykac@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680890}