[omnibox] Optimize GURL comparison for tab look-up
It has been observed that a single char press in the Omnibox can trigger
several memory allocations per open tab. This is partly due to the fact
that, for each Omnibox suggestion, we compare that URL against each
tab's URL to see if it is a tab switch candidate. When we compare 2
URLs, we convert both of them to their stripped version prior to the
comparison. However, since one of the URLs in these comparisons - the
Omnibox suggestions's - is not changing over the scan of the tabs, we
could convert it outside the loop, and only convert the candidate tab's
URL.

This CL adds a new URL comparison member function for this search, that
takes one URL to be stripped, and one already stripped. Using this
function instead avoids half the allocations for N tab comparisons.

Bug: 986597
Change-Id: Ib65e223087c0df4a030508d026e089866b754c41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1717911
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Kevin Bailey <krb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680998}