Reland "Plumb through parameters for integrating WebAPKs with adaptive icons"
This is the first part of integrating adaptive icon with WebAPKs.
This CL plumbs through the is_primary_icon_maskable parameter
for crafting create/update requests to our WebAPK minting server.

This CL also removes the requirement for a web app manifest to
have an "any" purposed icon to be eligible for being added as a
WebAPK, if adaptive icons are supported.

For adaptive icon to be a fully functioning feature, change in
shell apk, and their corresponding change in chromium is needed.
These changes are left for further commits.

Also, fix a lint error in WebappSplashDelegate.java

Bug: 977173
Change-Id: I85c30c275212d6aefd4c3823fd128bb7c8ad9cd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1712515
Commit-Queue: Henry Jian <hzjian@google.com>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681003}