[Flaky] Disable WebRtc GetDisplayMediaVideoAndAudio test on Linux
The flakes are common and we have seen it on Windows and ChromeOS
already. Interestingly, it times out often on Linux Tests (dbg) [1] and
causes several retries ASan but still works on msan and rel builds. I'll
disable it Linux to calm the bots down.

[1] https://test-results.appspot.com/dashboards/flakiness_dashboard.html#tests=WebRtcGetDisplayMediaBrowserTestWithPicker.GetDisplayMediaVideoAndAudio

TBR=guidou@chromium.org

Bug: 986033
Change-Id: I6dc8c8febdeede7c394475c731cf10f56a275540
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1718950
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681250}