Use of layer animation for page switch in app_list/contents_view
The contents_view is responsible for the transition between the apps
grid and the search results, and the transition is controlled as
a slide animation with bunch of bounds changes.

I believe that this can be done through animations without involving
views repainting and layouts, this CL achieves as follows;
- HorizontalPageContainer (containing apps grid); bounds change
  with ui::LayerAnimator
- SearchResultPage; set-bounds + clip-rect and rounded corners
  animation. Shadow changing with gfx::Transform.
- ExpandArrow; opacity change animation with ui::LayerAnimator
- SearchBox; transform change with a SlideAnimation

This CL also changes on PaginationModelObserver on the timing when
the page transition callbacks. Previously OnTransitionStarted()
was called right before the transition actually started (i.e.
PaginationModel::has_transition() returns false), so this CL renames
it to OnTransitionStarting() and adds a new OnTransitionStarted()
method.

This CL has significant improvement on test cases of
LauncherAnimations/LauncherFullscreenSearchTest.SearchAfterFullscreen:
  ~35-60% -> 65-80%
Also LauncherHalfTest:
  ~40-60% -> 65-90%

Bug: 980367
Test: interactive_ui_tests
Change-Id: I935114f2b973bd54babbee9568c2572bf4d2bbd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1719271
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681543}