Re-add gdi32 delayload and fix pdf font rendering
In order to get font data in the sandboxed process, InitializePDF()
creates a patch for GetFontData() which instead proxies the request
using Mojo to the browser process which has access to this method.
When GDI is delayloaded, calls into real GDI after these patches are
created result in the delayload handler overwriting the patches when
it resolves the delayed imports with the full import information. To
workaround this, we first call into GDI so that the delayed imports
get resolved, THEN we apply the patches.

On Win10 this isn't necessary because these methods are not used.

Note that this shouldn't be necessary on Win8 either, but it is due
to a current issue with the win32k process mitigations. That issue is
being tracked via crbug.com/988169.

Bug: 980577
Change-Id: I3d9726194a6f16e2f94cf794b441ae8d33cff4ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1718777
Commit-Queue: Cliff Smolinsky <cliffsmo@microsoft.com>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681597}