media/gpu/v4l2: make PrepareQueueBuffer() take const argument
V4L2WritableBufferRef::PrepareQueueBuffer() does not take ownership of
its argument, not does it modify it. So change it to take a const
reference instead.

On top of being semantically correct, this change also makes it easier
to port the V4L2SVDA to V4L2Queue.

Bug: 987852
Test: video_decode_accelerator_unittest passes on Minnie.
Test: video_decode_accelerator_tests passes on Minnie.
Change-Id: I85e17448535db91a42c52a0fd4e7bb1de2ff1efc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1720232
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681702}