heap: Replacing ephemeron_callbacks_ with a worklist.
I added a worklist for registering ephemeron callbacks.
ThreadHeap previously used a HashMap for that. The HashMap remains since it
is also used to verify callbacks are consistent.
MarkingVisitor adds entries to the worklist instead of the HashMap. ThreadHeap
then reads from the worklist to the HashMap, invoking callbacks as it goes.

Bug: 986235
Change-Id: Ie716281ba216a083146bff7a78abe06621befcba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1718371
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681733}