Navigation: Move |reload_type_| from NavigationHandleImpl to NavigationRequest.
This includes removing |pending_nav_entry_id_| from NavigationHandleImpl
since it is always equal to the |nav_entry_id_| in NavigationRequest,
and neither of them is modified after the creation of the handle.

Removing the logic for initializing |reload_type_| and |restore_type_|
in NavigationHandleImpl.

This does not change the current behavior, since if the
|pending_nav_entry_id_| was not 0, this means the |nav_entry_id_| is not
0, and we can simply initialize both |reload_type_| and |restore_type_|
in NavigationRequest.

Bug: 916537
Change-Id: Ia51aa8d64021e37d81ea20f6328b21315af6be9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1715081
Commit-Queue: Mohamed Abdelhalim <zetamoo@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Lowell Manners <lowell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681779}