Prevent setting attributes on or appending children to text nodes
The LogBuffer distinguishes between two types of node in a tree that is
ultimately converted to a DOM: Elements (think of this as HTML tags) and Text
Nodes. Before this CL it was possible to do the following:

Log() << "Foo" << "Bar";

This would create a text node with content "Foo" and give it a child "Bar". As
HTML does not allow a text node to have children, "Bar" would never appear
in chrome://autofill-internals.

The correct code in this case would be:

Log() << Tag{"div"} << "Foo" << "Bar";

To make the code clearer, the term "node" in the code has been replaced with
"element": In HTML, Nodes are the super set of Element (tags) and Text Nodes.

At some point, we could support HTML Fragments, which would enable use to
support the Log() << "Foo" << "Bar" syntax without requiring to explicitly add
a Tag{"div"} before the "Foo".

Bug: 928595
Change-Id: I778a025c853dc0a39f41675d57a381ba11c13975
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1722534
Reviewed-by: Maxim Kolosovskiy <kolos@chromium.org>
Commit-Queue: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681832}