Add progress bar to CrOS Lock Screen Media Controls
This CL creates a MediaControlsProgressView class and adds an instance
to the CrOS lock screen media controls UI.

LockScreenMediaControlsView uses MediaSessionPositionChanged to pass
the playback/position info to MediaControlsProgressView. The progress
view then continually updates the progress bar/time as time progresses.

The placement of this view on the UI isn't critical right now as it
will be changing in the near future.

Bug: 965573
Change-Id: I2da5e864130ec371a58c623006ab7773f9bbe09c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1721321
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Mia Bergeron <miaber@google.com>
Cr-Commit-Position: refs/heads/master@{#681851}