[Media Session] Add color type to bitmap
Add a color type to MediaImageBitmap. The default color
type (N32) is platform specific and is an alias for either
RGBA_8888 or BGRA_8888. When deserializing we will convert
all images to the platform default. The color type tells
us the original color type of the image which may be
different if it came from ARC or Chrome.

ARC only supports RGBA_8888 and images from ARC will
fallback to RGBA_8888 which is the default value.

BUG=987814

Change-Id: I9fed118c85e29a24c232bc3adc52cdf64988bd3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1717168
Commit-Queue: Becca Hughes <beccahughes@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681878}