Dump distillability info when DCHECK is on
When DCHECK is enabled, the distillability information is dumped to
the console in developer tools for easier debugging and access by
by the modeling scripts. The message is in verbose level, so it's not
visible by default.

This CL is reincarnated from:
https://codereview.chromium.org/1972503002/.

On desktop, add command line flag --enable-distillability-service to
enable this code path.

Bug: 610944
Change-Id: Id7b0a5d0a4d2d6313474171db70ea79dbbb50cf1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1717443
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681974}