Remove test util class BrowserAddedObserver
It used NOTIFICATION_BROWSER_OPENED.

Most of the time in these tests, the new browser is opened
with chrome::NewWindow or chrome::NewEmptyWindow, which
are usually synchronous, so waiting is not needed. The
exception is when they trigger session restore, and in
those cases it's sufficient to use
SessionRestoreTestHelper::Wait.

Bug: 268984
Change-Id: Ia1de99068a9c6828dbb653b7cedcd7bda31a4f0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1706952
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682047}