Do not reset searchbox state when closed
Currently state changes are ignored when the app-list is closed,
so I think it's better not to clear the search state there.
Instead this allows to clear the state on showing (ResetForShow).

This works fine with tablet mode so I'm leaving some of the
existing conditions as-is.

Ideally state change should be accepted after the closing animation
finishes -- that wouldn't affect the rendering and animation
smoothness. Modifying that requires larger changes, but this CL
wants to be small for merging to M77 branch.

Bug: 987584
Test: manually
Change-Id: Ie577941b662e68768b2ef03490eb6eb7324332d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1725190
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682092}