desktop-pwas: OpenApplication using Launch Service
The Launch Service can now be used to open a new application
window.

The numerous call sites of ::OpenApplication are being migrated to
use the Launch Service in subsequent CL
https://chromium-review.googlesource.com/c/chromium/src/+/1723794

Still to be implemented:
When the BMO flag DesktopPWAsWithoutExtensions is enabled,
we currently do not notify SiteEngagementService or
AppBannerSettingsHelper about web application launches.


Bug: 966288, 966290
Change-Id: Ia122eaa67561e3fa9b6373d4ad9e8692eb2e4f5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1722953
Commit-Queue: Eric Willigers <ericwilligers@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682186}