Introduce a GUID to BookmarkNode
BookmarkNodes currently have no GUID or ID consistent across different
clients and that remains stable throughout the lifetime of the bookmark.
This CL thus introduces such a GUID. This will allow for the insertion
of logic geared towards avoiding bookmark duplicates when synced across
different devices and also potentially substitute other identifier
fields later on. GUIDs are randomly generated for each added bookmark,
with the exception of BookmarkPermanentNodes, for whom hard-coded GUIDs
are used instead.

Objectively, this CL does the following:
1. Introduces an immutable GUID to the BookmarkNode and a corresponding
key to the persisted JSON object, in order to allow for
encoding/decoding.
2. Changes BookmarkNode constructors to always take in an ID and GUID to
be set to the newly created node.
3. Changes the BookmarkPermanentNode constructor to assign hard-coded
GUIDs to PermanentNodes (stored in a map) based on their Type.

Bug: 978430
Change-Id: I77909f33e24b454f026d5566f64cb731a08f2b68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1676167
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Pauline Leitao <psivieroleitao@google.com>
Cr-Commit-Position: refs/heads/master@{#682216}