[PM] Adding a working set trimmer policy.
As discussed this CL will add a WorkingSetTrimmer policy. The default
policy as implemented in working_set_trimmer_policy is identical to the
previous behavior on Windows in which we will always trim on frozen.

The Windows implementation uses that same default behavior but uses
PlatformSupportsWorkingSetTrim to check if it's flag enabled.

The ChromeOS implementation uses the same default plus it adds working
set trim on Moderate memory pressure. There are a number of flags used
to control this behavior.

Bug: 973963
Change-Id: I2dfd4a8e7e45497725cb905bc1254beb8fda3ae3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1706450
Commit-Queue: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682475}