IndexedDB: Refactor AddCursorBinding into CreateCursorBinding
This relands https://crrev.com/1725318 with a fix for the crash
that can happen under certain circumstances.  The error in that
CL was that dispatch_context() shouldn't be used outside of the
context of a Mojo connection.  Here's the updated commit message:

As part of cleanup from CL https://crrev.com/c/1626660, this CL
moves creation of CursorImpl into IndexedDBDispatcherHost by
refactoring AddCursorBinding into CreateCursorBinding.

As a result, the OpenCursor calling path no longer needs to pass
the sequenced task runner into the OpenCursorOperation task and
ownership of CursorImpl and the mojo binding steps is made
simpler.

Bug: 988868
Change-Id: I8444f53b0c4f362d3152cb9bea232d65f99375a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1728540
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Chase Phillips <cmp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682851}