[Fuchsia] Fix CleansUpContextJobs test to use its own job.
The test is supposed to create a new sub-job of its parent job, so that
it can watch for the job becoming empty. Instead it was just duplicating
the handle to its existing job, causing the test to flake if any other
test batch process happened to be running in that job at the time.

Bug: 989085
Change-Id: I17a287dcb6e5b0a7351294f6297ed355eb3e15ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1728440
Auto-Submit: Wez <wez@chromium.org>
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682871}