Modify PDF accessibility interfaces to send data for links and images
This change modifies the existing PDF interfaces to send data for links
and images. This extends the existing pipeline to accommodate links and
images which is part of the larger effort to make links and images
accessible in PDF. This extends the messaging pipeline between plugin
process and mimehandler process to send links and images.

This doesn't have any changes pertaining to actual data being populated
in the structures. In future CL's the data will be populated in the
structures for links and images.

Bug: 981448
Change-Id: Iee0b1ffae36df57772a739937c240aec5e01dd02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1703627
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Raymes Khoury <raymes@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#682921}