Introduce tests for OnFetchAccessTokenComplete in O2ATMtest
This CL is a part of moving access token management to
OAuth2AccessTokenManager.

It is specifically a step toward folding O2TS into PO2TS
now that O2TS is only one subclass of PO2TS.

It introduces OnFetchAccessTokenCompleteOnRequestCompleted
to test that OnFetchAccessTokenComplete is invoked when a
request is completed and
OnFetchAccessTokenCompleteOnRequestWithoutRefreshToken
to test that it is invoked when a request without a
refresh token is started.

Bug: 967598
Change-Id: If869a4e3647f496f66c58b6547ddb0e6b15a4eb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1728410
Commit-Queue: Julie Jeongeun Kim <jkim@igalia.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683153}