Lookalikes: Show the interstitial for only a subset of top sites
This CL displays the lookalike interstitial for domains that are lookalikes of top 500 domains instead of the full top 10K. It uses the same Huffman encoded trie that contains the top 10K list to do this. The CL adds a new "is_top_500" bit to each entry and marks the top 500 domains with one and the rest with zero.

The lookalike navigation throttle reads this bit and displays the interstitial for top domains with is_top_500 bit set. It still records UMA and UKM for other top domains.

An alternative approach is to use the top 500 domains array generated by the make_top_domain_list_for_edit_distance binary. This would require rewriting the list as a map so that lookups are constant time. This isn't particularly tricky, but we would essentially duplicate the functionality of the Huffman trie in a  roundabout way so I opted for the current approach.

Bug: 988885
Change-Id: I98381f4035ace891bfe8162427e7cc2d4080a991
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1724897
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683304}