Portals: Flush touch events in PassthroughEventQueue before activation
In addition to cancelling active touches before activation, we also
need to end active gestures to ensure the gesture stream remains valid.
To do this, we ack touch events that are pending acks directly from the
browser and generate the required gesture events.

Because we flush the touch event queue in PassthroughTouchEventQueue,
we no longer need to worry about any pending acks in TEAQ, and can
revert the changes made in https://crrev.com/c/1656808 and
https://crrev.com/c/1645159.

Bug: 985929
Change-Id: I7526b8ba0d92d05d518656347865d46c20dfa2d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1714372
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683331}