Avoid timeouts in mach_fuzzer::SendMessage()
Using mach_msg_send() will wait until the message is delivered. If the
receiver is on the same thread as the sender, then the kernel message
queue can fill up and block the sender indefinitely. Switch to using
mach_msg() with a 0-second timeout and return the error in that
scenario instead.

Bug: 950323
Change-Id: Ia9491034cca45239417f8954f80b1bfb1817e1b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1731478
Reviewed-by: Max Moroz <mmoroz@chromium.org>
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683353}