NS: clean up legacy net path in Safe Browsing component
No change to behavior, because the legacy net path is no longer
supported.

This trims out usage of URLRequestContext/URLRequestContextGetter in the
Safe Browsing component. This is only used in the legacy net path, so
it's safe to remove entirely. This updates chrome/ and AW/ to no longer
pass values for URLRequestContextGetter.

Bug: 989709
Test: autoninja monochrome_apk (both chrome & webview compile)
Test: Manual - start up chrome for desktop, verify SB on http://testsafebrowsing.appspot.com
Test: Manual - start up webview demo app, verify SB works
Change-Id: I84a39cdee3b0703ebf401f565a73ef590d82ef74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1730252
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683369}