[CrOS Cellular] Fix InternetDetailDialog for locked cellular SIMs.
This dialog failed to load fully when cellular SIM was locked due to a
JavaScript error trying to access CrOncStrings. This bug was exposed due
to another CL [1] which defaults the Cellular network's name to
"Cellular" if there is no SIM card (and, thus, no carrier name to show
in its place).

In this case, the "Cellular" string is actually a field of the
CrOncStrings object, but that object was not defined in the dialog.
This CL defines the object and eliminates the error.

Note that it may be a good idea to refactor CrOncStrings usage; I've
filed [2] to investigate this issue.

This CL also adds myself as an OWNER for this directory, since I'm now
the networking lead. This directory was accidentally skipped in [3].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1660263
[2] https://bugs.chromium.org/p/chromium/issues/detail?id=990121
[3] https://chromium-review.googlesource.com/c/chromium/src/+/1532905

Bug: 983484
Change-Id: I748f300616d54ee9e1ff167371396647e58cfc09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1731952
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Michael Giuffrida <michaelpg@chromium.org>
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Auto-Submit: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683421}