desks: Fix overview highlight related issues.
Fix highlights not being positioned correctly if the shelf is not bottom
aligned. This is because we use the mini-views layer target bounds,
which are in local bounds, so we need to account for the parents target
bounds, in this case the desk bar view.

Fix highlights not repositioning after adding an item.

Fix being able to highlight the new desk button even if it is disabled.

Also fixes a crash when the highlight is on mini view, and a desk other
than the highlighted one is deleted.

Also the rounded corner of the highlight matches the new desk button
when it is highlighting it.

Test: added tests
Bug: 988183, 988134, 988126, 989795
Change-Id: I6a44ded06bce1089946dcfdc793ac4b35be4ca4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1722193
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683695}