Fix the InputTiming metrics for re-used RenderWidgets
RenderWidget objects are re-used for same-domain navigations, meaning
the input handler lives on. The lifecycle state of the input handle
must be reset on navigation to correctly report input timing.

We also increment the histogram name to reflect the change in data
measurement. And rename some of the methods to better match other parts
of the code.

Bug: 958424
Change-Id: Ic43e3a1256000ca3d99e119b3cb8ceee66dc8499
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1729794
Auto-Submit: Stephen Chenney <schenney@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683716}