Make SSLClientSessionCache take a struct for a key.
When client certificate preferences change, we should clear all session
cache entries for the host. This allows SSLClientSessionCache to
determine the host portion of each key.

Note this changes it from a std::unordered_map-backed MRUCache to a
std::map-backed one. I'm assuming that, given the cache is bounded to
1024 entries, it doesn't matter either way.

Bug: 951205
Change-Id: Ia6cb77a7331fa8cfc8aaa86f366b42f845940cd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1707035
Commit-Queue: David Benjamin <davidben@chromium.org>
Reviewed-by: Steven Valdez <svaldez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683789}