Use system font as the default font
This CL is using the user configured font as the default font.

The users in the experiment "PlatformFontSkiaOnWindows" are using the
default skia font. The current default font while using the
PlatformFontSkia is "sans", which is mapped to the default Segeo UI
font.

Some users may change the default font and chrome should use the
configured font.

R=robliao@chromium.org, asvitkine@chromium.org

Bug: 944227, 989476
Change-Id: I3f7c6b91103f9b58388a08b212fd6484344a0079
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1733429
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683830}