BundledExchanges: Update SafeBundledExchangesParser interface
This class should hold the BundledExchangesParserFactory instance
so to keep the utility process alive. This fix leads a design change
to pass a service connector at constructor rather than other open
methods.

Also, this adds an error check to SafeBundledExchangesParser::OpenFile.
We could return a file open error and so on when
BundledExchangesReader::ReadMetadata is called.

Bug: 966753
Change-Id: I3174a88d0236f4afe0ebbad8348626800096b68d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1732518
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Kunihiko Sakamoto <ksakamoto@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683923}