Ensure that +setUpForTestCase is called for each test case.
Additionally, change appLaunchManagerDidRelaunchApp: API contract to
call only when app got relaunched and not call when app got launched for
the firts time. This is because +setUpHelper and -resetAppState should
not be called unless the app was shut down.

Bug: 989743
Change-Id: Id33e5e60a693580269a0f6d5504bfc0427b34c19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1730954
Auto-Submit: Eugene But <eugenebut@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684027}