IndexedDB: Simplify IndexWriter's member data
An IndexWriter was holding both the index's metadata (including id)
and an "IndexedDBIndexKeys" struct with the index's id and a vector
of keys.

Simplify to just holding the metadata and vector of keys, which
removes an indirection step in a few places in the code. Refactored
out of an upcoming change.

Change-Id: I8efefdc5350c48a61df58b46b51be4c685c94ca5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1733014
Commit-Queue: Joshua Bell <jsbell@chromium.org>
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Auto-Submit: Joshua Bell <jsbell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684107}