Add handling for two-up view in PDFiumEngine::Paint().
Currently PDFiumEngine::Paint() takes a region of space, |leftover|,
that needs to be redrawn and iterates through the visible pages, drawing
them. A part of this process is recomputing the region of space that
needs to be redrawn. In two-up view, a page does not span the whole
width of the document so subtracting the page's area from |leftover|
does not result in a rectangle.

This CL makes PDFiumEngine::Paint() only recompute |leftover| if
|two_up_view_| is false, allowing PDFiumEngine to properly draw two-up
view layouts.

Bug: 51472
Change-Id: Ia7175d3182bd208e1b20c6ffe0a7f510ba2cff26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1731925
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Jeremy Chinsen <chinsenj@google.com>
Cr-Commit-Position: refs/heads/master@{#684146}