Move document layout constants from PDFiumEngine to DocumentLayout.
Now that DocumentLayout exists, move the constants used for document
layout over. Reduce the number of constants to remove redundancies
and update users of the constants to compensate.

After this, it may make sense to move some users of those constants
from PDFiumEngine to DocumentLayout as well.

Change-Id: I8b5e5f0a83b5c0b1448f5f14f86a4df100e73f52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1733850
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684193}