Stop passing ParentExecutionContextTaskRunners to WorkerThread
The motivation of this change is to avoid creating
ParentExecutionContextTaskRunners while starting service worker.
It's a garbage collected object and we can't create it on the
IO thread.

Before this CL, WorkerThread took ParentExecutionContextTaskRunners
as a parameter of Start(). For service/shared workers, we
create a dummy version of task runners which always
fall back to the default task runner that is bound to the
current thread. The task runners are used to schedule a
task to terminate script execution outside the worker thread
(e.g. force terminating infinite loop).

Since the termination task doesn't rely on parent context,
we can just use the default task runner of the current
thread (mostly the main thread, could be the IO thread
in the future).

Another option is to pass a SingleThreadTaskRunner as a
parameter of Start(). I didn't take this approach because:
* I wanted to aggregate the call of
  Thread::Current()->GetTaskRunner() as it's discouraged to
  use widely.
* For dedicated workers we could pass
  ParentExecutionContextTaskRunners::Get(kInternalDefault)
  but this could be stale when the parent context is
  destroyed. Keeping it without ContextLifecycleObserver
  isn't safe.

Bug: 988335
Change-Id: I3fe27bfc184619f5c4b1be516ad794190588eed0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1735591
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684252}