DNR: Remove the "redirectUrl" key.
Remove the "redirectUrl" key. Redirect actions should use the new Redirect
dictionary. Note that this is a breaking change to the API, but it should be
fine since the API is not yet launched to Stable.

Also, add more documentation for the Redirect dictionary and make it publicly
visible.

BUG=983685

Change-Id: I4f07d7389a0a8997c87caf735fd3b4b82bfe9923
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1722190
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684642}