[BackgroundTaskScheduler] Metrics for expiration feature
The expiration feature was added to BackgroundTaskScheduler, allowing the caller to decide if they want their task to expire after the time window set.

No metrics were added along the support for this feature. We are unable to measure how this works in real scenarios.

This CL adds new enumerated histograms for the scenarios we are interested in and also adds tests for these metrics. The metrics apply for both schedulers used (JobScheduler and GcmNetworkManager).

Bug: 963830
Change-Id: I5c3ecce64751519b3265c4634d69f5fad822b8d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1705915
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Richard Knoll <knollr@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Mugdha Lakhani <nator@chromium.org>
Commit-Queue: Ioana Stefan <ioanastefan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684676}