Extend sync DataTypeControllers with explicit readiness states
We introduce the concept of datatype preconditions, which enforces the
right mindset and allows datatypes to think, for each of the relevant
conditions, whether sync metadata should be cleared or not.

The implementation leverages the existing
SyncError::DATATYPE_POLICY_ERROR for this, which should be renamed or
repurposed in the future.

Bug: 990802
Change-Id: I8f1bba3e43d67fadc52c50a074cb36ed1ac23cd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1736679
Auto-Submit: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: sebsg <sebsg@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684735}