[ios] Preparation for removing the TextFieldStyling protocol (1/5).
This CL lays the groundwork for removing the TextFieldStyling and
TextFieldValidation protocols.  All direct uses of these protocols are
removed.  Bookmarks was implementing a TextFieldValidation method that
was no longer ever called.  Autofill was calling setUseErrorStyling, but
was always passing NO, so the associated code could safely be removed.

Adds a new ChromeBrowserProvider method to provide a text field that
doesn't implement TextFieldStyling, as preparation for removing the
protocol entirely.

BUG=991397

Change-Id: I8d77005c1f99ac841ce51030caa62667e244a5b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1740566
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684782}