WebApkIntegrationTest cleanup
Various changes hoping to either fix or make it easier to diagnose the
flakiness in
https://bugs.chromium.org/p/chromium/issues/detail?id=983769

This fixes a potential race condition in checking for the splash screen
hiding - it might never have shown.

Also, we get rid of the duplicate EmbeddedWebServer, using the
ChromeActivityTestRule's server.

Bug: 983769
Change-Id: I7daadb61de86746c30258e2e0367ac2a44ad0726
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1739118
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684816}