Implement UIA Page Navigation for PDF documents
This change is built upon the paragraph navigation functionality
that was added recently. Page boundaries are defined via a new
attribute 'kIsPageBreakingObject', and page navigation methods
in ax_position.h are very similar to paragraph navigation.

Additionally, paginated documents are tracked via the
'has_pagination_support_' member variable in AXTree that this
change adds. This is set on a tree when a node with
'kIsPageBreakingObject' is encountered upon serialization.

The PDF root node was changed from a kGroup role into a
kDocument role. This allows for obtaining a TextPattern for
the entire document, move-by-document UIA traversal.

Bug: 985604
Change-Id: I5f1e5bdb0a879ce77408d3c16d9187685293db21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1699147
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684854}