[Chromecast] Use a single USE_UNIX_SOCKETS buildflag
Mixer service and capture service defined their own USE_UNIX_SOCKETS
buildflag headers. Move the flag to their parent folder, so there will
be a single one.

Also, remove the use of this buildflag in constants headers, when the
constants are just string and int but not related to socket details.

Merge-With: eureka-internal/301237
Bug: internal: 133880006
Test: Compile and unit tests.
Change-Id: Ia93bb72b14521cf39d7bd905a10e5775502f7798
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1740109
Commit-Queue: Kehuang Li <kehuangli@chromium.org>
Reviewed-by: Kenneth MacKay <kmackay@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684874}