Add support for sending the system color info to the renderer process.
The system color info is sent to the renderer process:
1. When a renderer process is created by calling ThemeHelper::SendSystemColorInfo
from RenderProcessHostImpl::OnProcessLaunched
2. When the native theme was updated (e.g. HC turned on) by using the native
theme observer implemented by ThemeHelper

In order to encapsulate the system color info a new struct
mojo::UpdateSystemColorInfoParams is added.

For sending the system color info the mojo::Renderer interface is extended
by adding UpdateSystemColorInfo.

A new enum with the currently supported system colors was added to NativeTheme
(SystemThemeColor) and a map was used to store the system theme colors.

NativeThemeWin was changed to use the map from NativeTheme to store the
system theme colors and a SysColorToSystemThemeColor converter was added.

This CL will be followed by another CL to use the system colors stored
in the renderer by changing NativeThemeAura::GetSystemColor which currently
returns hardcoded colors. After the system colors are actually used tests
are going to be added to ensure that the renderer receives the right system
color info.

Bug: 989768
Change-Id: I525d61e4f788e10b074c196541befe71d5c694c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1728056
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684886}