[LCP] Add debugging rects into the trace
Largest Contentful Paint algorithm depends on the paint rects of layout
objects mainly to decide whether an object is the "largest" and
whether it is "contentful". As the paint rects are important, it is
helpful to add them into the trace in case debugging is needed.

This CL adds |PaintTimingVisualizer| to encapsulate how to generate
the trace events with the paint rects and other debugging information.
Visualization tool can be built on top of these trace events to
view the rects.

The enabling of the visualizer is enabled by the tracing category
"loading". The condition is tried when
A. |PaintTimingDetector| is constructed, or
B. at the end of a paint tree walk
(PaintTimingDetector::NotifyPaintFinished()).
B assumes that the tracing category can be changed during the visit
of a page.

Note that in |ImagePaintTimingDetector|,
* the size computation is added to get the rect when an image finishes
loading. The extra computation is not needed by LCP but only benefits
the trace. The extra cost should be deemed as tolerable compared with
the overhead of enabling the "loading" category.
* the event is calculated without considering the intrinsic size,
because the visualizer does not visualize intrinsic size yet.

Bug: 981400

Change-Id: Ifcd43adfad38be63b7f726aba98bc17ab969f0dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1730622
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684951}