Add a component registration call to the autotest API
As part of refactoring crostini tests, we need the ability to assert
that a compatible component exists (which was shipped as a data
dependency to the device in tast). This call allows us to add an entry
to the component manager, which prevents it from trying to download
that component itself.

Bug: 980048
Change-Id: Id2c935bd28dc24c81b906616b52bee1e702856d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1739232
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685037}