[ozone/wayland] Stop using wl_display_roundtrip
According to the Wayland documentation, wl_display_roundtrip is a
blocking call that can block if the event queue is empty. That is,
to be able to process buffer swap requests synchronously,
that call blocks until a frame callback comes or a wl_buffer
is attached. This results in blocking the CrBrowserMain thread.

This change alters the logic so that the buffer manager processes
one pending buffer at a time asynchronously avoiding blocking
the thread if synchronous processing is not possible due to
two conditions:
1) a frame callback is not received,
2) a wl_frame buffer is not created.

That is, on a frame no 1, the manager does not have a
frame callback installed and it may not process the buffer
synchronously if a wl_buffer is not created.
(note that wl_buffers can be created synchronously or asynchronously
depending on the protocol version of the zwp_linux_dmabuf[1]).
If such condition satisfies (there is no wl_buffer created), that
buffer is stored as a pending one and is processed on a
AttachWlBuffer call. Otherwise, the manager just processes
the buffer swap for the frame no 1 synchronously.

On frames no 2 and onwards, the manager may not process
buffer swaps synchronously if a frame callback is set, but not
received. That is, whenever the manager attaches
buffers to surfaces and commits them, it sets frame callbacks,
which Wayland compositor runs whenever it is time to send next
frame. If such condition satisfies, the manager stores the buffer
and processes it on next OnFrameCallback call. Otherwise, the swap
request processed synchronously.

[1] https://cs.chromium.org/chromium/src/ui/ozone/platform/wayland/host/wayland_zwp_linux_dmabuf.cc?g=0&l=67

Bug: 987821
Test: WaylandBufferManagerTest.TestCommitBufferConditions
Change-Id: I153b2071a453856a40e23a7f0a15f61dbc01d27f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1726051
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#685143}