Remove DCHECK which can fire spuriously in a race.
It seems we may be hitting a race condition where (1) a BrowserPlugin
is created in a mainframe, but (2) before the browser knows about the
BrowserPlugin it routes an event to a main frame that ends up
targeting the newly created BrowserPlugin. In this scenario the
HandleInput function will correctly return kNotHandled, but the DCHECK
firing really doesn't indicate a problem.

Since BrowserPlugin will go away soon, let's just remove this DCHECK.

Bug: 985944
Change-Id: I6fcaf54a2cecdcc4a53c79cbc7fa260bf6697c6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1742368
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685150}