Implement interactive UI test to measure paint perf of the browser UI
This is a native UI performance test that reports the average CPU time
of each step in the pipeline reporter (e.g.
SingleThreadedCompositorLatency.SendBeginMainFrameToCommit), in
microseconds. Note that SendBeginMainFrameToCommit is of particular
interest because the metric contains View::Paint times. In the test,
we hover over each tab in a browser by mousing slowly over the
tab strip from left to right. Hover cards and tooltips are disabled.

Note that the test currently only builds for chromeos. This is because
the DragEventGenerator has the option to use touch events, and these
touch events are only compatible with chromeos currently. If you
want to run the test on Linux or Windows, move the test file and
DragEventGenerator files in the BUILD file such that they build for
the desired platform, and comment out the SendTouchEvent() and
SendTouchEventNotifyWhenDone() from the DragEventGenerator file.

Bug: 974349

Change-Id: Iaf7c5975cca158eb79175927a60ca50ca82d1ba8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1662970
Commit-Queue: Annie Su <anniesu@google.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Reviewed-by: Collin Baker <collinbaker@chromium.org>
Reviewed-by: Malay Keshav <malaykeshav@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685292}