Add more testing for Group Policy: MaxConnectionsPerProxy
Setting this policy should affect the state of the network service as it
is getting created. This test verifies that the policy does indeed
modify the state of the network service. It does not experimentally
verify the number of connections per proxy. That sort of test is covered
in /services code.

Bug: 981598
Change-Id: Icf376f922d7d8907fb22abfad62ffd7daf33c6b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1738342
Commit-Queue: Nicolas Arciniega <niarci@microsoft.com>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685329}