Show a dialogue when we fail to load cros-termina
Currently we do this if we know we're not online and we need a major
update, but if we fail after that check nothing is displayed to the
user. This could happen if, for example, their network connection was
dropped during the update. Change this to show the same dialogue on
any install failure, since we expect most failures to be caused by
network issues.

Change-Id: I55541ceb3674cceb406fe357cb78b19ab29933e0
Bug: 990227
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1743440
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Fergus Dall <sidereal@google.com>
Cr-Commit-Position: refs/heads/master@{#685419}