clang tot bots: Disable Wreturn-stack-address
The warning got smarter on trunk, so turn it off while we fix.

We have to fix before we roll over this, else we lose a valuable existing
warning -- the improvements to the warning aren't behind a new flag.

Instead of just disabling, make this one warning non-fatal so that we
can use the tot bots to see how many things are left to fix.

Bug: 992458
Change-Id: I2daa2ba62b2fe706786bc5ce9c2e03b2f1e8dc72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1745627
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Reid Kleckner <rnk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685621}