Plumb the running LocalFrame to MediaStreamVideoCapturer ctor
This CL addresses [1], by haraken@:

"
GetCurrentContext() should not be used in places where you're
100% sure that you're in a right context (e.g., V8 binding layer).
If you happen to be in a wrong context, it will end up leaking
cross-context objects, which will cause security issues.
"

[1] https://crrev.com/c/1742117/4/third_party/blink/renderer/modules/mediastream/media_stream_video_capturer_source.cc#23

It plumbs the LocalFrame* instance calling through
MediaStreamVideoCapturer, so code does not need to fallback to
calls to v8's GetCurrentContext().

PS: in html_media_element_capture.cc, the function
CreateHTMLAudioElementCapturer() gets a LocalFrame* parameter
added although it is not used, to it keeps symmetry with
CreateHTMLVideoElementCapturer().

BUG=704136
R=guidou@chromium.org, haraken@chromium.org

Change-Id: I8a26a0488487753c6377499b0167c9545a1880ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1744373
Commit-Queue: Antonio Gomes <tonikitoo@igalia.com>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685623}