Add integration test targets to chrome_cleaner_unittests
This will cause chrome_cleaner_unittests to also run the integration
tests on the internal builders, allowing us to deprecate the
internal-only integration test suite. A followup will make most of
those integration tests public.

Exposes "is_internal_chrome_cleaner_build" as a buildflag.

There were two near-identical engine_definitions targets.
chrome_cleaner:engine_definitions was in a more logical place, but
had a typo that meant targets using it on the internal builders were
linked with the wrong definitions. By chance it was only referenced
from a test that worked anyway.

This fixes the implementation of chrome_cleaner:engine_definitions and
removes chrome_cleaner/executables:engine_definitions.

This also removes the complicated OverrideTestCommandLine function
which was used to add a longer timeout for the integration tests. The
timeout was de-facto internal-only because the integration test binary
was only used on the internal builders. Now we explicitly add the
timeout in test_main.cc when running on the internal builders.

BUG=949669

Change-Id: Ieff42b1cce452485358f005f2f2d9b67723dc2a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1742688
Reviewed-by: proberge <proberge@chromium.org>
Commit-Queue: Joe Mason <joenotcharles@google.com>
Cr-Commit-Position: refs/heads/master@{#685640}