[Locked Fullscreen] change external extensions installation stack to queue
The problem is basically that the assessment assistant extension takes a lot of time to be installed, although its downloaded already.
After Investigating the external extensions downloading order & installation order, I found that the extensions installation order doesn't follow the extensions downloading completion order. And this happens because the fetched crx files are put into stack not queue, which is quite the opposite order.

Bug: 965686
Change-Id: I6891231ae278b508f28622229b05e95f3062ef13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1741817
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Commit-Queue: Aya Elsayed <ayaelattar@google.com>
Cr-Commit-Position: refs/heads/master@{#685709}