Ozone/DRM: Make HostDrmDevice call OnGpuProcessLaunched from IO thread
DrmGpuPlatformSupportHost already calls OnGpuProcessLaunched from the
IO thread but HostDrmDevice calls it on the UI thread. It's crucial to
call this method on the IO thread because it causes
DrmDisplayHostManager to send the file descriptor for the DRM device
to the GPU process before any GMB allocation can occur. If we wait for
the UI thread to send the fd and a GBM allocation is requested in the
meantime, the GPU process will crash itself [1].

[1] https://cs.chromium.org/chromium/src/ui/ozone/platform/drm/gpu/drm_thread.cc?rcl=513c695a543d9e70c052e3193ba0eaa5824d47df&l=156

Bug: 620927,981721
Change-Id: Ib219b81bc7139987ed0922de38b95bd5f69c77b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1742385
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Commit-Queue: Saman Sami <samans@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685728}