clang docs: mention llvm_force_head_revision
Also add a gn-time assert that says that llvm_force_head_revision
can't work with use_goma. (The real assert would be "compilers not
uploaded to goma don't work with goma", but we don't have a good way
to check for that. Most non-head clangs don't work with goma, and
head clang can work with goma if it was uploaded, but for all but maybe 4
people the assert will always be true when it fires.)

Bug: none
Change-Id: Ib8b528eb52e1ea976045bc6026a4a38efca3340f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1747199
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685976}