Activate safety tips when URL blocked via component updater.
This CL changes safety tips to activate only on sites that are on the
component-updater-provided blocklist instead of literally every site
with low engagement.

To permit wider blocking, matching against the blocklist uses similar
suffix/prefix enumeration as Safe Browsing. See
https://developers.google.com/safe-browsing/v4/urls-hashing#suffixprefix-expressions .

This CL also sneaks in a little housekeeping to emphasize that the
remote blocklist is not the only thing that will (eventually) trigger
the safety tip UI.

Bug: 991589,984070
Change-Id: Ied0f516227fc45e1c1d703ace89aa14236b3e7ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1745597
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Joe DeBlasio <jdeblasio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686015}