[signin]: Add a way to debug unexpected service response errors.
We noticed that we are seeing a lot of errors in parsing the
/ListAccounts responses on Windows, Linux and macOS. We also noticed we
record two samples per failure. For this kind of problems, we do not
have any custom output printing logic that might trigger on only a
subset of requests. All responses will have the same format and there is
no way to filter them.

To solve this problem, this CL allows to append a suffix to the request
source which will provide a way to distinguish between requests. Also, a
one time retry request with the suffix |UnexpectedServiceResponse| is
issued on ListAccounts failure for that reason. We expect the retry
request will fail as well but it will allow us to filter those requests
and debug the root cause of these errors.

Bug: 992401
Change-Id: Ifd15e7fc88485bfb642804a3454c3b9fbc72e3fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1745998
Reviewed-by: Thomas Tangl <tangltom@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Commit-Queue: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686016}