Treat alpha channel differently when compositing animations
Currently our implementation of InterpolationType::Composite is too
general and could result in errors when applied to color.

Let's use an example to demonstrate how it works.
Suppose we have to keyframes to animation
keyframe[0]: (10, 0, 0, 1), composite: add
keyframe[1]: (30, 0, 0, 1), composite: replace
and the underlying color of the animation target is (50, 0, 0, 1).
We want to calculate the interpolation result at 1.2, and we will
use alpha channel because that's easier to demonstrate.

The first step is computing the interpolated result at 1.2 for the
two keyframes, and our implementation gives correct result where the
alpha is 1 * (-0.2) + 1 *1.2 = 1. In other words, the result is
(34, 0, 0, 1).

Next step is to composite the interpolated result onto the underlying
value. The alpha channel is 1 * (-0.2) + 1 = 0.8.
We can see that the alpha channel is wrong, and because of that, the
other channel would be wrong because they operate on pre-multiplied
numbers.

This CL treats alpha channel differently for color. When
the underlying and the other one has the same alpha channel, we keep
that value and do not apply the formula.

Here is the spec indicates that alpha channel needs to be always
clamped to range [0, 1]:
https://www.w3.org/TR/css-color-3/#alphavaluedt

Bug: 981326
Change-Id: I34429d0a463bdc83099d58ad23874086d861dc9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1715786
Commit-Queue: Xida Chen <xidachen@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686064}