[flags]: add about://flags entry for WebP YUV decoding
This change adds a switch to explicitly enable or disable the YUV
decoding feature for WebP images: decode-webp-images-to-yuv. We target
M81 for flag expiry, by which point we hope the Finch trial will have
ramped up to 100% on Stable and YUV decoding for WebP images will be
the default.

Bug: 900264
Change-Id: I8bc48894b18b9e38fa3b9cfe7e8fdd21828a8c7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1749142
Commit-Queue: Madeleine Barowsky <mbarowsky@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686107}