[IndexedDB] Extract connection request code into a separate class
The connection request code lived in the IndexedDBDatabase, which
already has a lot of it's own code and caused that class to be huge
and confusing.

This patch is the first step towards refactoring that part of the
code out into its own class. It is a purely non-functional change
which moves the code into a new class
'IndexedDBConnectionCoordinator' and simply delegates calls from
IndexedDBDatabase to this new class.

R=cmp@chromium.org

R: cmp@chromium.org
Bug: 992474
Change-Id: Ia3d394e097fe245fc6ecf0b57ca35986faa45670
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1745840
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Chase Phillips <cmp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686133}