Add metrics for non sync users; show non sync warning dialog.
This is a follow up CL after cr/1673863 and cr/1719975.
The CL is large because adding functionality for non sync users
require retrieving and passing around the username. Previously,
the primary account is easily accessed. Now, to get the non syncing
accounts, all the signed in accounts needs to be traversed to figure
out which account password was reused.

This CL includes several functionality:
-Recording metrics for non-sync users
-Applies the same functionality as the primary account to the
content area accounts
-Add two malicious states and two safe browsing states that replaces
the original SIGN_IN_PASSWORD_REUSE state
-Refactored most methods to take in ReusedPasswordAccountType instead
of PasswordType and username.
-Changed kSafeBrowsingUnhandledSyncPasswordReuses to
kSafeBrowsingUnhandledGaiaPasswordReuses
-Saves a username variable inside the ChromePasswordProtectionService
in order to retrieve the content area account info to see if the
account is GMAIL, GSUITE, syncing, etc.

Bug: 914410
Change-Id: I86d807c6fb25c5bb6e9075411be44bf42451a859
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1693188
Commit-Queue: Bettina Dea <bdea@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Reviewed-by: Hector Carmona <hcarmona@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Christopher Thompson <cthomp@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686232}