Make GetImportantRegisterableDomains not crash in incognito mode.
ImportantSitesUtil::GetImportantRegisterableDomains tried to access the
SearchPermissionsService under certain circumstances in incognito mode,
but the service is not instantiated for off-the-record profiles.

This CL only fixes the crash, and does not otherwise change the logic,
which also seems to be broken, as documented in crbug.com/993021.

Bug: 989890
Change-Id: I9b93bd5007855e1906fcc67c43ed625030ed7142
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1747011
Reviewed-by: Raymes Khoury <raymes@chromium.org>
Commit-Queue: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686359}