[Code health] Add flat_map.h to native_theme.h
This class is using base::flat_map, which seems to be added by this CL:
https://chromium-review.googlesource.com/c/chromium/src/+/1728056

Notice that the flat_map.h is not included. It works fine for now, but
can be a potential problem. For example, when I am working on my own CL:
https://chromium-review.googlesource.com/c/chromium/src/+/1614120

There is a compilation error if flat_map is not included in this file.

Bug: None
Change-Id: I8d9be82945a322afba3d0a07bd8d79d7a9910495
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1746293
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686451}