Don't allow plugins to change file extensions arbitrarily
Currently extensions are allowed to use the download API to override
downloaded file extensions. This causes some security issues
as file with dangerous extensions can be override to safe extensions
and bypassing dangerous prompt. And there are also other possible
attacks using this feature.
This CL blocks plugings from changing the file extensions to an
arbitrary one. The mime type will be used to determine the final
extension. However, it is possible for plugins to remove the
extensions from the final file name.

BUG=989078

Change-Id: Idd28510d3db191f40bbe24256d64449856e4644f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1740048
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Ben Hayden <benjhayden@chromium.org>
Commit-Queue: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686498}