media/gpu/test: Fix video decoder tests build when no decoder type is set.
When the video_decode_accelerator_(perf_)tests are build for ChromeOS the video
test code expects at least the 'use_v4l2_codec' or 'use_vaapi' flag to be set.
This CL ensures the build works if none of these flags are set. The test won't
work, but the build will not be broken.

BUG=993467

TEST=Built for ChromeOS without 'use_v4l2_codec' or 'use_vaapi' set

Change-Id: Ic7b543907ed1b37bcac2b9f2bc3dc006109a8166
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1752268
Reviewed-by: Manu Cornet <manucornet@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686677}