media/gpu/test: Change dropped frame rate to dropped frame percentage.
The new video_decode_accelerator_perf_tests collect various performance metrics
while decoding a video. To give an idea of the relative amount of frames dropped
the 'frame drop rate' is calculated, which is the ratio between frames dropped
and frames not-dropped. This metric is unfortunately rather confusing to
interpret (e.g. 50% of frames dropped is a frame drop rate of 1). The drop rate
was initially used rather than drop percentage to keep in sync with the old
video_decode_accelerator_unittest.

This CL changes the 'frame drop rate' metric to the more readable 'frame drop
percentage', which is also the metric collected by various other tests such as
the Tast playback_perf_* tests.

TEST=new vda performance tests on eve

BUG=None

Change-Id: I674e47b9d540df0348ee53f4cf71f5cc19fb5db8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1750534
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686680}