arc: ime: Restore ARC IME focus after CustomTabs.
Because of CustomTab, it is possible for ARC IME to lose IME focus even
though window focus does not change. To prevent that, we should ensure
the IME focus is set again when the custom tab child window is removed.

TEST=manual(Open a custom tab in Gmail, type something in web contents,
close the tab, and type again in Gmail search box in the app.)
BUG=b:135281220

Change-Id: I7ee383cdf952c349264725e5691323ce0ee609fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1752272
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Commit-Queue: Tetsui Ohkubo <tetsui@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686710}