Open dialer for Click to Call messages on pre Q.
On pre Q Android devices, we can simply open the dialer instead of
having to display a notification first. This adds a flag so we can test
user behavior with this new flow. We still keep the notification to
notify users if the screen is off or locked.

Bug: 992879
Change-Id: I63e3dfe1e294377705a371b3cb76384ad4c0eaff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1748750
Reviewed-by: Peter Conn <peconn@chromium.org>
Reviewed-by: Michael van Ouwerkerk <mvanouwerkerk@chromium.org>
Commit-Queue: Richard Knoll <knollr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686749}