[Autofill Assistant] Fix animation rendering crash on Payment Request.
Summary: The animation crashes on the payment section when switching
between billing addresses with and without a postcode.

When switching the require_billing_postal_code flag to true,
and using a profile with a billing address that does not have a
postcode, the Payment Request section will show an error message
requiring a billing postcode.
If users then edit the payment method and choose a billing address with
a postcode, the animation of the payment method section causes the app
to crash because the error message TextView is also being hidden (using
View.GONE). This is an animation rendering bug.

Fix: Disable animation on the payment method section for children of
the section choice list. The animation is not required and causes the
crash. Also setting the text of the error message to empty string when
there is no missing information.

Bug: b/139265901
Change-Id: I4b81b3dcbc08690a372748f2bb09d74968b33ccd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1753005
Reviewed-by: Jordan Demeulenaere <jdemeulenaere@chromium.org>
Commit-Queue: Marian Fechete <marianfe@google.com>
Cr-Commit-Position: refs/heads/master@{#686782}