Sync: Add SECURITY_EVENTS to AlwaysPreferredUserTypes
Before this CL, the SECURITY_EVENTS type was enabled only if the
"Sync everything" toggle was on. That doesn't make much sense, and was
never the intention.
This CL adds SECURITY_EVENTS to the AlwaysPreferredUserTypes set, so
it will always be enabled if Sync is on (unless the data type itself
decides that it shouldn't be enabled).

Bug: 993119
Change-Id: I54e5a444dbfaa05647aff740f20b0c8764ceb504
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1751986
Commit-Queue: Markus Heintz <markusheintz@chromium.org>
Auto-Submit: Marc Treib <treib@chromium.org>
Reviewed-by: Markus Heintz <markusheintz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686799}