[BackgroundTaskScheduler] Task ids - class names mapping
BackgroundTaskScheduler currently keeps information about both the id
and the class name of the tasks it schedules. This information is
redundant, since a task id can be associated with its class.

This CL is the first out of several that will remove the retention of
BackgroundTask class names. This CL modifies the way in which
the corresponding BackgroundTask class is retrieved. The
selection of the correct class will be done based on a BackgroundTask
factory. The caller of BackgroundTaskScheduler is responsible for setting
this factory so that it will take into consideration all the needed task
ids.

This work exposes an interface for the BackgroundTask factory and a
concrete implementation for //chrome, the main user of this component.
Specific code calls and tests are updated accordingly.

Bug: 916875
Change-Id: Idf2bfa18e83316df9fbd03d84571ba5c9d1fca03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1742164
Commit-Queue: Ioana Stefan <ioanastefan@chromium.org>
Reviewed-by: Richard Knoll <knollr@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686801}