[Fuchsia] Fix memory leak in gin::PageAllocator::ReleasePages()
On all non-posix platforms ReleasePages() was calling
base::DecommitSystemPages() instead of freeing these pages. On Windows
that doesn't cause the leak because the pages will be freed later when
the whole region is released. That's not the case on other platforms,
particularly Fuchsia. As result ReleasePages() was essentially leaking
memory on Fuchsia.

Bug: 993541
Change-Id: I6d3c14184d553ea32ef28e64843e1f1915979ee6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1752877
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686941}