[ios] Remove VC walk in recent tabs
The utility method |TopPresentedViewController()| performs a recursive
walk up the view hierarchy to find the top presented view controller.
This method was used in recent tabs to present the long press action
sheet for remote devices.

This CL removes this unnecessary call from this use case.

The rationale for removal:
1) The action sheet should only be displayed above the recent tabs UI.
This is the only context in which the action sheet options make sense.
If another modal UI, such as settings, was above the recent tabs UI,
this action sheet should not be shown.

2) The use of |TopPresentedViewController()| is deprecated and is bad
practice.

This passes all existing tests, and there are no useful additional tests
that I can think of adding.

Bug: 754642
Change-Id: I2b2e698261147f7fc3d0610b21a6f7986d6cf2a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1754423
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686995}