media/gpu/v4l2ip: get correct number of planes for V4L2 format
When creating an ImageProcessor, we convert the desired fourcc into
its VideoFrameLayout counterpart. However, the conversion was incorrect
if the requested format was single-planar: V4L2 pixel formats
differenciate between single and multi-planar layouts, while
VideoPixelFormat does not. Therefore, when converting the V4L2 format,
the resulting VideoFrameLayout was always considered to be multi-planar.

And when we finally create the V4L2 image processor and the
VideoFrameLayout is converted back to its V4L2 format counterpart, we
would receive a multi-planar format even though the image processor only
supports single-planar formats, resulting in an error.

Fix this issue by querying the number of planes on the V4L2 pixel
format (which contains this information), not the VideoPixelFormat
(which does not).

Bug: b:123551776
Test: VDA unittest passes on Hana and Kukui.
Change-Id: I9262f3d37b5bbc4e2af973a58e5416931718434d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1753794
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687001}