DeviceScheduledUpdateChecker: Choose correct schedule on retry
This change fixes a bug in the retry logic. Previously, time would be
recalculated from current time on a retry. This was wrong and would
cause missed schedule. Instead the original calculated schedule should
be used on a retry, if that time hasn't passed. This change also fixes a
test that should have been checking for this.

BUG=924762
TEST=Unit tests.

Change-Id: I932a869e960f011bd5e6e4fa0304da7263af3344
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1730590
Commit-Queue: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687046}