Remove DataReductionProxyIOData
Now that network service is launched and navigation lives on the UI
thread, there is no need for DRP to keep state on the IO thread. This CL
cleans up all DRP state that was still on the IO thread. Most of the
logic that used to live in DataReductionProxyIOData has been moved to
DataReductionProxyService.

Also removes some uses of DataReductionProxyEnabledWithNetworkService
in browsertests since this should be removed soon.

Bug: 934009, 978614
Change-Id: Id2b5891106a8ea6e3dd13520a22922f9719612b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1746682
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687065}