Enable ExtensionManagementTest.InstallSameVersion with flakiness fix.
The test queries the version information of the newer extension
through ExecuteScript (from IsExtensionAtVersion). However, there
is no guarantee that the background page's WebContents has started
loading or it has seen DidStopLoad. This is the cause of this test's
flakiness. This CL waits for the extension WebContents to see
DidStopLoading, after the older extension version's ExtensionHost
is destroyed.

Bug: 89727
Change-Id: Ib27c1c481c3d1a72fcf9bb174b0974b400dadc1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1725120
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Commit-Queue: Istiaque Ahmed <lazyboy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687068}