service worker: Prepare context wrapper for UI thread core.
The plan is to move ServiceWorkerContextCore to the UI thread. This
prepares for the change by allowing more callsites to use
ServiceWorkerContextWrapper on any thread, while internally using
core on the core thread. The end goal is for both core and wrapper
to be on the UI thread only (and perhaps removing wrapper totally).

More detail: methods that were previously IO thread only are moved
to be callable on any thread, and methods that were previously UI thread
only (or callback'd on the UI thread) are kept that way since it's
aligned with the end goal. In some cases methods cannot be moved,
because, e.g., they have a synchronous return value so either way
the callsite will have to adjust for the new world.

This CL is a bit rough with, e.g., consistent order of methods in the
.h and .cc file because this is largely temporary.

Bug: 824858
Change-Id: I598433f9852ab8a9b7a068242c5d12e770f21794
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1753843
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687149}