Initializes the AppIconSource from the AppServiceProxy.
This makes it possible to use chrome://app-icon/ urls from pages other
than AppManagement. It will make it possible to make FileHandling on
ChromeOS BMO agnostic.
https://chromium-review.googlesource.com/c/chromium/src/+/1743289

(we need this so we can get an icon url regardless of whether the app
is a WebApp or a BookmarkApp).

Bug: 938103
Change-Id: I171908db99dc41f225ecfe53285778b6c524eae2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1753996
Commit-Queue: Jay Harris <harrisjay@chromium.org>
Reviewed-by: Nigel Tao <nigeltao@chromium.org>
Reviewed-by: Jeevan Shikaram <jshikaram@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687173}