generate_breakpad_symbols: fix symlink
Fix symlink creation for cases when the binary does not have build-id.

When --symbols-dir is a relative path, the symlink under discussion
would look like this:

'path/to/libfoo.so/000000000000000000000000000000000' -> (pointing to)
        'path/to/libfoo.so/<HASH>'

It works for absolute paths, but not for relative ones like 'path/to'.
It will work for both absolute and relative paths if the link itself is
relative and points to -> '<HASH>'. Fortunately it is easy to do.

Done.

Bug: None
Change-Id: Ic764f0380063e2f2e15cf484afd5468dcca8f51f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1745986
Reviewed-by: Christopher Grant <cjgrant@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Auto-Submit: Egor Pasko <pasko@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687283}