Update input type 'date', 'datetime-local', and 'week' ControlType, LocalizedControlType, and AriaRole
The html-aam spec doesn't specify values for ControlType,
LocalizedControlType, or ARIA role for input date,
datetime-local, or week.  However, the current values for these
don't really reflect the way that the user interacts with the controls.
The current values are:
ControlType: UIA_ListControlTypeId
LocalizedControlType: "list"
AriaRole: "list"

The values should be changed to the following to better reflect the
way that the user interacts with the control (basically as a textbox
that restricts its input to valid dates), with LocalizedControlTypes
that distinguish between the different types:
ControlType: UIA_EditControlTypeId
LocalizedControlType: "date picker" for input date, "local date and time picker"
for datetime-local, "week picker" for week.
AriaRole: "textbox"

Note that the above refers to the textbox part of the control,
not the calendar popup.

Bug: 993549
Change-Id: I1071d9c69701cb312426c8cded4875ac339c57ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1751975
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Commit-Queue: Dan Clark <daniec@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#687313}