Add fractional scroll offsets to about:flags
This will allow us to easily turn the feature on and off for testing.
This CL adds a Feature for the Blink RuntimeEnabledFeature and plumbs
it through to about:flags.

Today, Blink stores scroll offsets for scrollers using integral
values. For user scrolls, we can apply subpixel precision deltas
(e.g. on high DPI screens or if the user zooms) since the compositor
stores a floating-point scroll offset. However, when the offset is
sent to/received from Blink it is truncated. Therefore, script set
and read offsets are always integer values (of physical pixels).

The FractionalScrollOffsets feature allows Blink to store offsets in
a floating point representation. This will reduce bugs and maintenance
burdens related to the truncation between CC/Blink. It also helps fix
issues in the web platform as WebAPIs will now store and return any
number in the floating point range.

Bug: 414283
Change-Id: I036402977204821ba92b592c1c88c2b46e64c7dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1753546
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687346}