[Rendering Core Metrics] Split StyleAndLayout into Style and Layout
In various investigations we have struggled to identify whether a
perf regression is due to a style change or a layout change. Improve
this situation by splitting the metric.

There are some consequences to this. UMA will report layout and style
for each sub-frame as a separate entry, rather than the previous behavior
of recording an UMA entry for the main frame layout plus all child frames,
and probably one for each child frame too. The result might be clearer
now.

We also losing coverage of some tasks within the UpdateStyleAndLayoutRecursive
call because we do not want the intervals to become intertwined (in other words,
we want to be sure we are not double counting).

And building the layout tree is included in Style, because that's how the
code is written.

Finally, the ForcedStyleAndLayout metrics remain unchanged. A follow on
task will attempt to disambiguate those from document lifecycle updates.

No UKM privacy review is required because we are not gathering any additiomal information.

R=eae@chromium.org, futhark@chromium.org, ikilpatrick@chromium.org, rkaplow@chromium.org

Bug: None
Change-Id: Ifc05e653c3d557b24178a61c54508ab6919f1759
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1752084
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687400}