grit: policy: sort policy_definitions output for stability
The policy logic sorts a bunch of its keys already to keep the output
stable, especially for unittests.  The policy_definitions section is
not currently sorted which leads to unittest failures on Python 3 as
it happens to iterate differently.  Sort the keys to fix it.

Bug: 983071
Test: `./grit/test_suite_all.py` passes
Change-Id: Iacd30d0edc2a597ada265b4b69664a8d19d9629c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1755389
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687406}