Measure 3rd parties accessing localStorage / sessionStorage on a page load
How: Add a OnDomStorageAccessed method in MetricsWebContentsObserver,
and call it from ChromeRenderMessageFilter::OnAllowDOMStorage. Move
TabSpecificContentSettings::OnDomStorageAccessed to the same site as
they require the same web_contents. Route this signal from
MetricsWebContentsObserver to ThirdPartyMetricsObserver that's able to
store the counter and log to histograms.

Bug: 986782
Change-Id: I9c8e5d210751152ca83272a78edf86714a3e0b96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1655996
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Commit-Queue: Yao Xiao <yaoxia@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687558}