[LayoutNG] Fix painting background of NGPhysicalBoxFragment
This patch fixes painting background and borders when painting
|NGPhysicalBoxFragment|. The new logic matches to the one for
block children in |BlockPainter::PaintContents|.

This patch reduces the failures when enabled[1] down to 793.

This change should not run by default yet. This is part of
the code path for when |NGBoxFragmentPainter| finds block
children, which occurs only when 1) NGPhysicalBoxFragment
paint is enabled, or 2) NG multicol is enabled.

[1] https://crrev.com/c/1719913

Bug: 988015
Change-Id: Ie1565bc8f51734bf7133136d59b2de372458acd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1757569
Commit-Queue: Koji Ishii <kojii@chromium.org>
Commit-Queue: Emil A Eklund <eae@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687724}