[ios] Call updateNewWebState:withWebStateList: in BadgeMediator init
If an app is updated and then launched, the tabs will be saved.
However, the LocationBar logic is completely reloaded, so BadgeMediator
will be passed a WebStateList that may have an active WebState. This
change ensures that the active WebState's InfobarBadgeTabHelper's delegate
will be set and the badges for that WebState will be displayed immediately.

Bug: 976901
Change-Id: I8f1272bd5b8fd59a735a889f7f290a2c4b8748f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1758508
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687732}