AW: fix testSHA1LocalAnchorsAllowed for Android Q
No change to production logic. Minor change to test.

This resolves a TODO with testSHA1LocalAnchorsAllowed. This test was, at
the time of writing, broken for the first Android OS to distrust SHA1
local anchors. As Android Q finalizes, we know this will be the first
release to do so, so it's time to update the test.

Bug: 919749
Test: run_webview_instrumentation_test_apk -f AwNetworkConfigurationTest.testSHA1LocalAnchorsAllowed (on Q)
Change-Id: Ib55f65446573158ac85ee2bf9494a990b3b56d0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1757079
Auto-Submit: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Changwan Ryu <changwan@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687733}