components/arc/video_accelerator: Verify arguments and create GMBHandle from them
VerifyVideoFrame is used in GpuArcVideoDecodeAccelerator (GAVDA) and
GpuArcVideoEncodeAccelerator (GAVEA) so as to verify that arguments sent from
Android that stands for video frame buffer. It returns bool value. This CL
changes the function to create GpuMemoryBufferHandle. If the arguments are
invalid, it returns base::nullopt. This is useful because GAVDA can pass the
GMBHandle to VideoDecodeAccelerator as-is, and GAVEA can briefly construct
media::VideoFrameLayout from GMBHandle. This should not change the original code
logic in GAVDA and GAVEA.

Bug: None
Test: video.ArcVideo*
Change-Id: Idd2f3feab192ec68bcb51ee37158d9264f36d66b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1728311
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688016}