ResourceLoadPriority tests to no longer rely on exact timing of messages
Split off from https://crrev.com/c/1751324

The ResourceLoadPriority tests send Priority and a Status result via
postMessage. Due to the way this is implemented the test is unnecesarially
sensitive to the loading order and can fail even when the thing it's
actually testing passes.  This patch changes the message to a key-value
pair which is more robust.

Bug: 971191
Change-Id: Ieb7d7d736c363b75a0b1e2e1f1a2ad95cfaddbff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1760226
Commit-Queue: Alex Clarke <alexclarke@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688137}