Hover Cards: show loading message as page loads.
Currently, it's possible for a tab hover card to be blank if it's
displayed when a page is loading for the first time. This CL changes the
behavior so that in that state, the title that is shown is "Loading..."
(or localized equivalent) unless the page has crashed in which case a
"Crashed." message appears. The URL that is displayed in this case is
the visible URL, not the committed URL, so it's obvious to the user
*which* page is loading.

Bug: 981103
Change-Id: I082088ac42bf32c37a2dc3ed220c68ba5d5400da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1760669
Commit-Queue: Dana Fried <dfried@chromium.org>
Auto-Submit: Dana Fried <dfried@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688266}