Extensions Tab API: Move permissions checks to its own method
This CL moves the permission checks that used to exist in the
ExtensionTabUtil::ScrubTabForExtension method into its own util method
so that the level of scrubbing of data that needs to happen can be
determined in the context from which it is requested. This will
ultimately help in allowing WebUI to get non-scrubbed data.

Bug: 994337
Change-Id: I6c7cd37b1cb0f64233759ccac7aa9b8097234482
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1755052
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688610}