[omnibox] Sort submatches consistently with parent match
Submatches normally have the same relevance as their parents
and are thus sorted positionally with them. However, if
there are demote-by-type rules in effect, a parent (or a
submatch) can have an altered relevance and thus be
sorted away from their family.

This CL records within each submatch the type of its parent
so that it can be sorted identically.

Change-Id: I396dba33ba8808a2bae62b79352962f611feb23e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1721041
Commit-Queue: Kevin Bailey <krb@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688625}