Add `Blink.UseCounter.File.Features` to `histograms.xml`
https://chromium-review.googlesource.com/c/chromium/src/+/1386464 added
a histogram to track use counters on `file:` URLs, but neglected to add
that new histogram to the XML file. This patch takes care of that minor
oversight.

We'd like to keep this histogram around long enough to make a decision
about whether/how to deprecate and remove `file:` URLs' capabilities
over time to reduce the attack surface we're exposing to locally-hosted
pages. Setting the expiry to the end of the year seems like a
reasonable timeline to aim for.

Bug: 794098
Change-Id: I0e659be03ceb7c86477d5a54f30e4b58e81c7b50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1760208
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688764}