[JJI] Implement Java representation of injected JS object for replying
This CL implements a Java class for replying directly to the injected
JavaScript object without JavaScript passing a MessagePort to Java.

The purpose is to make the API more intuitive to use, without requiring
app developer to think about MessagePort when not absolutely necessary.

Bug: 918065
Change-Id: Ib42fad47ad68f1d9c124fb0d8c6e3558979be300
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1719453
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Changwan Ryu <changwan@chromium.org>
Commit-Queue: Shimi Zhang <ctzsm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688808}