Fix test for NTLM auth over H2: HttpNetworkTransactionTest.NTLMOverHttp2
The test was simulating the H2 server hanging up after sending an H2
challenge, and then negotiating a new H1 connection, regardless of
whether net/ tried to negotiate H2 or H1 when re-establishing the
connection. This CL keeps the H2 session open, and explicitly checks
that when re-establishing a connection, we only try to negotiate H1.

Bug: 996042
Change-Id: Icb2d0e39382f9173308582b19d836a9cd26f047f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1762458
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Bence Béky <bnc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689020}