Refactor bottom Inset API onto ViewAndroid.
Refactors the logic that allows insetting the Visual Viewport.
Now the control is moved from RWHV to ViewAndroid and its delegates.
This allows the setting to be established for an object with the
same lifetime as a WebContents, so when a new RWHV is swapped in
it will still apply.

BUG=986384, 968150

Change-Id: I5cb3601ff460637f99e98a4edf090ad28321bc29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1714286
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Auto-Submit: Donn Denman <donnd@chromium.org>
Commit-Queue: Donn Denman <donnd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689037}