Surface "Setup not completed" error in settings page
As part of making sync manual_start for android the user needs to be aware
that the initial setup process was interrupted and sync is in an
incofigured state. This should be reflected on the main settings page
and the user can either start sync or stop sync from syncAndServicesPreferences
page.

Either way FirstSetupComplete will be true after user has seen and
interacted with it in the syncAndServicesPreferences page.

Screenshot: https://crbug.com/991522#c2

Bug: 991522
Change-Id: Id65cccc62933dd6a7446ad577e066b909bf28eec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1744487
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Tanmoy Mollik <triploblastic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689080}