android: track background UMA sessions
Added a new histogram 'Session.Background.TotalDuration'. This is recorded
in every log which contains a background session.

When the UMABackgroundSessions feature is enabled, we also attempt to open
new UMA logs on foreground/background transitions to segment
foreground/background UMA.

The primary goal is to learn which whether any given UMA histogram is
being reported in the foreground, or the background.

Bug: 954310
Change-Id: I7cb5bdbdc9d153e32bc283065f8147d548a82f45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1650526
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Commit-Queue: Dan H <harringtond@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689081}