Remove NullValue() from CalculateDirectedProgress
NullValue() is represented as a NaN in a double. We could better return
a base::Optional<double> and use the undefined base::Optional as null
value. Also there is no need for the IsNull() call as we could just use
the returned value as a boolean.

Bug: 994811
Change-Id: I2c5eb88ba563486f4864480b8f842ce349c610f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1760965
Commit-Queue: Sergio Villar <svillar@igalia.com>
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689144}