Payments: Prepare for service worker core on the UI thread.
The thread ServiceWorkerContextCore lives on (the "core thread") will
move from the IO thread to the UI thread.

This CL makes PaymentAppBrowserTest tests pass when run with
ServiceWorkerOnUI enabled. It is fairly mechanical. One tricky
part is whether the callbacks here need to be called asynchronously.
In some cases we may already be on the thread the callback needs
to run on, the CL still uses PostTask() to run the callback in case
sync callbacks are not ok.

Bug: 824858
Change-Id: I195d4ce93d267d9417a1e671d64b3d6ae4a8acc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1763551
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689181}