Remove NullValue() from CalculateOverallProgress
NullValue() is represented as a NaN in a double. We could better return
a base::Optional<double> and use the undefined base::Optional as null
value. Also there is no need for the IsNull() call as we could just use
the returned value as a boolean.

Bug: 994811
Change-Id: I236fb8dbf2f0fd6ea879a6996e2a11a143406b09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1761272
Commit-Queue: Sergio Villar <svillar@igalia.com>
Reviewed-by: Yi Gu <yigu@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689197}