Fix inverted backdrop mask problem for non-flipped framebuffer
On non-flipped framebuffer devices (e.g. Mac in some cases),
the mask image was previously being wrapped into an SkImage
with the incorrect value for flip_texture. This is now
correct, which fixes a bug with masking of backdrop-filter.
I also improved the backdrop-with-mask WPT test to catch this
problem.

Bug: 974857
Change-Id: I5c7cdedac29dfe1a9d9c214169fde718799649fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1761557
Commit-Queue: Mason Freed <masonfreed@chromium.org>
Reviewed-by: enne <enne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689253}