Worker: Stop directly binding RenderFrameHost with callbacks for mojo receiver requests
In PopulateBinderMapWithContext(), callbacks for mojo receiver requests are
bound with RenderFrameHost, and then added to the binder map indirectly owned by
DedicatedWorkerHost. This is not safe because RenderFrameHost can be destroyed
before the binder map and the callbacks, and that may result in dangling pointer
access.

To avoid it, this CL binds DedicatedWorkerHost with the callbacks, and makes it
check if RenderFrameHost exists before proceeding with the operations. This is
safe because DedicatedWorkerHost outlives the binder map and the callbacks.

Bug: 936482
Change-Id: If32fbe142ad544e519a153082f1e1aadcab9f6f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1763514
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689358}