[gamepad] Shut down XInput haptics on Windows
When the gamepad service is shut down, vibration effects may still
be active. Chrome tries to stop all vibration on the gamepad before
disconnecting from the device. However, this was not done for XInput
gamepads on Windows. This could allow haptic effects to continue for
a few seconds after the gamepad is disconnected. It also causes a
DCHECK failure in AbstractHapticGamepad which checks that the Shutdown
method was called before the gamepad is destroyed.

This CL ensures that the Shutdown method is called when the data
fetcher is destroyed.

BUG=996038

Change-Id: Ib62566ca2a81c2e68a0d15830250a7f529f396d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1762952
Reviewed-by: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Commit-Queue: Matt Reynolds <mattreynolds@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689699}